Identify Records with Identical lat/lon

Removes or flags records with equal latitude and longitude coordinates, either exact or absolute. Equal coordinates can often indicate data entry errors.

cc_equ(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  test = "absolute",
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

test

character string. Defines if coordinates are compared exactly (“identical”) or on the absolute scale (i.e. -1 = 1, “absolute”). Default is to “absolute”.

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = runif(100, -180, 180), 
                decimalLatitude = runif(100, -90,90))

cc_equ(x)
#> Testing equal lat/lon
#> Removed 0 records.
#>     species decimalLongitude decimalLatitude
#> 1         a       160.269109      27.2956089
#> 2         b        65.216888      82.9237247
#> 3         c       -84.629054     -70.4130701
#> 4         d        36.690602      33.0630807
#> 5         e      -106.208232      71.5430645
#> 6         f      -101.489814     -82.0024422
#> 7         g       143.608857      24.0499472
#> 8         h        58.314551       3.3895738
#> 9         i       -23.582318      22.7689531
#> 10        j       -61.904052      34.7949683
#> 11        a        46.378090     -81.2431297
#> 12        b       153.755733     -58.9495659
#> 13        c       142.981592     -82.0904771
#> 14        d       153.892902       3.8788759
#> 15        e        29.346947     -42.9537470
#> 16        f       -72.720573     -84.2246225
#> 17        g       113.766022     -40.5930452
#> 18        h       100.947146     -44.2503193
#> 19        i        23.161850     -10.6767552
#> 20        j       157.989773     -67.7825205
#> 21        a       -18.511653      28.5614028
#> 22        b        14.515776      73.6548520
#> 23        c       -83.197276      30.5310454
#> 24        d        16.847353     -36.8500016
#> 25        e       -16.661030     -35.4200974
#> 26        f       -62.277060     -24.4469457
#> 27        g      -122.557924     -30.7358326
#> 28        h      -160.996533      -9.2211041
#> 29        i        85.355319     -55.7804615
#> 30        j        66.045137      37.4150015
#> 31        a        74.031279      -0.9529211
#> 32        b        27.910924      50.1067747
#> 33        c        73.970530      -8.7430499
#> 34        d       -26.202273     -47.3140208
#> 35        e      -132.700720     -87.0472465
#> 36        f      -118.506494      66.9127950
#> 37        g      -146.915065     -19.2183272
#> 38        h      -116.518264      23.0103279
#> 39        i       131.867487      22.2948842
#> 40        j       173.542407      60.0956942
#> 41        a       146.993527      49.9779457
#> 42        b      -167.278417      34.3859275
#> 43        c        64.081680     -48.1830633
#> 44        d      -108.811220      58.9500939
#> 45        e       -80.210599      51.4744740
#> 46        f        53.286341      78.2814287
#> 47        g       146.583448     -10.4935775
#> 48        h       139.786501     -16.0466313
#> 49        i      -144.970702      -7.0545405
#> 50        j        43.759288     -55.0748226
#> 51        a        81.976647     -44.2848129
#> 52        b       -74.659282     -55.0076477
#> 53        c      -109.440935      45.5037838
#> 54        d       -53.277203     -38.6157833
#> 55        e      -117.884206      -6.6504411
#> 56        f      -138.456886     -71.6869299
#> 57        g       -73.515591      76.0628745
#> 58        h       -42.958319     -18.7241687
#> 59        i       123.182188      -5.0062751
#> 60        j      -178.992989      -4.0451872
#> 61        a       120.978559      -1.6646206
#> 62        b       -64.015845      83.4120601
#> 63        c        45.965201      72.0918052
#> 64        d       -83.941872     -40.1918851
#> 65        e        24.744688     -80.5835380
#> 66        f       -75.324533     -62.1894143
#> 67        g        69.978439      84.9171606
#> 68        h        77.380708      59.9778840
#> 69        i      -134.234864     -16.7514387
#> 70        j       113.122179      59.3873878
#> 71        a       123.557669     -87.9055432
#> 72        b       -97.881135     -17.8710419
#> 73        c       137.749045      31.3103173
#> 74        d       141.166809     -61.2909550
#> 75        e        43.507649     -54.8303317
#> 76        f      -146.995718      27.0387670
#> 77        g        73.602938      52.3560371
#> 78        h        94.122627     -76.2501489
#> 79        i       175.813169      64.6438487
#> 80        j      -161.829971     -45.0095183
#> 81        a        39.987200      64.8544599
#> 82        b       177.432772      32.1143207
#> 83        c      -102.652072      62.4744191
#> 84        d        55.171711      -9.0649890
#> 85        e       136.165413     -45.9835032
#> 86        f      -115.389040     -57.0743660
#> 87        g      -154.545604      37.3320860
#> 88        h       -71.288596      79.8105190
#> 89        i       118.143745     -46.3611634
#> 90        j       -20.102337      -8.6654727
#> 91        a       109.506133      17.3913500
#> 92        b       -36.304626     -89.9302917
#> 93        c       167.145099     -64.1513419
#> 94        d       161.324596      42.4307689
#> 95        e       -38.753879     -63.3270072
#> 96        f       -11.793364      20.9857288
#> 97        g         4.433776      56.9689823
#> 98        h       138.654422      64.3225797
#> 99        i        96.429019     -28.8693632
#> 100       j        38.302281     -61.5896825
cc_equ(x, value = "flagged")
#> Testing equal lat/lon
#> Flagged 0 records.
#>   [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [16] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [31] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [46] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [61] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [76] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [91] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE