Identify Records Assigned to GBIF Headquarters

Removes or flags records within 0.5 degree radius around the GBIF headquarters in Copenhagen, DK.

cc_gbif(
  x,
  lon = "decimallongitude",
  lat = "decimallatitude",
  species = "species",
  buffer = 1000,
  geod = TRUE,
  verify = FALSE,
  value = "clean",
  verbose = TRUE
)

Arguments

x
data.frame. Containing geographical coordinates and species names.

lon
character string. The column with the longitude coordinates. Default = “decimallongitude”.

lat
character string. The column with the latitude coordinates. Default = “decimallatitude”.

species
character string. The column with the species identity. Only required if verify = TRUE.

buffer
numerical. The buffer around the GBIF headquarters, where records should be flagged as problematic. Units depend on geod. Default = 100 m.

geod
logical. If TRUE the radius is calculated based on a sphere, buffer is in meters. If FALSE the radius is calculated in degrees. Default = T.

verify
logical. If TRUE records are only flagged if they are the only record in a given species flagged close to a given reference. If FALSE, the distance is the only criterion

value
character string. Defining the output value. See value.

verbose
logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Details

Not recommended if working with records from Denmark or the Copenhagen area.

See also

Examples

x <- data.frame(species = "A", 
                decimallongitude = c(12.58, 12.58), 
                decimallatitude = c(55.67, 30.00))
                
cc_gbif(x)

#> Testing GBIF headquarters, flagging records around Copenhagen

#> Removed 1 records.

#>   species decimallongitude decimallatitude
#> 2       A            12.58              30

cc_gbif(x, value = "flagged")

#> Testing GBIF headquarters, flagging records around Copenhagen

#> Flagged 1 records.

#>     1     2 
#> FALSE  TRUE