Flag Records in the Vicinity of Biodiversity Institutions

Flag records assigned to the location of zoos, botanical gardens, herbaria, universities and museums, based on a global database of ~10,000 such biodiversity institutions. Coordinates from these locations can be related to data-entry errors, false automated geo-reference or individuals in captivity/horticulture.

cc_inst(x, lon = "decimallongitude", lat = "decimallatitude",
  buffer = 0.001, ref = NULL, value = "clean", verbose = TRUE)

Arguments

x
a data.frame. Containing geographical coordinates and species names.

lon
a character string. The column with the longitude coordinates. Default = “decimallongitude”.

lat
a character string. The column with the longitude coordinates. Default = “decimallatitude”.

buffer
numerical. The buffer around each province or country centroid, where records should be flagged as problematic, in decimal degrees. Default = 0.001 (= ~ 100m).

ref
a SpatialPointsDataframe. Providing the geographic gazetteer. Can be any SpatialPointsDataframe, but the structure must be identical to institutions. Default = institutions

value
a character string. Defining the output value. See value.

verbose
logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic. Default = “clean”.

Details

Note: the buffer radius is in degrees, thus will differ slightly between different latitudes.

Note

See https://github.com/azizka/CoordinateCleaner/wiki for more details and tutorials.

Examples

x <- data.frame(species = letters[1:10],
                decimallongitude = runif(100, -180, 180),
                decimallatitude = runif(100, -90,90))

cc_inst(x, buffer = 5)#large buffer for demonstration

#> Testing biodiversity institutions

#> Flagged 25 records.

#>     species decimallongitude decimallatitude
#> 1         a       165.142044      83.2349348
#> 2         b       -50.414356     -69.8778162
#> 4         d        81.462714      87.0320793
#> 5         e      -119.472318     -57.7570492
#> 6         f       -98.466721     -65.1653206
#> 7         g       124.394420     -79.8194545
#> 9         i      -155.764862      -6.0291358
#> 10        j      -170.272997      71.3481074
#> 11        a      -137.503955     -86.6643420
#> 12        b      -175.069667      63.1753612
#> 13        c        84.862573       0.2562162
#> 14        d       152.625876     -60.0908803
#> 15        e       173.588616     -25.0189018
#> 16        f       -34.102727      27.3600331
#> 17        g       -82.909932      67.3518304
#> 19        i      -161.617517     -15.7548860
#> 20        j      -179.108341     -61.3387294
#> 22        b       -92.235138      69.1864384
#> 23        c       -87.930267     -18.7203396
#> 24        d        18.834760     -83.1346580
#> 25        e       -89.432054      81.2512740
#> 29        i       146.337012     -58.0033208
#> 30        j      -146.890874     -24.7470649
#> 31        a        85.606255     -19.5384165
#> 33        c        70.735887     -40.2377672
#> 34        d        72.276301     -64.9153228
#> 35        e      -165.621364      -8.7594652
#> 36        f       -24.651935     -70.1406272
#> 37        g       113.123292      68.5722564
#> 38        h        23.971060       6.0916869
#> 41        a        78.954467     -42.3932538
#> 42        b       157.114402      31.2941461
#> 43        c       -25.404619     -52.2681147
#> 45        e        97.431546     -54.9432872
#> 47        g      -122.661091     -87.2941899
#> 48        h       -35.424718      78.1510543
#> 49        i      -179.407264     -74.8014961
#> 50        j       -29.532227     -83.8705909
#> 51        a        95.016370     -34.0272911
#> 53        c       -92.668673     -39.5279016
#> 54        d        93.301543     -60.2489170
#> 55        e      -169.522505     -11.9529336
#> 56        f       -22.557318      56.5281359
#> 58        h       -25.260714      49.9151094
#> 59        i       111.346905     -73.5838245
#> 60        j       152.123263      43.6723012
#> 61        a        63.008450     -18.5986844
#> 62        b      -146.171451     -36.7149199
#> 64        d        -8.803755      -6.0283009
#> 66        f        13.708542     -86.0189309
#> 67        g       135.137884      56.7041759
#> 69        i       -99.084446      77.7232557
#> 70        j        68.730419       7.2775826
#> 72        b        -6.012656      -4.0342745
#> 73        c      -136.153210      36.8035184
#> 74        d        68.402605     -88.6490004
#> 76        f       113.827566     -42.1571584
#> 77        g        16.529558      84.7873685
#> 78        h       161.034186     -55.7054784
#> 79        i      -166.130494     -86.4547400
#> 81        a      -151.337156       4.6316536
#> 82        b       -24.144923      24.2136076
#> 83        c        93.921528     -18.8193673
#> 84        d       -74.246818     -58.5237203
#> 85        e      -141.360519     -31.0836309
#> 87        g       -51.726752     -75.7500460
#> 89        i        64.075701     -34.8693643
#> 92        b      -144.727900      44.6126844
#> 93        c      -144.356505     -64.9932261
#> 94        d        98.928570     -22.0782518
#> 95        e      -125.461599      25.4490184
#> 96        f       -37.941631      88.9550414
#> 98        h        86.252243     -29.2772609
#> 99        i        48.840524      74.4633798
#> 100       j       -58.240854      73.6956650

cc_inst(x, value = "flagged", buffer = 5)

#> Testing biodiversity institutions

#> Flagged 25 records.

#>     1     2     3     4     5     6     7     8     9    10    11    12    13 
#>  TRUE  TRUE FALSE  TRUE  TRUE  TRUE  TRUE FALSE  TRUE  TRUE  TRUE  TRUE  TRUE 
#>    14    15    16    17    18    19    20    21    22    23    24    25    26 
#>  TRUE  TRUE  TRUE  TRUE FALSE  TRUE  TRUE FALSE  TRUE  TRUE  TRUE  TRUE FALSE 
#>    27    28    29    30    31    32    33    34    35    36    37    38    39 
#> FALSE FALSE  TRUE  TRUE  TRUE FALSE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE 
#>    40    41    42    43    44    45    46    47    48    49    50    51    52 
#> FALSE  TRUE  TRUE  TRUE FALSE  TRUE FALSE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE 
#>    53    54    55    56    57    58    59    60    61    62    63    64    65 
#>  TRUE  TRUE  TRUE  TRUE FALSE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE  TRUE FALSE 
#>    66    67    68    69    70    71    72    73    74    75    76    77    78 
#>  TRUE  TRUE FALSE  TRUE  TRUE FALSE  TRUE  TRUE  TRUE FALSE  TRUE  TRUE  TRUE 
#>    79    80    81    82    83    84    85    86    87    88    89    90    91 
#>  TRUE FALSE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE  TRUE FALSE  TRUE FALSE FALSE 
#>    92    93    94    95    96    97    98    99   100 
#>  TRUE  TRUE  TRUE  TRUE  TRUE FALSE  TRUE  TRUE  TRUE