Identify Invalid lat/lon Coordinates

Removes or flags non-numeric and not available coordinates as well as lat >90, lat <-90, lon > 180 and lon < -180 are flagged.

cc_val(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Details

This test is obligatory before running any further tests of CoordinateCleaner, as additional tests only run with valid coordinates.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = c(runif(106, -180, 180), NA, "13W33'", "67,09", 305), 
                decimalLatitude = runif(110, -90,90))
                
cc_val(x)
#> Testing coordinate validity
#> Removed 4 records.
#>     species decimalLongitude decimalLatitude
#> 1         a      167.8200355       0.7527452
#> 2         b      165.7386564     -41.0133861
#> 3         c       35.1801717      44.2162359
#> 4         d      -65.5442393      80.7795499
#> 5         e       18.5613791      57.9227699
#> 6         f       76.0563049      15.0801763
#> 7         g      -33.6689577     -85.4419995
#> 8         h      -56.6991843     -61.0000567
#> 9         i     -126.9604005      30.5167549
#> 10        j      -12.5353516      63.9933407
#> 11        a      -94.9127060     -18.4434249
#> 12        b       25.3534425     -53.0948402
#> 13        c     -127.1075550       5.4051999
#> 14        d     -142.8804054     -54.0522192
#> 15        e     -130.7924646     -18.1945335
#> 16        f      122.5912078     -23.7073015
#> 17        g       99.4483794     -49.8969898
#> 18        h      175.3416432      68.0772010
#> 19        i      -56.2574407     -30.2070127
#> 20        j      112.4314028     -63.0833523
#> 21        a       52.3691387       6.4411277
#> 22        b      174.6374420      83.3535925
#> 23        c      111.2917144     -50.5744650
#> 24        d     -145.1468320     -72.5941046
#> 25        e       79.0469632      43.4855816
#> 26        f     -176.1422086     -56.5588692
#> 27        g     -116.4139933     -89.9816583
#> 28        h      -80.7983404     -68.5416074
#> 29        i      142.1069728      -7.5950451
#> 30        j     -103.6098733     -80.2197668
#> 31        a      -56.8383534     -29.0767759
#> 32        b     -128.6674049     -44.7535313
#> 33        c      -41.4632938     -72.4560455
#> 34        d       34.8511557      52.3868312
#> 35        e     -137.7963813      70.4586631
#> 36        f      -28.7814414     -79.3579670
#> 37        g      -53.3865287      28.4922759
#> 38        h      151.5788780     -24.9129824
#> 39        i      -48.2010783     -50.3568879
#> 40        j     -141.7833306     -88.6940565
#> 41        a       31.1605714      14.2723814
#> 42        b     -107.9367973      58.4662674
#> 43        c       67.7361463      15.3019758
#> 44        d       -7.0150350     -68.9251232
#> 45        e      120.6165878      -9.6798790
#> 46        f      138.2204784     -50.1321608
#> 47        g       77.9999845     -59.3442410
#> 48        h      111.1400756      -3.7334207
#> 49        i     -119.4696563      48.3690460
#> 50        j       93.8247430      72.6792598
#> 51        a       67.0121131       1.9533077
#> 52        b     -169.3551757      -5.7418488
#> 53        c      159.2645930      84.0624636
#> 54        d       66.5317347     -56.6378947
#> 55        e      -55.6778532      21.4327862
#> 56        f      174.4429331     -12.2089390
#> 57        g     -124.5456944     -81.0901986
#> 58        h      101.7980519     -22.0293941
#> 59        i     -107.6949932      60.9100241
#> 60        j      -27.3506639      16.0018165
#> 61        a     -125.5694057      -8.4857929
#> 62        b      104.7669681     -72.9988479
#> 63        c     -161.0293615       2.3075734
#> 64        d       94.4773572      20.6886480
#> 65        e      -54.9047433      23.0832822
#> 66        f      128.8263262     -34.9313677
#> 67        g      155.9582775     -37.3318005
#> 68        h       66.4384827      44.6162401
#> 69        i       88.2017188     -76.2952368
#> 70        j      -14.8862185      70.3133516
#> 71        a     -120.7618655      36.3156771
#> 72        b      -60.2524713      56.7958936
#> 73        c      -11.5924435     -31.6973475
#> 74        d        0.9176927     -79.3926008
#> 75        e      -38.3486870     -52.9448336
#> 76        f     -116.9585577      23.6122459
#> 77        g       49.3587388      79.7384745
#> 78        h     -163.1792561      16.6474016
#> 79        i      -21.4193593       3.6848462
#> 80        j      150.0074811      62.9095896
#> 81        a      175.2304372     -38.8241223
#> 82        b     -140.0265148      64.5724470
#> 83        c      -27.7687284      49.0217110
#> 84        d       35.3241917      23.3897270
#> 85        e       22.6140467      19.7364372
#> 86        f     -166.3619597     -37.4749799
#> 87        g      -71.7923062     -56.6691065
#> 88        h      145.7949044     -87.1351522
#> 89        i      173.3064415       6.7583454
#> 90        j      157.3886831      36.7918379
#> 91        a       82.2642515     -18.3303715
#> 92        b      154.8836432     -71.6974054
#> 93        c      143.8325026      85.7973214
#> 94        d      115.3384420      72.2078461
#> 95        e       88.4675117      84.7889067
#> 96        f      -92.1562303      60.3208527
#> 97        g       11.9483422     -71.9264776
#> 98        h      132.9117122      66.9223168
#> 99        i       77.3347929     -76.8244828
#> 100       j       89.5204893      56.2373807
#> 101       a        3.9434377      59.9372385
#> 102       b     -179.4414527     -38.9119614
#> 103       c      100.4053469     -14.2666447
#> 104       d     -118.0788353      14.3719628
#> 105       e     -132.4770147     -19.0624234
#> 106       f      -22.7200959     -26.9618367
cc_val(x, value = "flagged")
#> Testing coordinate validity
#> Flagged 4 records.
#>   [1]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [13]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [25]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [37]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [49]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [61]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [73]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [85]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [97]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE
#> [109] FALSE FALSE