Identify Coordinates in Artificial Hotspot Occurrence Inventory

Removes or flags records within Artificial Hotspot Occurrence Inventory. Poorly geo-referenced occurrence records in biological databases are often erroneously geo-referenced to highly recurring coordinates that were assessed by Park et al 2022. See the reference for more details.

cc_aohi(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  species = "species",
  taxa = c("Aves", "Insecta", "Mammalia", "Plantae"),
  buffer = 10000,
  geod = TRUE,
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

species

character string. The column with the species identity. Only required if verify = TRUE.

taxa

Artificial Hotspot Occurrence Inventory (AHOI) were created based on four different taxa, birds, insecta, mammalia, and plantae. Users can choose to keep all, or any specific taxa subset to define the AHOI locations. Default is to keep all: c("Aves", "Insecta", "Mammalia", "Plantae").

buffer

The buffer around each capital coordinate (the centre of the city), where records should be flagged as problematic. Units depend on geod. Default = 10 kilometres.

geod

logical. If TRUE the radius around each capital is calculated based on a sphere, buffer is in meters and independent of latitude. If FALSE the radius is calculated assuming planar coordinates and varies slightly with latitude. Default = TRUE. See https://seethedatablog.wordpress.com/ for detail and credits.

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

References

Park, D. S., Xie, Y., Thammavong, H. T., Tulaiha, R., & Feng, X. (2023). Artificial Hotspot Occurrence Inventory (AHOI). Journal of Biogeography, 50, 441–449. doi:10.1111/jbi.14543

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = c(runif(99, -180, 180), -47.92), 
                decimalLatitude = c(runif(99, -90,90), -15.78))
cc_aohi(x)
#> Testing Artificial Hotspot Occurrence Inventory
#> Removed 0 records.
#>     species decimalLongitude decimalLatitude
#> 1         a        64.147015     -37.5876667
#> 2         b        60.349796      47.1684101
#> 3         c      -125.123634     -17.8607673
#> 4         d      -127.865791      78.9006630
#> 5         e       -62.806929       0.6871756
#> 6         f       -71.566416       8.4782048
#> 7         g        24.921600     -22.9816552
#> 8         h       -89.701074      -1.9179661
#> 9         i      -153.138687      15.2405396
#> 10        j       -16.780628      38.0773643
#> 11        a       133.272172     -31.8068255
#> 12        b       119.580213      76.5042293
#> 13        c      -166.185798      -0.4144688
#> 14        d       -55.275007      50.8581928
#> 15        e      -160.013672      81.6400785
#> 16        f      -102.356848     -89.9844370
#> 17        g      -111.929036      49.7330658
#> 18        h       153.666430     -39.0878563
#> 19        i       -71.398707      -4.5530125
#> 20        j       119.581428      46.4956120
#> 21        a       -82.144655      72.6725621
#> 22        b        44.795055     -19.9090418
#> 23        c      -126.513145     -55.7567016
#> 24        d       -41.182623     -72.4709846
#> 25        e       -31.682694      21.9621235
#> 26        f      -147.087819     -54.6440937
#> 27        g      -115.377719     -62.6435326
#> 28        h       111.951932     -68.0702968
#> 29        i       168.984282      60.4309084
#> 30        j        80.145321     -10.3324255
#> 31        a        -1.427081     -31.1826084
#> 32        b      -163.988751      22.9373918
#> 33        c        19.627881     -20.8378050
#> 34        d       -42.154944      21.0778589
#> 35        e        19.552942     -11.3146987
#> 36        f       -17.397006     -19.4056039
#> 37        g      -178.911788      75.3883312
#> 38        h       -80.961661     -38.6585902
#> 39        i       147.626026     -54.4042500
#> 40        j        79.114218     -37.1618146
#> 41        a      -102.075730     -54.1241171
#> 42        b        35.090333     -36.6283035
#> 43        c       -78.534711      56.6407946
#> 44        d       -34.203728     -67.0945854
#> 45        e        19.427838      27.3415501
#> 46        f      -106.882331     -66.3162172
#> 47        g      -157.372217     -18.1184641
#> 48        h      -161.937511      41.1528953
#> 49        i        43.164113     -58.0985746
#> 50        j       -86.771657     -79.1798827
#> 51        a       100.399961      42.0499022
#> 52        b       158.280787      87.5335772
#> 53        c       -60.240849      44.2016348
#> 54        d      -123.862455     -52.6438690
#> 55        e        92.380488     -68.6173415
#> 56        f       174.047782      16.7648612
#> 57        g       104.012648      78.8277255
#> 58        h       123.926517     -22.4624404
#> 59        i        51.828655     -84.4619566
#> 60        j      -139.511500      36.1765754
#> 61        a        71.079948      65.4307672
#> 62        b        23.870060     -14.8704016
#> 63        c       -32.804141     -81.7376643
#> 64        d      -153.852499     -16.2331557
#> 65        e       101.178832      36.9503775
#> 66        f      -112.744550       9.9229794
#> 67        g      -123.494220      52.2204265
#> 68        h       -74.759182     -37.4477254
#> 69        i       102.730890     -34.3272275
#> 70        j        36.373370      72.5469128
#> 71        a        53.299649      79.0217713
#> 72        b         8.101166     -35.8051087
#> 73        c       106.618946     -65.1235262
#> 74        d       147.317455      63.3900572
#> 75        e        26.355182     -31.0466534
#> 76        f      -157.715859     -79.0705044
#> 77        g       166.276960     -35.3175676
#> 78        h        44.099090     -88.6063794
#> 79        i        86.689944      46.6222475
#> 80        j        61.262930     -52.0727434
#> 81        a        -8.097652     -21.6780280
#> 82        b      -138.568971      55.2386848
#> 83        c        -8.595039     -22.3839769
#> 84        d       -84.581456      22.8746303
#> 85        e       -63.746544      10.5910518
#> 86        f       168.001454     -84.2563249
#> 87        g        32.730836      79.7213388
#> 88        h       167.021003     -86.3705375
#> 89        i      -122.383848     -89.5388865
#> 90        j       -80.259775     -70.9638489
#> 91        a        72.730720     -54.7031472
#> 92        b       -90.274396      50.8211393
#> 93        c       163.632740      67.3439492
#> 94        d       138.511506      55.3802340
#> 95        e        -6.886823      30.1359776
#> 96        f      -155.277539      50.7968678
#> 97        g       106.865937     -88.2658800
#> 98        h        38.531554      57.1925561
#> 99        i       -53.472211      82.4245251
#> 100       j       -47.920000     -15.7800000