Identify Coordinates in Artificial Hotspot Occurrence Inventory

Removes or flags records within Artificial Hotspot Occurrence Inventory. Poorly geo-referenced occurrence records in biological databases are often erroneously geo-referenced to highly recurring coordinates that were assessed by Park et al 2022. See the reference for more details.

cc_aohi(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  species = "species",
  taxa = c("Aves", "Insecta", "Mammalia", "Plantae"),
  buffer = 10000,
  geod = TRUE,
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

species

character string. The column with the species identity. Only required if verify = TRUE.

taxa

Artificial Hotspot Occurrence Inventory (AHOI) were created based on four different taxa, birds, insecta, mammalia, and plantae. Users can choose to keep all, or any specific taxa subset to define the AHOI locations. Default is to keep all: c("Aves", "Insecta", "Mammalia", "Plantae").

buffer

The buffer around each capital coordinate (the centre of the city), where records should be flagged as problematic. Units depend on geod. Default = 10 kilometres.

geod

logical. If TRUE the radius around each capital is calculated based on a sphere, buffer is in meters and independent of latitude. If FALSE the radius is calculated assuming planar coordinates and varies slightly with latitude. Default = TRUE. See https://seethedatablog.wordpress.com/ for detail and credits.

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

References

Park, D. S., Xie, Y., Thammavong, H. T., Tulaiha, R., & Feng, X. (2023). Artificial Hotspot Occurrence Inventory (AHOI). Journal of Biogeography, 50, 441–449. doi:10.1111/jbi.14543

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = c(runif(99, -180, 180), -47.92), 
                decimalLatitude = c(runif(99, -90,90), -15.78))
cc_aohi(x)
#> Testing Artificial Hotspot Occurrence Inventory
#> Removed 0 records.
#>     species decimalLongitude decimalLatitude
#> 1         a      -26.5250210       4.0267364
#> 2         b     -116.6167847      60.8293673
#> 3         c      111.1434450      85.6243042
#> 4         d     -159.9233881     -39.7291072
#> 5         e       74.7227771      65.3264759
#> 6         f     -123.3636128       2.8298876
#> 7         g      -84.6038439     -56.7913391
#> 8         h      -87.3944503     -60.2249849
#> 9         i       12.6168823      14.6963845
#> 10        j      -10.2123331      66.8725008
#> 11        a      -23.8579466      66.7716199
#> 12        b      158.6138181     -36.9199256
#> 13        c       89.1249881      88.0166110
#> 14        d      -77.1748993     -45.9386710
#> 15        e      -53.0177577      77.1312057
#> 16        f     -177.4246884      74.8869185
#> 17        g      -70.2533920     -68.8266402
#> 18        h      -76.0814287      70.1167657
#> 19        i       60.2683821     -47.7492981
#> 20        j       57.5697266      20.8827530
#> 21        a      -18.2142582      19.7346421
#> 22        b      -93.0347685     -82.3803818
#> 23        c      142.4289350      62.9700714
#> 24        d      -50.5532112      14.4821842
#> 25        e     -149.8041325     -76.9088423
#> 26        f      169.5782185      29.5827777
#> 27        g       31.2202314      53.9851449
#> 28        h      -63.6067725      19.4555446
#> 29        i      -25.7692877      -0.9265465
#> 30        j      162.8931888      61.0082774
#> 31        a     -162.3318363       8.8450656
#> 32        b      -48.4988947      63.6785533
#> 33        c      -60.0790583      21.1422264
#> 34        d      -10.2041795     -49.9991264
#> 35        e      166.0327597      -5.6497445
#> 36        f      171.3915606      74.7518153
#> 37        g      -75.0866507      50.1046202
#> 38        h      -60.4040343      -1.1886876
#> 39        i       88.4374632      36.0619040
#> 40        j      115.4215769      23.1649995
#> 41        a     -167.9494554     -53.4656485
#> 42        b     -156.1984548     -55.2976425
#> 43        c      -70.2593737     -28.5864465
#> 44        d      166.2392996      16.8774409
#> 45        e       39.1340956      33.0827841
#> 46        f      -38.1192809      42.6227890
#> 47        g     -116.0806415     -39.1845424
#> 48        h     -103.1207617      65.5950111
#> 49        i       11.5074385      37.3274580
#> 50        j      -38.2016672     -15.1750717
#> 51        a      -44.4297789      52.7409072
#> 52        b       32.6548616      38.7602908
#> 53        c      -96.1961451     -10.6383789
#> 54        d       94.7467603     -70.2830307
#> 55        e      155.4143972      77.4338166
#> 56        f     -127.4982267      63.9888094
#> 57        g     -120.2330453      81.3051420
#> 58        h     -155.1838074     -87.4514039
#> 59        i     -129.1129788     -66.7626641
#> 60        j      -45.6862832      69.5213772
#> 61        a      -19.1569681      22.2116568
#> 62        b      158.8338304     -24.7002085
#> 63        c       67.9480847      58.1022586
#> 64        d     -105.8502168      87.8234995
#> 65        e     -100.1433867     -47.0627692
#> 66        f       71.8658366     -36.3659685
#> 67        g      149.1554171      34.5111313
#> 68        h       -4.8162587      30.0702953
#> 69        i      135.4692928     -18.8882065
#> 70        j     -157.0177817      77.0503464
#> 71        a        0.6129734     -15.5539247
#> 72        b     -148.0660713      68.9389165
#> 73        c     -168.9724746     -48.7598259
#> 74        d      -61.0373698       1.3777877
#> 75        e       -1.0174117      -2.3038621
#> 76        f       70.1987109       3.4757226
#> 77        g       29.8733585      -1.1322760
#> 78        h     -140.9931499     -43.0571051
#> 79        i      158.8733010      34.7741065
#> 80        j     -167.1372227     -71.1130957
#> 81        a     -130.8732787      71.9242841
#> 82        b     -105.6665555      41.0414289
#> 83        c      -51.8644908     -81.7656545
#> 84        d     -103.1490174      85.6161560
#> 85        e        2.3461231     -11.2593740
#> 86        f     -163.4514510      34.7910198
#> 87        g      117.7848760     -42.0229751
#> 88        h     -155.1178008     -69.7589166
#> 89        i       95.3152173      47.0784829
#> 90        j     -128.5677019     -51.2509710
#> 91        a     -122.3863149      19.0655161
#> 92        b      -65.6866635     -29.6909838
#> 93        c     -104.1744847     -59.3055324
#> 94        d     -114.3676872     -17.6299500
#> 95        e      -49.9573637      -0.8113309
#> 96        f     -128.2259279     -43.6620382
#> 97        g      -20.1252972     -61.6691945
#> 98        h      179.2157875     -24.2918829
#> 99        i      -61.1335749     -23.4686152
#> 100       j      -47.9200000     -15.7800000