Identify Coordinates in Artificial Hotspot Occurrence Inventory

Removes or flags records within Artificial Hotspot Occurrence Inventory. Poorly geo-referenced occurrence records in biological databases are often erroneously geo-referenced to highly recurring coordinates that were assessed by Park et al 2022. See the reference for more details.

cc_aohi(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  species = "species",
  taxa = c("Aves", "Insecta", "Mammalia", "Plantae"),
  buffer = 10000,
  geod = TRUE,
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

species

character string. The column with the species identity. Only required if verify = TRUE.

taxa

Artificial Hotspot Occurrence Inventory (AHOI) were created based on four different taxa, birds, insecta, mammalia, and plantae. Users can choose to keep all, or any specific taxa subset to define the AHOI locations. Default is to keep all: c("Aves", "Insecta", "Mammalia", "Plantae").

buffer

The buffer around each capital coordinate (the centre of the city), where records should be flagged as problematic. Units depend on geod. Default = 10 kilometres.

geod

logical. If TRUE the radius around each capital is calculated based on a sphere, buffer is in meters and independent of latitude. If FALSE the radius is calculated assuming planar coordinates and varies slightly with latitude. Default = TRUE. See https://seethedatablog.wordpress.com/ for detail and credits.

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

References

Park, D. S., Xie, Y., Thammavong, H. T., Tulaiha, R., & Feng, X. (2023). Artificial Hotspot Occurrence Inventory (AHOI). Journal of Biogeography, 50, 441–449. doi:10.1111/jbi.14543

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = c(runif(99, -180, 180), -47.92), 
                decimalLatitude = c(runif(99, -90,90), -15.78))
cc_aohi(x)
#> Testing Artificial Hotspot Occurrence Inventory
#> Removed 0 records.
#>     species decimalLongitude decimalLatitude
#> 1         a      -102.444756     -67.9142178
#> 2         b         6.633454     -59.4818929
#> 3         c      -158.501649     -62.4810201
#> 4         d       -77.301194      42.9720624
#> 5         e        -6.982040     -18.2215759
#> 6         f       169.245493      40.2589019
#> 7         g         5.585960      57.1191775
#> 8         h       -82.892491      32.5140481
#> 9         i      -113.302243      58.7866727
#> 10        j        66.891333     -83.5900682
#> 11        a       148.767159     -30.7130134
#> 12        b       -46.479870      64.3101572
#> 13        c        22.482662     -56.0437360
#> 14        d       151.911133     -22.6681637
#> 15        e       119.048824      31.0174689
#> 16        f        17.638579     -47.0443919
#> 17        g       -59.800275      31.5580338
#> 18        h        35.865107       9.0855622
#> 19        i        52.442980       9.0898423
#> 20        j       128.780775      61.4353686
#> 21        a      -107.809691     -37.9752458
#> 22        b       174.051903      -5.1715969
#> 23        c       -38.369112     -31.9029887
#> 24        d      -169.152005     -41.6713602
#> 25        e        10.910921      68.5373891
#> 26        f        24.980039      57.8327944
#> 27        g       -86.860940     -11.7065613
#> 28        h       104.711662      55.9340697
#> 29        i       112.127422     -88.6309517
#> 30        j      -100.117019      69.6806372
#> 31        a        24.125551     -33.4183706
#> 32        b      -171.414089     -86.3869216
#> 33        c        -3.629386      20.8329193
#> 34        d        76.866963       0.5509546
#> 35        e      -131.466205     -77.2334552
#> 36        f      -169.764151      80.6850101
#> 37        g       147.380993      88.4577851
#> 38        h        44.259704     -38.2256582
#> 39        i       153.012350      -9.8837140
#> 40        j      -172.307609     -25.1187391
#> 41        a       -85.222959     -57.2614432
#> 42        b      -153.713993      85.5556662
#> 43        c       -41.984095      53.0905053
#> 44        d       115.776813     -45.2501237
#> 45        e       -77.810879       5.8236371
#> 46        f       -26.214419      54.1696849
#> 47        g      -119.171170      80.8820151
#> 48        h      -172.700852      73.4662763
#> 49        i        24.946838      10.2839259
#> 50        j      -160.675071      -4.9100518
#> 51        a       -31.461164      55.5212018
#> 52        b        15.134922     -32.9778001
#> 53        c       -99.091272     -15.9615712
#> 54        d       142.990335       8.3781636
#> 55        e        23.507833       8.3122934
#> 56        f      -160.812822      73.1762245
#> 57        g        50.319632     -53.3117985
#> 58        h         6.631961     -18.6071240
#> 59        i       -63.232329      66.5153514
#> 60        j       -44.909011     -34.0547295
#> 61        a      -150.329575     -52.1028866
#> 62        b        53.007560     -75.2064423
#> 63        c       -58.638774     -71.1089119
#> 64        d       -42.909008      62.1277783
#> 65        e        52.001873      84.3860597
#> 66        f        -5.699570     -29.0382871
#> 67        g       -60.693549      81.5109343
#> 68        h       -67.228566      57.5970553
#> 69        i       -47.397835      83.8038155
#> 70        j       -28.397039      61.0220101
#> 71        a        43.025460     -26.4566965
#> 72        b      -168.350628      46.2702419
#> 73        c       -43.436184      84.3342664
#> 74        d       -31.933118      14.9344489
#> 75        e       -15.443398     -76.6882161
#> 76        f       105.376069     -15.5620084
#> 77        g       156.890488     -67.9877812
#> 78        h        21.334338     -64.8752720
#> 79        i       -90.302300     -59.9278006
#> 80        j       166.640805      72.4662721
#> 81        a      -151.778909      16.8958194
#> 82        b        51.273374      81.7098124
#> 83        c        42.302677      39.3758056
#> 84        d        70.393780       7.5351093
#> 85        e        24.518857     -60.4422248
#> 86        f      -165.794770     -26.0531459
#> 87        g       -81.829900      37.0496394
#> 88        h      -131.568031      -3.6792810
#> 89        i        89.186688      64.8096767
#> 90        j       107.976785      89.4606694
#> 91        a        56.849793       9.4008437
#> 92        b       157.294133      19.3239901
#> 93        c      -127.010474     -58.8222142
#> 94        d       105.280704     -86.6468881
#> 95        e       -65.283136     -33.7024382
#> 96        f      -152.316325      40.1623155
#> 97        g       -83.822487     -74.7969052
#> 98        h      -166.068259     -60.9875779
#> 99        i       146.529118     -69.9638868
#> 100       j       -47.920000     -15.7800000