Identify Coordinates in Artificial Hotspot Occurrence Inventory

Removes or flags records within Artificial Hotspot Occurrence Inventory. Poorly geo-referenced occurrence records in biological databases are often erroneously geo-referenced to highly recurring coordinates that were assessed by Park et al 2022. See the reference for more details.

cc_aohi(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  species = "species",
  taxa = c("Aves", "Insecta", "Mammalia", "Plantae"),
  buffer = 10000,
  geod = TRUE,
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

species

character string. The column with the species identity. Only required if verify = TRUE.

taxa

Artificial Hotspot Occurrence Inventory (AHOI) were created based on four different taxa, birds, insecta, mammalia, and plantae. Users can choose to keep all, or any specific taxa subset to define the AHOI locations. Default is to keep all: c("Aves", "Insecta", "Mammalia", "Plantae").

buffer

The buffer around each capital coordinate (the centre of the city), where records should be flagged as problematic. Units depend on geod. Default = 10 kilometres.

geod

logical. If TRUE the radius around each capital is calculated based on a sphere, buffer is in meters and independent of latitude. If FALSE the radius is calculated assuming planar coordinates and varies slightly with latitude. Default = TRUE. See https://seethedatablog.wordpress.com/ for detail and credits.

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

References

Park, D. S., Xie, Y., Thammavong, H. T., Tulaiha, R., & Feng, X. (2023). Artificial Hotspot Occurrence Inventory (AHOI). Journal of Biogeography, 50, 441–449. doi:10.1111/jbi.14543

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = c(runif(99, -180, 180), -47.92), 
                decimalLatitude = c(runif(99, -90,90), -15.78))
cc_aohi(x)
#> Testing Artificial Hotspot Occurrence Inventory
#> Removed 0 records.
#>     species decimalLongitude decimalLatitude
#> 1         a        85.271352      31.4698387
#> 2         b      -126.800238      29.7802857
#> 3         c      -143.017507     -57.4485735
#> 4         d       -21.020269      14.8337039
#> 5         e       -37.557567      67.2469978
#> 6         f       -26.358515      28.0343644
#> 7         g       -72.024202      83.9770154
#> 8         h      -168.333797      12.3515425
#> 9         i       -13.768919      -1.4833400
#> 10        j       -31.770203       7.1139744
#> 11        a        19.945221     -79.1834145
#> 12        b      -163.085421     -52.6238743
#> 13        c       -81.173784     -57.8150113
#> 14        d        95.312267     -53.0020135
#> 15        e      -105.873249     -74.2835421
#> 16        f        -5.803793     -77.1478418
#> 17        g       142.892290     -32.9485349
#> 18        h      -167.091361      38.3277820
#> 19        i       136.037781      47.7472602
#> 20        j        30.739763     -62.7122266
#> 21        a        57.768487     -80.8173356
#> 22        b      -126.571204     -55.2367902
#> 23        c         6.112589     -84.0787311
#> 24        d       -46.909128     -89.6447246
#> 25        e       -96.547364      32.5190304
#> 26        f       157.028650      77.1999742
#> 27        g      -149.950034     -42.7778384
#> 28        h       179.003711     -57.7728274
#> 29        i       178.810902     -35.9811165
#> 30        j       -77.307947      65.2654021
#> 31        a        -2.724008      52.8531649
#> 32        b       -19.634357      50.6976990
#> 33        c       -88.597456     -36.9499248
#> 34        d       109.623449      64.7731987
#> 35        e        44.115881     -36.4719399
#> 36        f       -59.276596     -16.8663002
#> 37        g      -152.026272     -79.3157239
#> 38        h       157.045490      53.3655744
#> 39        i       -90.947169      73.6321330
#> 40        j       -72.006525     -59.8220501
#> 41        a       -77.996266      80.3438808
#> 42        b       -46.480010     -74.4917062
#> 43        c       -84.188943      42.2882718
#> 44        d       -87.163755     -57.3928763
#> 45        e       -93.221285      17.6897863
#> 46        f      -119.903435      81.2486437
#> 47        g       -56.063063     -82.4616063
#> 48        h       -65.253403      64.1885244
#> 49        i       -35.687960      34.0618526
#> 50        j       -62.987449      56.2477703
#> 51        a        19.530504     -19.1289283
#> 52        b       -37.823556      67.2541774
#> 53        c       -20.705673      67.0647773
#> 54        d       -55.828080     -20.9691305
#> 55        e      -141.489607      52.5852966
#> 56        f      -121.314849      39.9174919
#> 57        g       -99.965771     -21.5535289
#> 58        h        92.042553      -4.6596156
#> 59        i         1.495173      75.4914849
#> 60        j       -60.814501      33.1590650
#> 61        a       -95.556031      37.5973982
#> 62        b       -74.331229     -78.8041486
#> 63        c       -62.310140      81.7233915
#> 64        d        33.767733     -76.4933954
#> 65        e       -45.369504     -86.1872210
#> 66        f      -170.386316      31.1150972
#> 67        g        12.151984      41.2707482
#> 68        h      -174.146021      37.7323204
#> 69        i       -75.503489      57.9158972
#> 70        j        97.203556      77.2379119
#> 71        a      -150.735654       9.1882392
#> 72        b        -2.220906     -14.1506125
#> 73        c       -20.611316      40.1587826
#> 74        d      -132.913303      24.8524336
#> 75        e      -134.618627      11.9459834
#> 76        f      -175.029631     -73.4647437
#> 77        g       -43.301089      14.7143056
#> 78        h      -138.722252     -60.9486610
#> 79        i        34.303997     -22.9203896
#> 80        j       -31.956453     -60.3763876
#> 81        a       -25.686373     -59.1360353
#> 82        b      -159.991969      23.4022871
#> 83        c        39.122918     -49.1754706
#> 84        d      -135.598276      -7.6211445
#> 85        e      -153.605285      -7.7045308
#> 86        f       179.045912     -53.3901947
#> 87        g       121.104518     -61.5593510
#> 88        h       159.521847      49.8370374
#> 89        i       113.061336      87.7179267
#> 90        j       -53.671180       2.0626675
#> 91        a      -121.614612      46.1279685
#> 92        b       176.739002     -35.9690616
#> 93        c       105.115562     -47.6504744
#> 94        d       -81.449586     -41.5018547
#> 95        e        79.314434     -62.3986557
#> 96        f        91.389147      -0.9169858
#> 97        g        28.092744     -61.9125066
#> 98        h      -135.076764     -64.0494027
#> 99        i       -21.037637       2.6078775
#> 100       j       -47.920000     -15.7800000