Identify Coordinates in Artificial Hotspot Occurrence Inventory

Removes or flags records within Artificial Hotspot Occurrence Inventory. Poorly geo-referenced occurrence records in biological databases are often erroneously geo-referenced to highly recurring coordinates that were assessed by Park et al 2022. See the reference for more details.

cc_aohi(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  species = "species",
  taxa = c("Aves", "Insecta", "Mammalia", "Plantae"),
  buffer = 10000,
  geod = TRUE,
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

species

character string. The column with the species identity. Only required if verify = TRUE.

taxa

Artificial Hotspot Occurrence Inventory (AHOI) were created based on four different taxa, birds, insecta, mammalia, and plantae. Users can choose to keep all, or any specific taxa subset to define the AHOI locations. Default is to keep all: c("Aves", "Insecta", "Mammalia", "Plantae").

buffer

The buffer around each capital coordinate (the centre of the city), where records should be flagged as problematic. Units depend on geod. Default = 10 kilometres.

geod

logical. If TRUE the radius around each capital is calculated based on a sphere, buffer is in meters and independent of latitude. If FALSE the radius is calculated assuming planar coordinates and varies slightly with latitude. Default = TRUE. See https://seethedatablog.wordpress.com/ for detail and credits.

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

References

Park, D. S., Xie, Y., Thammavong, H. T., Tulaiha, R., & Feng, X. (2023). Artificial Hotspot Occurrence Inventory (AHOI). Journal of Biogeography, 50, 441–449. doi:10.1111/jbi.14543

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = c(runif(99, -180, 180), -47.92), 
                decimalLatitude = c(runif(99, -90,90), -15.78))
cc_aohi(x)
#> Testing Artificial Hotspot Occurrence Inventory
#> Removed 0 records.
#>     species decimalLongitude decimalLatitude
#> 1         a      -142.936980       85.640165
#> 2         b       -49.668961      -65.911499
#> 3         c      -170.500903       41.532576
#> 4         d        -2.128664      -43.481075
#> 5         e       -18.380406       89.490995
#> 6         f       172.526572       67.278413
#> 7         g      -126.877395      -37.721137
#> 8         h       167.770920      -67.592024
#> 9         i       -78.743126      -64.935926
#> 10        j       -52.743888       70.202411
#> 11        a       110.216049       67.563085
#> 12        b       146.826453       33.977513
#> 13        c       107.525907      -57.491472
#> 14        d       125.475590      -80.731008
#> 15        e       140.252009       41.407505
#> 16        f       -76.170982       41.269403
#> 17        g        77.505699      -16.102837
#> 18        h      -146.781788       45.479261
#> 19        i        46.488800       47.321069
#> 20        j       145.814876       54.371134
#> 21        a       167.038948      -38.077849
#> 22        b       153.437049       83.315111
#> 23        c       -80.634819       -4.930198
#> 24        d        38.693967       73.274306
#> 25        e      -118.123921      -67.127156
#> 26        f        37.463570      -64.566217
#> 27        g       102.605898      -84.699716
#> 28        h      -172.895109       21.959342
#> 29        i       132.598154       89.174094
#> 30        j        12.396487      -14.208227
#> 31        a       112.149250      -61.989829
#> 32        b      -121.878549       34.285969
#> 33        c      -165.088697       72.876901
#> 34        d        42.178581      -15.981047
#> 35        e      -110.613115        1.291127
#> 36        f      -131.004762      -86.266919
#> 37        g       129.811062       49.699741
#> 38        h       171.693334      -55.800784
#> 39        i       140.669934      -75.866726
#> 40        j      -118.127711      -55.924545
#> 41        a        -2.621668       76.200309
#> 42        b        78.216742       20.733444
#> 43        c       138.421265        6.534669
#> 44        d       112.703536      -59.725326
#> 45        e      -139.039099      -36.868728
#> 46        f       151.364139       51.809521
#> 47        g        69.579481      -30.131752
#> 48        h       -56.022358      -35.883135
#> 49        i      -166.942470        3.686954
#> 50        j        63.123109      -13.802441
#> 51        a      -105.671206       35.196314
#> 52        b       101.879529      -66.809378
#> 53        c       -58.277137       68.581432
#> 54        d       163.909477       32.768640
#> 55        e       -74.960005      -50.205525
#> 56        f      -104.831673       28.102323
#> 57        g       -99.809532      -45.517037
#> 58        h       -69.228660      -23.290194
#> 59        i       -87.759439      -82.075001
#> 60        j       -90.075100      -43.790298
#> 61        a       -55.423219        7.685144
#> 62        b       128.549972       83.525391
#> 63        c       -17.493172       12.217635
#> 64        d       -13.588646       -9.442634
#> 65        e        16.035660       25.318003
#> 66        f       106.969108       80.839103
#> 67        g        73.203742      -43.835658
#> 68        h      -102.206317      -87.123657
#> 69        i       161.869738      -80.747693
#> 70        j      -135.212202        6.134840
#> 71        a        45.101961       37.041357
#> 72        b        91.318538       44.378729
#> 73        c      -105.926103      -82.681743
#> 74        d      -175.797343       11.738101
#> 75        e        26.697127       27.855957
#> 76        f       -87.025961      -63.536100
#> 77        g         7.000247      -61.374438
#> 78        h      -171.194685      -78.334702
#> 79        i        98.934190       23.978925
#> 80        j       132.208185        2.041450
#> 81        a       170.147429      -51.786940
#> 82        b      -139.445578       64.013848
#> 83        c        -2.762345       73.913891
#> 84        d       -89.324401       38.615406
#> 85        e      -113.621654       22.302480
#> 86        f      -162.469346      -79.003914
#> 87        g      -173.632662      -58.257804
#> 88        h       177.500454      -73.839117
#> 89        i         8.974187       17.163675
#> 90        j      -120.610697       42.876391
#> 91        a       155.443671      -85.201474
#> 92        b       -50.275473      -48.151974
#> 93        c       -72.042830       89.334708
#> 94        d       155.006700      -53.662788
#> 95        e       104.741325       52.178026
#> 96        f      -140.755441       50.675285
#> 97        g       -47.889156       19.055169
#> 98        h       -43.559183      -87.394613
#> 99        i       145.529197       19.604552
#> 100       j       -47.920000      -15.780000