Identify Coordinates in Artificial Hotspot Occurrence Inventory

Removes or flags records within Artificial Hotspot Occurrence Inventory. Poorly geo-referenced occurrence records in biological databases are often erroneously geo-referenced to highly recurring coordinates that were assessed by Park et al 2022. See the reference for more details.

cc_aohi(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  species = "species",
  taxa = c("Aves", "Insecta", "Mammalia", "Plantae"),
  buffer = 10000,
  geod = TRUE,
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

species

character string. The column with the species identity. Only required if verify = TRUE.

taxa

Artificial Hotspot Occurrence Inventory (AHOI) were created based on four different taxa, birds, insecta, mammalia, and plantae. Users can choose to keep all, or any specific taxa subset to define the AHOI locations. Default is to keep all: c("Aves", "Insecta", "Mammalia", "Plantae").

buffer

The buffer around each capital coordinate (the centre of the city), where records should be flagged as problematic. Units depend on geod. Default = 10 kilometres.

geod

logical. If TRUE the radius around each capital is calculated based on a sphere, buffer is in meters and independent of latitude. If FALSE the radius is calculated assuming planar coordinates and varies slightly with latitude. Default = TRUE. See https://seethedatablog.wordpress.com/ for detail and credits.

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

References

Park, D. S., Xie, Y., Thammavong, H. T., Tulaiha, R., & Feng, X. (2023). Artificial Hotspot Occurrence Inventory (AHOI). Journal of Biogeography, 50, 441–449. doi:10.1111/jbi.14543

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = c(runif(99, -180, 180), -47.92), 
                decimalLatitude = c(runif(99, -90,90), -15.78))
cc_aohi(x)
#> Testing Artificial Hotspot Occurrence Inventory
#> Removed 0 records.
#>     species decimalLongitude decimalLatitude
#> 1         a       -60.614739      69.8037751
#> 2         b       -76.186168      41.4666950
#> 3         c      -135.598116      -3.7958537
#> 4         d       -26.505474      89.3032404
#> 5         e      -103.456803     -82.0704419
#> 6         f       -62.188323     -21.7674561
#> 7         g       166.509195      -0.7642725
#> 8         h       -73.869476     -29.1937553
#> 9         i       -22.861572      25.5790046
#> 10        j      -145.986453      45.6946261
#> 11        a       145.721623      -1.8185689
#> 12        b      -118.561003     -65.8575294
#> 13        c       100.198456       2.3289766
#> 14        d        -8.162139     -37.2345444
#> 15        e      -166.774472      48.8943071
#> 16        f       146.537943      52.9794582
#> 17        g       -94.689302      69.2544224
#> 18        h      -117.648622     -40.8552066
#> 19        i       147.206452     -18.5821806
#> 20        j        19.020835     -83.3507467
#> 21        a         1.173687     -69.4669274
#> 22        b       123.190996     -65.2700056
#> 23        c       -12.852102     -77.8053982
#> 24        d       122.973993     -52.4724948
#> 25        e      -152.631857      44.6168402
#> 26        f       -26.393329      31.1681064
#> 27        g        -7.620308     -47.8607922
#> 28        h       -17.382955     -26.4799218
#> 29        i       104.376511     -65.9246974
#> 30        j        88.104248      -1.6233564
#> 31        a        96.025397     -62.4314564
#> 32        b       -73.386194      21.7611183
#> 33        c      -153.053086     -89.7586470
#> 34        d        74.913422     -81.8691212
#> 35        e       -46.425843      19.4187618
#> 36        f         9.287558     -67.9365754
#> 37        g       143.364155      56.2799582
#> 38        h        26.381284     -34.6861251
#> 39        i       -62.189749     -72.4631026
#> 40        j        94.472348     -88.4005194
#> 41        a      -171.884510     -54.8303588
#> 42        b      -136.402751      45.6323203
#> 43        c       -56.396130      21.7330908
#> 44        d       175.863048     -65.3460810
#> 45        e        63.154674     -34.4282660
#> 46        f        68.988764      15.4242234
#> 47        g       129.249761     -62.4944622
#> 48        h      -133.037138      43.2562256
#> 49        i      -123.508035     -63.0720917
#> 50        j       163.875180      43.7435178
#> 51        a       -68.322592     -80.4768770
#> 52        b        12.771377     -83.6134612
#> 53        c       143.248624      53.0137331
#> 54        d       -65.779684      -2.2092545
#> 55        e      -172.093776      49.2382544
#> 56        f      -167.966980     -23.4763778
#> 57        g       -87.421111      49.2313775
#> 58        h         4.819786      25.6155119
#> 59        i       -64.367384      31.7086656
#> 60        j       -46.370719     -50.3154272
#> 61        a       -54.707647      70.1846945
#> 62        b       -62.885373      46.4365752
#> 63        c       171.837361     -60.9367551
#> 64        d       -22.382906      21.6034827
#> 65        e        39.470256      42.9730756
#> 66        f       107.599831     -63.8456112
#> 67        g      -152.712366      66.6333216
#> 68        h        51.744977       4.9198072
#> 69        i        61.596638     -81.9770626
#> 70        j       133.319239       6.2433631
#> 71        a        39.991089      88.0539959
#> 72        b        51.962831      12.9889366
#> 73        c      -114.840411       2.5008643
#> 74        d       151.593962     -43.9793544
#> 75        e        33.167891     -49.8558808
#> 76        f      -168.877077      49.4507040
#> 77        g       -30.379011     -88.3532251
#> 78        h       -14.164873      50.5781294
#> 79        i        44.767004      23.9203632
#> 80        j        17.414134     -48.3834361
#> 81        a      -106.917748      46.9392199
#> 82        b       -60.948698      44.9359355
#> 83        c        44.924127     -74.5325690
#> 84        d       -91.883649      44.7879672
#> 85        e       -65.797593      89.1104025
#> 86        f       -76.058795      18.0393650
#> 87        g       -25.575630     -38.4439361
#> 88        h       113.426338      63.8326282
#> 89        i      -163.833146     -39.7489681
#> 90        j       114.107864     -28.4439430
#> 91        a       175.882027     -78.8977246
#> 92        b       -14.733118      70.4852810
#> 93        c       -42.230610     -77.4339603
#> 94        d      -110.719953     -50.1356676
#> 95        e       178.163966      44.7497847
#> 96        f       -91.716404     -21.9541636
#> 97        g       -82.200940      21.4590512
#> 98        h       127.583157      53.8256330
#> 99        i      -126.859067      44.9484599
#> 100       j       -47.920000     -15.7800000