Identify Coordinates in Artificial Hotspot Occurrence Inventory

Removes or flags records within Artificial Hotspot Occurrence Inventory. Poorly geo-referenced occurrence records in biological databases are often erroneously geo-referenced to highly recurring coordinates that were assessed by Park et al 2022. See the reference for more details.

cc_aohi(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  species = "species",
  taxa = c("Aves", "Insecta", "Mammalia", "Plantae"),
  buffer = 10000,
  geod = TRUE,
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

species

character string. The column with the species identity. Only required if verify = TRUE.

taxa

Artificial Hotspot Occurrence Inventory (AHOI) were created based on four different taxa, birds, insecta, mammalia, and plantae. Users can choose to keep all, or any specific taxa subset to define the AHOI locations. Default is to keep all: c("Aves", "Insecta", "Mammalia", "Plantae").

buffer

The buffer around each capital coordinate (the centre of the city), where records should be flagged as problematic. Units depend on geod. Default = 10 kilometres.

geod

logical. If TRUE the radius around each capital is calculated based on a sphere, buffer is in meters and independent of latitude. If FALSE the radius is calculated assuming planar coordinates and varies slightly with latitude. Default = TRUE. See https://seethedatablog.wordpress.com/ for detail and credits.

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

References

Park, D. S., Xie, Y., Thammavong, H. T., Tulaiha, R., & Feng, X. (2023). Artificial Hotspot Occurrence Inventory (AHOI). Journal of Biogeography, 50, 441–449. doi:10.1111/jbi.14543

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = c(runif(99, -180, 180), -47.92), 
                decimalLatitude = c(runif(99, -90,90), -15.78))
cc_aohi(x)
#> Testing Artificial Hotspot Occurrence Inventory
#> Removed 0 records.
#>     species decimalLongitude decimalLatitude
#> 1         a       -58.115295     -52.3153036
#> 2         b        80.073378     -69.9964436
#> 3         c        99.958795      39.0979547
#> 4         d      -116.700602       5.1111889
#> 5         e      -165.432025     -29.2779475
#> 6         f       -48.060586      89.7458722
#> 7         g       124.164267     -10.9916269
#> 8         h      -152.131187     -61.6684135
#> 9         i        17.594311     -36.3003906
#> 10        j        -8.292747     -41.9030603
#> 11        a       151.233883       6.7965728
#> 12        b       128.895799       0.7370308
#> 13        c        89.358549      58.6599060
#> 14        d      -162.106939     -53.1836256
#> 15        e      -104.783346      86.5073460
#> 16        f       -37.258703     -21.4740891
#> 17        g        35.921369      76.0743404
#> 18        h      -155.851859      79.1958372
#> 19        i      -118.035635      66.5588282
#> 20        j        94.700890      -5.7429709
#> 21        a      -108.160872      47.8469663
#> 22        b       159.186013      78.6015861
#> 23        c       172.375546     -43.3268759
#> 24        d        60.450388      89.0077189
#> 25        e        13.189372      58.0724052
#> 26        f       -83.686230     -53.7135952
#> 27        g       161.155456      32.5979634
#> 28        h       -77.353071       1.7145866
#> 29        i        68.226317      -5.7887634
#> 30        j      -161.129883      -7.1149798
#> 31        a       -48.725126      51.1903530
#> 32        b      -137.324331     -75.3461156
#> 33        c       142.982975      59.9717635
#> 34        d        29.248052      57.3413224
#> 35        e       171.362159     -68.2837211
#> 36        f        11.206368     -76.1918532
#> 37        g        63.722987      62.8534124
#> 38        h       162.945246     -46.3016666
#> 39        i       129.354501     -82.3109469
#> 40        j        84.705657     -21.3727432
#> 41        a         3.168819      -5.5841839
#> 42        b        49.373176     -80.6294592
#> 43        c        82.074947      62.4158222
#> 44        d       132.257666      52.1450607
#> 45        e        24.002834     -82.3151845
#> 46        f      -102.674688     -86.0515913
#> 47        g        63.647894      84.9743482
#> 48        h       -62.496739     -58.9462027
#> 49        i       163.938775      43.0506527
#> 50        j        25.940025     -69.4136181
#> 51        a       160.755698      75.8406466
#> 52        b      -147.883204       0.8914734
#> 53        c       -50.295837      83.6451317
#> 54        d      -153.386383     -11.1363562
#> 55        e        56.500440      76.6501037
#> 56        f        74.468285      24.5454070
#> 57        g        84.995797      55.1203331
#> 58        h        64.053835     -86.9080277
#> 59        i       -72.306905      32.7804552
#> 60        j        33.216004      16.0447809
#> 61        a        99.676877     -11.9781512
#> 62        b        -7.816572      70.3088265
#> 63        c        40.213235      55.4336214
#> 64        d      -178.198620     -21.4495006
#> 65        e      -130.842231     -54.6516550
#> 66        f       138.808176     -78.0661729
#> 67        g       -94.940190     -59.2026727
#> 68        h       -81.827303      60.5313151
#> 69        i       -51.818101      65.7724698
#> 70        j       114.619116     -30.0970102
#> 71        a       -50.794901     -31.6614483
#> 72        b      -125.659019      67.1834955
#> 73        c        91.514835      19.4516079
#> 74        d        80.775744      39.1385613
#> 75        e      -156.095196      76.8254517
#> 76        f       -75.641457       4.6070561
#> 77        g      -164.883284     -25.0781467
#> 78        h       135.463637      42.6733278
#> 79        i        89.622154      15.3989272
#> 80        j       -34.284846     -17.3592091
#> 81        a       -99.991397      16.5692856
#> 82        b       -90.145713      73.3136712
#> 83        c        48.159614     -58.2705791
#> 84        d       -18.996044      20.7439732
#> 85        e      -150.395767      42.2097644
#> 86        f       -76.739443     -87.6545590
#> 87        g       -15.179165     -38.7465618
#> 88        h        96.907364      86.7681328
#> 89        i        64.241400     -70.2209677
#> 90        j        23.504622      23.9538758
#> 91        a      -173.878542      74.0348338
#> 92        b        -2.928541      88.0003999
#> 93        c      -127.798474     -89.2410863
#> 94        d       144.476262      48.0680033
#> 95        e       -30.888721      62.7517692
#> 96        f       150.010810      65.9754559
#> 97        g       -60.455123      55.2294781
#> 98        h       -50.153336      -2.9497669
#> 99        i      -110.317020     -77.6363034
#> 100       j       -47.920000     -15.7800000