Identify Coordinates in Artificial Hotspot Occurrence Inventory

Removes or flags records within Artificial Hotspot Occurrence Inventory. Poorly geo-referenced occurrence records in biological databases are often erroneously geo-referenced to highly recurring coordinates that were assessed by Park et al 2022. See the reference for more details.

cc_aohi(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  species = "species",
  taxa = c("Aves", "Insecta", "Mammalia", "Plantae"),
  buffer = 10000,
  geod = TRUE,
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

species

character string. The column with the species identity. Only required if verify = TRUE.

taxa

Artificial Hotspot Occurrence Inventory (AHOI) were created based on four different taxa, birds, insecta, mammalia, and plantae. Users can choose to keep all, or any specific taxa subset to define the AHOI locations. Default is to keep all: c("Aves", "Insecta", "Mammalia", "Plantae").

buffer

The buffer around each capital coordinate (the centre of the city), where records should be flagged as problematic. Units depend on geod. Default = 10 kilometres.

geod

logical. If TRUE the radius around each capital is calculated based on a sphere, buffer is in meters and independent of latitude. If FALSE the radius is calculated assuming planar coordinates and varies slightly with latitude. Default = TRUE. See https://seethedatablog.wordpress.com/ for detail and credits.

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

References

Park, D. S., Xie, Y., Thammavong, H. T., Tulaiha, R., & Feng, X. (2023). Artificial Hotspot Occurrence Inventory (AHOI). Journal of Biogeography, 50, 441–449. doi:10.1111/jbi.14543

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = c(runif(99, -180, 180), -47.92), 
                decimalLatitude = c(runif(99, -90,90), -15.78))
cc_aohi(x)
#> Testing Artificial Hotspot Occurrence Inventory
#> Removed 0 records.
#>     species decimalLongitude decimalLatitude
#> 1         a      -124.067562        5.654371
#> 2         b       -31.310469      -66.815282
#> 3         c       -20.743404       -1.428842
#> 4         d        63.765131       52.738149
#> 5         e       -73.448248       18.483682
#> 6         f        18.582992      -52.517673
#> 7         g        44.935214      -46.935303
#> 8         h        48.075296      -78.176794
#> 9         i       -32.761081      -15.234237
#> 10        j       -55.873661       11.895320
#> 11        a       -78.867160       78.707731
#> 12        b       -21.175747       55.164935
#> 13        c      -110.968931        7.122720
#> 14        d       -70.595229      -45.434356
#> 15        e       158.466455      -87.227864
#> 16        f        40.131223       68.607014
#> 17        g      -152.739480      -18.902375
#> 18        h       -20.750828       43.901598
#> 19        i       -75.762128      -36.299889
#> 20        j       -70.255178      -19.106995
#> 21        a        53.482575      -63.556637
#> 22        b       166.431985       43.049891
#> 23        c       -74.329110       15.179101
#> 24        d       150.064641      -34.700659
#> 25        e      -109.916481      -67.879593
#> 26        f       -54.536800       22.409022
#> 27        g       125.958804       36.783823
#> 28        h         6.746698       66.319440
#> 29        i       -47.516703      -66.066921
#> 30        j       158.097421      -13.198322
#> 31        a      -172.305421       40.731879
#> 32        b       -29.465728      -51.793730
#> 33        c        60.435883       58.378195
#> 34        d       143.420767      -15.302251
#> 35        e        70.493783       32.285657
#> 36        f       -33.114283       57.115429
#> 37        g        64.096572      -80.828497
#> 38        h       -85.272569      -59.001160
#> 39        i        65.384082       62.819772
#> 40        j       123.831139       19.746460
#> 41        a        38.951416      -59.603679
#> 42        b        -9.011294      -40.205239
#> 43        c        51.637793        2.775327
#> 44        d       134.760135       75.333482
#> 45        e      -173.443028       41.643507
#> 46        f       141.140112       63.051536
#> 47        g      -140.553042       40.472100
#> 48        h       119.251415      -25.942841
#> 49        i       102.119283      -42.754139
#> 50        j       -99.068345       32.708669
#> 51        a       138.506340      -26.438176
#> 52        b      -138.711720       72.849442
#> 53        c       -52.635804       32.067155
#> 54        d       -10.573707       10.483686
#> 55        e        10.914444       47.302845
#> 56        f       -57.944926      -10.804608
#> 57        g       -75.321570      -24.381269
#> 58        h       -88.419947       67.109722
#> 59        i       158.646469       -7.821537
#> 60        j       169.512730      -14.851402
#> 61        a        24.208098       66.071028
#> 62        b       155.943824       18.280957
#> 63        c         2.854003       71.286583
#> 64        d       158.748505      -55.963064
#> 65        e      -114.297643       55.835923
#> 66        f       -72.169955      -66.720718
#> 67        g       157.578411      -32.790674
#> 68        h      -155.555986      -33.032861
#> 69        i        82.867007      -14.475624
#> 70        j       159.455864      -44.361470
#> 71        a        52.267624      -67.399127
#> 72        b        80.391386      -42.745828
#> 73        c       -52.666503       76.367730
#> 74        d       150.505544       -7.938652
#> 75        e      -107.126832       28.886491
#> 76        f        65.094897       73.679514
#> 77        g         2.185947      -37.670100
#> 78        h      -104.312518        3.547749
#> 79        i       -75.085005       89.267058
#> 80        j      -163.125875       38.212046
#> 81        a        95.351736      -62.349931
#> 82        b      -130.774470       17.739587
#> 83        c      -103.565901       23.064021
#> 84        d        28.179431       18.778976
#> 85        e        79.949201       38.045709
#> 86        f       165.629971      -52.344343
#> 87        g       -58.137529        5.723581
#> 88        h       -44.111899      -36.717383
#> 89        i       116.108202      -45.326634
#> 90        j      -155.696728       12.484339
#> 91        a       149.973753      -20.172877
#> 92        b        27.203575       75.585953
#> 93        c        82.523339       54.655664
#> 94        d        53.786185       32.182805
#> 95        e      -121.790715      -13.270228
#> 96        f        88.344864      -50.349567
#> 97        g      -178.183719      -10.239761
#> 98        h        27.596492      -38.584708
#> 99        i        66.372109       34.086879
#> 100       j       -47.920000      -15.780000