Identify Coordinates in Vicinity of Country Capitals.

Removes or flags records within a certain radius around country capitals. Poorly geo-referenced occurrence records in biological databases are often erroneously geo-referenced to capitals.

cc_cap(
  x,
  lon = "decimallongitude",
  lat = "decimallatitude",
  species = "species",
  buffer = 10000,
  geod = TRUE,
  ref = NULL,
  verify = FALSE,
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimallongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimallatitude”.

species

character string. The column with the species identity. Only required if verify = TRUE.

buffer

The buffer around each capital coordinate (the centre of the city), where records should be flagged as problematic. Units depend on geod. Default = 10 kilometres.

geod

logical. If TRUE the radius around each capital is calculated based on a sphere, buffer is in meters and independent of latitude. If FALSE the radius is calculated assuming planar coordinates and varies slightly with latitude, in this case buffer is in degrees. Default = TRUE. See https://seethedatablog.wordpress.com/ for detail and credits.

ref

SpatialPointsDataFrame. Providing the geographic gazetteer. Can be any SpatialPointsDataFrame, but the structure must be identical to countryref. Default = countryref.

verify

logical. If TRUE records are only flagged if they are the only record in a given species flagged close to a given reference. If FALSE, the distance is the only criterion

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimallongitude = runif(100, -180, 180), 
                decimallatitude = runif(100, -90,90))

cc_cap(x)
#> Testing country capitals
#> Removed 0 records.
#>     species decimallongitude decimallatitude
#> 1         a       -41.939572       62.359088
#> 2         b        21.951665       42.377644
#> 3         c        54.398724       -5.838524
#> 4         d      -178.573584       23.870413
#> 5         e      -178.842711       37.182852
#> 6         f        72.176482      -83.515199
#> 7         g        55.646191      -10.000901
#> 8         h        70.656686       87.881174
#> 9         i       139.927442       22.023402
#> 10        j        -3.014722       76.971120
#> 11        a         3.531642      -24.592543
#> 12        b      -124.786184       79.289820
#> 13        c      -126.363795      -68.019948
#> 14        d      -176.701679       51.041824
#> 15        e        71.211253       39.358117
#> 16        f       135.047347       39.942513
#> 17        g        18.311152       53.388482
#> 18        h      -169.401642        6.583027
#> 19        i       -35.669819       20.081322
#> 20        j       167.557009       36.412058
#> 21        a       110.935345      -50.556764
#> 22        b       -16.347055       27.735854
#> 23        c       112.238861       67.913951
#> 24        d       -48.351879      -63.978040
#> 25        e       127.808902      -86.398343
#> 26        f       -99.756953      -79.535709
#> 27        g       -60.265671       40.594013
#> 28        h        72.126797      -83.021517
#> 29        i       -63.720974      -67.079720
#> 30        j        18.499839       21.732971
#> 31        a      -161.035692      -42.918113
#> 32        b       -44.872972      -86.452698
#> 33        c        29.175446       31.612243
#> 34        d      -161.115136       69.128964
#> 35        e       104.035082      -85.682825
#> 36        f       112.669017       64.153517
#> 37        g       -36.029042       71.415539
#> 38        h       -94.470912      -82.849434
#> 39        i       -67.382859        8.134903
#> 40        j       -78.291886      -88.574437
#> 41        a        87.776671      -35.896093
#> 42        b       106.524201       43.228703
#> 43        c        76.473031       14.700266
#> 44        d      -160.410858       84.920355
#> 45        e      -111.003093      -49.757121
#> 46        f        59.809794        5.614472
#> 47        g      -115.651398       21.793456
#> 48        h       -77.012248      -63.542094
#> 49        i       144.648517      -88.108032
#> 50        j      -150.855492      -13.854989
#> 51        a      -173.396856        6.209216
#> 52        b      -140.809156      -85.681558
#> 53        c       119.433565       76.900936
#> 54        d      -111.422064      -54.662062
#> 55        e        57.300102       83.545898
#> 56        f        28.340550       72.983021
#> 57        g       -21.715208      -56.509688
#> 58        h       -23.485541       87.273911
#> 59        i       115.034342       60.716197
#> 60        j      -174.859847       61.568991
#> 61        a       -91.362230      -79.973906
#> 62        b       -18.012839      -42.227469
#> 63        c         5.440202       25.273663
#> 64        d        84.974508       -5.453833
#> 65        e       -78.611705      -51.598491
#> 66        f      -169.667688       61.870822
#> 67        g        85.777243      -64.184333
#> 68        h      -131.106901       16.858644
#> 69        i       -69.991999       -2.831785
#> 70        j       114.320407       85.835314
#> 71        a       139.952661      -34.533155
#> 72        b       138.423370       68.047761
#> 73        c       176.441050      -44.450850
#> 74        d        19.195300      -82.536398
#> 75        e       -20.055698      -50.179712
#> 76        f      -121.809028      -14.752030
#> 77        g       179.493507       14.806384
#> 78        h       -88.292221       10.692588
#> 79        i        -6.999949       87.280376
#> 80        j       -46.707561      -87.786450
#> 81        a      -106.320370       79.667751
#> 82        b       -69.381454       14.039474
#> 83        c       -17.142216       87.017501
#> 84        d      -131.489731      -87.244683
#> 85        e       -22.704125       57.166046
#> 86        f       -74.443527       24.590785
#> 87        g        35.937953      -61.919868
#> 88        h       113.983184       83.971792
#> 89        i        27.606134      -18.372260
#> 90        j       -51.478146       74.411310
#> 91        a      -106.543995       62.016189
#> 92        b       -42.454784       -8.428213
#> 93        c        20.404256      -56.103238
#> 94        d       -53.331617      -88.624471
#> 95        e      -170.473142       33.087618
#> 96        f       -96.888126       42.273861
#> 97        g      -100.901470       89.102425
#> 98        h        -2.184551       76.074847
#> 99        i       -54.407010       25.342549
#> 100       j        14.710079      -51.305670
cc_cap(x, value = "flagged")
#> Testing country capitals
#> Flagged 0 records.
#>    1    2    3    4    5    6    7    8    9   10   11   12   13   14   15   16 
#> TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE 
#>   17   18   19   20   21   22   23   24   25   26   27   28   29   30   31   32 
#> TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE 
#>   33   34   35   36   37   38   39   40   41   42   43   44   45   46   47   48 
#> TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE 
#>   49   50   51   52   53   54   55   56   57   58   59   60   61   62   63   64 
#> TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE 
#>   65   66   67   68   69   70   71   72   73   74   75   76   77   78   79   80 
#> TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE 
#>   81   82   83   84   85   86   87   88   89   90   91   92   93   94   95   96 
#> TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE 
#>   97   98   99  100 
#> TRUE TRUE TRUE TRUE