Identify Coordinates in Vicinity of Country Capitals.

Removes or flags records within a certain radius around country capitals. Poorly geo-referenced occurrence records in biological databases are often erroneously geo-referenced to capitals.

cc_cap(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  species = "species",
  buffer = 10000,
  geod = TRUE,
  ref = NULL,
  verify = FALSE,
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

species

character string. The column with the species identity. Only required if verify = TRUE.

buffer

The buffer around each capital coordinate (the centre of the city), where records should be flagged as problematic. Units depend on geod. Default = 10 kilometres.

geod

logical. If TRUE the radius around each capital is calculated based on a sphere, buffer is in meters and independent of latitude. If FALSE the radius is calculated assuming planar coordinates and varies slightly with latitude. Default = TRUE. See https://seethedatablog.wordpress.com/ for detail and credits.

ref

SpatVector (geometry: polygons). Providing the geographic gazetteer. Can be any SpatVector (geometry: polygons), but the structure must be identical to countryref. Default = countryref.

verify

logical. If TRUE records are only flagged if they are the only record in a given species flagged close to a given reference. If FALSE, the distance is the only criterion

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

See also

Examples

x <- data.frame(species = letters[1:10],
                decimalLongitude = c(runif(99, -180, 180), -47.882778),
                decimalLatitude = c(runif(99, -90, 90), -15.793889))

cc_cap(x)
#> Testing country capitals
#> Removed 1 records.
#>    species decimalLongitude decimalLatitude
#> 1        a       81.5247704      41.7323865
#> 2        b       38.7255181     -57.8277744
#> 3        c      160.5282250      -0.9970314
#> 4        d      -30.1481970      30.5180681
#> 5        e       10.2471123     -85.7245381
#> 6        f     -161.4308235      51.6653202
#> 7        g       37.5240207      -2.6567573
#> 8        h        6.8511781     -58.6618783
#> 9        i      145.2848039     -10.1810884
#> 10       j      174.7444472      74.7448000
#> 11       a       98.3589772     -85.6574253
#> 12       b      -43.5047870     -54.0091391
#> 13       c       37.8557099     -42.8721593
#> 14       d      156.8602022      63.3277240
#> 15       e      -22.3861550     -55.1587213
#> 16       f      -33.9813611     -59.1748766
#> 17       g      -96.5852130     -82.1605012
#> 18       h      119.8641348     -76.4435037
#> 19       i       15.0957701     -57.0970963
#> 20       j       -0.9678232     -52.5908156
#> 21       a       28.7767751      -0.2311357
#> 22       b       74.0239551     -67.8370381
#> 23       c       29.9063695      42.4868718
#> 24       d      -65.5736050      70.4138891
#> 25       e     -106.6476073     -82.3854272
#> 26       f     -113.7518543      81.1910751
#> 27       g      -88.0042062     -40.1825914
#> 28       h      136.0551668      23.0980683
#> 29       i      -77.5390313      17.2703665
#> 30       j      -40.3225972     -20.6335347
#> 31       a     -146.2063635      50.7748589
#> 32       b      -16.1996097      85.2424881
#> 33       c       97.8030283     -43.5292853
#> 34       d      -28.4014861      12.1384115
#> 35       e      165.0571006       2.9472348
#> 36       f     -174.3508969     -49.8341593
#> 37       g       -5.7593477      73.6183297
#> 38       h      176.7725170      66.3655820
#> 39       i      -46.7129342     -75.0510551
#> 40       j     -159.9161615      36.4207194
#> 41       a      -14.6965648      16.6009816
#> 42       b      -30.4503952     -58.8498113
#> 43       c       36.9053636      33.7465461
#> 44       d      125.3509878     -10.4301728
#> 45       e     -155.8588769      48.9649191
#> 46       f        9.7712237     -41.1445042
#> 47       g        8.6602517     -48.6984440
#> 48       h      -15.0702864      18.6622092
#> 49       i      118.4401656      12.3810661
#> 50       j      -48.9504869     -80.3138460
#> 51       a     -147.9848860     -77.5037615
#> 52       b       -4.8887914     -16.5563048
#> 53       c       11.2440127      -7.4428996
#> 54       d       36.0622967      21.6620054
#> 55       e       77.9430850      50.3386944
#> 56       f     -100.1505323      50.1834901
#> 57       g     -140.1604976      -7.9390430
#> 58       h      102.6676837     -60.1613919
#> 59       i     -174.4882028      19.6191447
#> 60       j      120.4548603      89.5623330
#> 61       a     -175.1737021     -62.8437581
#> 62       b       38.7406841     -82.2330097
#> 63       c      -65.0106726      65.8941563
#> 64       d       31.4201257      76.5263354
#> 65       e       13.0134414     -46.8559399
#> 66       f      114.8424425      78.8743139
#> 67       g       94.1705138      89.0217749
#> 68       h       52.1839333     -60.3214408
#> 69       i      -31.5329086       3.6315766
#> 70       j     -160.6524838      62.6018201
#> 71       a      -52.8052896     -65.6457054
#> 72       b      -87.8806280     -46.3256808
#> 73       c      144.9952874      34.3582977
#> 74       d      165.4218022      45.5977784
#> 75       e       50.7756624       8.4116731
#> 76       f     -166.9289716     -15.2830291
#> 77       g      -60.8745402      57.4072359
#> 78       h      123.6095070      37.2618096
#> 79       i      -32.6584762     -55.8685846
#> 80       j     -138.3695894      55.9883002
#> 81       a     -177.1328469      52.7760359
#> 82       b      -52.3633198     -53.6375326
#> 83       c       23.6658863      43.4568412
#> 84       d     -156.9655215     -41.4727177
#> 85       e     -178.3428313      76.5622410
#> 86       f       55.8691948      49.4950305
#> 87       g       88.9998834     -55.8594228
#> 88       h      -45.8737188     -68.8155983
#> 89       i     -104.0506138     -27.6943471
#> 90       j     -102.8248651      76.2903495
#> 91       a     -174.7796162     -86.2270353
#> 92       b      178.8555837      84.4558112
#> 93       c       23.9792356       3.0805948
#> 94       d       90.1079354      79.1683298
#> 95       e      104.9448339      61.9667377
#> 96       f     -140.8631702      22.9960644
#> 97       g      131.1347631     -44.3792023
#> 98       h      -52.7108638      78.5704716
#> 99       i      161.3818682     -41.7832210
cc_cap(x, value = "flagged")
#> Testing country capitals
#> Flagged 1 records.
#>   [1]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [13]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [25]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [37]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [49]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [61]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [73]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [85]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [97]  TRUE  TRUE  TRUE FALSE