Identify Coordinates in Vicinity of Country Capitals.

Removes or flags records within a certain radius around country capitals. Poorly geo-referenced occurrence records in biological databases are often erroneously geo-referenced to capitals.

cc_cap(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  species = "species",
  buffer = 10000,
  geod = TRUE,
  ref = NULL,
  verify = FALSE,
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

species

character string. The column with the species identity. Only required if verify = TRUE.

buffer

The buffer around each capital coordinate (the centre of the city), where records should be flagged as problematic. Units depend on geod. Default = 10 kilometres.

geod

logical. If TRUE the radius around each capital is calculated based on a sphere, buffer is in meters and independent of latitude. If FALSE the radius is calculated assuming planar coordinates and varies slightly with latitude. Default = TRUE. See https://seethedatablog.wordpress.com/ for detail and credits.

ref

SpatVector (geometry: polygons). Providing the geographic gazetteer. Can be any SpatVector (geometry: polygons), but the structure must be identical to countryref. Default = countryref.

verify

logical. If TRUE records are only flagged if they are the only record in a given species flagged close to a given reference. If FALSE, the distance is the only criterion

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

See also

Examples

x <- data.frame(species = letters[1:10],
                decimalLongitude = c(runif(99, -180, 180), -47.882778),
                decimalLatitude = c(runif(99, -90, 90), -15.793889))

cc_cap(x)
#> Testing country capitals
#> Removed 1 records.
#>    species decimalLongitude decimalLatitude
#> 1        a       104.122074    -73.37887678
#> 2        b      -155.649257     31.52842950
#> 3        c      -158.310083    -51.43270496
#> 4        d       149.754955    -32.81806359
#> 5        e      -163.095710     63.73619234
#> 6        f        53.705093    -21.68571055
#> 7        g        81.691581    -31.24992618
#> 8        h      -166.233161    -11.69988411
#> 9        i        81.662445     48.74249525
#> 10       j        46.076395     71.42801309
#> 11       a       -42.058001    -58.46599821
#> 12       b        38.527702     75.19854966
#> 13       c       -88.498581    -37.20097773
#> 14       d        52.946536     64.75090394
#> 15       e      -137.495220    -53.71832297
#> 16       f       124.089705     32.42633400
#> 17       g       112.969490     -9.73265395
#> 18       h      -152.168240     -1.46898067
#> 19       i      -148.089160     52.08791619
#> 20       j      -168.740603      3.31523804
#> 21       a       -57.339224    -11.62981291
#> 22       b        80.992549    -24.30043272
#> 23       c      -147.848281     26.87616269
#> 24       d      -174.566030     78.44223915
#> 25       e       -94.424116    -23.16799364
#> 26       f      -168.490104     -0.54292629
#> 27       g       123.686849    -21.10276059
#> 28       h       145.571834     -1.14390293
#> 29       i       -86.579627     52.46423691
#> 30       j       104.088885     28.77232202
#> 31       a       131.805532     -5.96260508
#> 32       b      -128.607125    -28.55030150
#> 33       c       163.904194    -23.39942578
#> 34       d      -140.311367    -87.57262263
#> 35       e       138.852719    -77.48005559
#> 36       f       175.516256     -5.23309406
#> 37       g       -27.338759    -60.32766673
#> 38       h        61.307312     72.86611780
#> 39       i       142.790940    -24.08316277
#> 40       j        63.014490      0.05965434
#> 41       a       -56.341491    -33.87413498
#> 42       b       169.406298    -34.77000562
#> 43       c       -24.219366    -75.53614359
#> 44       d        27.773263    -24.35324047
#> 45       e      -130.418092     54.58834775
#> 46       f        41.295846    -58.09625432
#> 47       g       -33.076794     26.90508687
#> 48       h       138.915859     52.35594100
#> 49       i       -33.663699    -62.94370544
#> 50       j       176.353156      0.61506519
#> 51       a       -96.799768     44.95382879
#> 52       b        30.956841    -39.80246068
#> 53       c      -119.412991     77.20719757
#> 54       d       152.382219      3.59500959
#> 55       e      -125.471088     88.89188441
#> 56       f      -116.015673     38.23536811
#> 57       g       -61.220178    -43.67550247
#> 58       h       116.432744     85.08736651
#> 59       i        99.951004     59.26661246
#> 60       j      -118.979022     18.06797478
#> 61       a        -7.468987    -51.67401102
#> 62       b       -25.111837     49.55950844
#> 63       c       172.459759     74.16741069
#> 64       d       -52.732565      1.42750454
#> 65       e        65.074301    -75.74537886
#> 66       f       116.900739     44.96941678
#> 67       g      -131.656835    -68.80636475
#> 68       h       -20.307490     68.22827733
#> 69       i        46.423825     76.07245866
#> 70       j       154.178242     31.77963610
#> 71       a       -53.127601    -73.57954688
#> 72       b        35.394917     81.59897243
#> 73       c       152.378490     42.07026785
#> 74       d      -107.851607    -24.53090685
#> 75       e      -155.507551    -61.26688567
#> 76       f       140.216420    -26.15536373
#> 77       g       -51.851614    -57.04602834
#> 78       h       171.342310    -60.18798240
#> 79       i         3.861790     -0.13143662
#> 80       j      -168.218740     79.22251072
#> 81       a       -73.874440    -32.18645022
#> 82       b        81.004526      1.04316163
#> 83       c       125.651871     10.05138353
#> 84       d        26.985060    -52.91194734
#> 85       e      -146.662339     15.36322515
#> 86       f        54.082825     58.97488045
#> 87       g        13.161319     -1.05721238
#> 88       h       165.802318     35.60793764
#> 89       i       -15.827872     77.49480254
#> 90       j        71.506305     66.93425242
#> 91       a       178.964248    -45.79609602
#> 92       b        79.610180    -41.10190520
#> 93       c      -105.739120     84.74853197
#> 94       d        12.705904     77.54469259
#> 95       e       173.436638     -3.00059380
#> 96       f      -140.681001     16.36361661
#> 97       g        69.383329    -31.36466509
#> 98       h         1.277710     75.27671048
#> 99       i       103.983228     48.70590179
cc_cap(x, value = "flagged")
#> Testing country capitals
#> Flagged 1 records.
#>   [1]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [13]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [25]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [37]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [49]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [61]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [73]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [85]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [97]  TRUE  TRUE  TRUE FALSE