Identify Coordinates in Vicinity of Country Capitals.

Removes or flags records within a certain radius around country capitals. Poorly geo-referenced occurrence records in biological databases are often erroneously geo-referenced to capitals.

cc_cap(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  species = "species",
  buffer = 10000,
  geod = TRUE,
  ref = NULL,
  verify = FALSE,
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

species

character string. The column with the species identity. Only required if verify = TRUE.

buffer

The buffer around each capital coordinate (the centre of the city), where records should be flagged as problematic. Units depend on geod. Default = 10 kilometres.

geod

logical. If TRUE the radius around each capital is calculated based on a sphere, buffer is in meters and independent of latitude. If FALSE the radius is calculated assuming planar coordinates and varies slightly with latitude. Default = TRUE. See https://seethedatablog.wordpress.com/ for detail and credits.

ref

SpatVector (geometry: polygons). Providing the geographic gazetteer. Can be any SpatVector (geometry: polygons), but the structure must be identical to countryref. Default = countryref.

verify

logical. If TRUE records are only flagged if they are the only record in a given species flagged close to a given reference. If FALSE, the distance is the only criterion

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

See also

Examples

x <- data.frame(species = letters[1:10],
                decimalLongitude = c(runif(99, -180, 180), -47.882778),
                decimalLatitude = c(runif(99, -90, 90), -15.793889))

cc_cap(x)
#> Testing country capitals
#> Removed 1 records.
#>    species decimalLongitude decimalLatitude
#> 1        a      -125.065174    -47.24359303
#> 2        b      -157.922955     45.77578001
#> 3        c       121.481343     28.18202552
#> 4        d        73.835205    -31.62725255
#> 5        e       112.267278     88.30095080
#> 6        f       -79.795053     35.31236684
#> 7        g        15.552557     82.22687295
#> 8        h      -111.324815     10.96258379
#> 9        i      -134.940285     45.57833753
#> 10       j       148.450419    -65.53245502
#> 11       a       157.560736     16.39758084
#> 12       b        75.112272    -53.79881225
#> 13       c      -151.173391     44.36225917
#> 14       d       100.486429    -87.33017600
#> 15       e       -47.086769     45.00154780
#> 16       f      -132.160276    -79.39848773
#> 17       g        49.524620    -82.41011107
#> 18       h       -51.500179    -27.22285907
#> 19       i       -18.673171     15.00717460
#> 20       j       -69.249593    -60.14168696
#> 21       a        94.837149     12.95349655
#> 22       b        47.786037    -56.20505477
#> 23       c        28.087465    -29.12279055
#> 24       d        95.243311    -68.51664389
#> 25       e       169.691993    -66.50932958
#> 26       f        70.211315     82.33644701
#> 27       g       148.738369    -22.97199985
#> 28       h       -38.489755    -53.12940479
#> 29       i       166.023079     87.02222262
#> 30       j        37.501914     10.15796756
#> 31       a       127.014423    -16.43859118
#> 32       b        38.551272    -12.38862744
#> 33       c      -136.763521     32.91822716
#> 34       d       -24.822434     88.24552509
#> 35       e        -9.760371    -66.16610745
#> 36       f      -101.086697    -20.22195420
#> 37       g       -35.443057    -37.95094428
#> 38       h       -17.124210     -5.44042869
#> 39       i         4.246097      2.98149853
#> 40       j       -31.335624     89.31714856
#> 41       a        39.370588     18.26433406
#> 42       b       -72.251801    -65.49650277
#> 43       c        -5.151897    -71.08199676
#> 44       d        98.116120    -38.09661803
#> 45       e       174.046617     11.46760028
#> 46       f      -165.351642     43.34876851
#> 47       g      -107.124156     -1.64638101
#> 48       h      -136.267618     14.22615180
#> 49       i       105.818151    -46.40345878
#> 50       j        65.485132    -68.82471139
#> 51       a       150.217829    -50.10604071
#> 52       b      -120.792336     38.34742859
#> 53       c        70.679579    -78.55114044
#> 54       d      -144.396844    -74.39092428
#> 55       e       103.836895    -23.74989373
#> 56       f       -76.249306     84.81389820
#> 57       g        90.989557    -52.42681141
#> 58       h      -173.968461     -9.90192866
#> 59       i        19.907295    -10.45543577
#> 60       j      -162.221932     84.47904821
#> 61       a       102.886175    -35.31481780
#> 62       b       172.430162    -62.86522945
#> 63       c       175.188242    -33.53438007
#> 64       d        84.791849    -65.06638281
#> 65       e       111.419708    -45.17432485
#> 66       f       -45.308533    -19.14472106
#> 67       g       -91.505605    -51.66929113
#> 68       h       -95.180973    -82.79764556
#> 69       i      -117.395517    -14.54997548
#> 70       j        21.031818     -2.60343526
#> 71       a        -3.739617    -59.84139489
#> 72       b      -174.384728     57.99176598
#> 73       c       -51.870760      7.66173692
#> 74       d       -86.075266    -10.88387378
#> 75       e       113.751560    -32.07548761
#> 76       f      -100.691796     -3.57294196
#> 77       g        95.300065      6.99916160
#> 78       h       101.219461     44.47632379
#> 79       i        47.748275    -14.61893516
#> 80       j       147.423369     43.64017120
#> 81       a      -160.355574    -54.32863318
#> 82       b        49.802210    -23.58065346
#> 83       c       171.383879    -31.19625836
#> 84       d       -78.290185    -25.16522858
#> 85       e       -44.177341    -31.75039646
#> 86       f       139.441088    -80.54914379
#> 87       g      -129.760841     23.49890665
#> 88       h       -76.059110     -0.08630804
#> 89       i      -175.935620    -57.14761621
#> 90       j       -44.407771    -27.15756597
#> 91       a       -12.649269     78.45015908
#> 92       b      -107.566300     53.48711408
#> 93       c       177.200751     80.91000530
#> 94       d      -166.003059    -43.02304847
#> 95       e      -173.678116    -29.63599487
#> 96       f       167.951354     85.25986155
#> 97       g        57.957418    -24.51460920
#> 98       h      -159.897287    -87.71863325
#> 99       i       169.172694     10.38958511
cc_cap(x, value = "flagged")
#> Testing country capitals
#> Flagged 1 records.
#>   [1]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [13]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [25]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [37]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [49]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [61]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [73]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [85]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [97]  TRUE  TRUE  TRUE FALSE