Identify Coordinates in Vicinity of Country and Province Centroids

Removes or flags records within a radius around the geographic centroids of political countries and provinces. Poorly geo-referenced occurrence records in biological databases are often erroneously geo-referenced to centroids.

cc_cen(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  species = "species",
  buffer = 1000,
  geod = TRUE,
  test = "both",
  ref = NULL,
  verify = FALSE,
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

species

character string. The column with the species identity. Only required if verify = TRUE.

buffer

numerical. The buffer around each province or country centroid, where records should be flagged as problematic. Units depend on geod. Default = 1 kilometre.

geod

logical. If TRUE the radius around each capital is calculated based on a sphere, buffer is in meters and independent of latitude. If FALSE the radius is calculated assuming planar coordinates and varies slightly with latitude. Default = TRUE. See https://seethedatablog.wordpress.com/ for detail and credits.

test

a character string. Specifying the details of the test. One of c(“both”, “country”, “provinces”). If both tests for country and province centroids.

ref

SpatVector (geometry: polygons). Providing the geographic gazetteer. Can be any SpatVector (geometry: polygons), but the structure must be identical to countryref. Default = countryref.

verify

logical. If TRUE records are only flagged if they are the only record in a given species flagged close to a given reference. If FALSE, the distance is the only criterion

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = c(runif(99, -180, 180), -47.92), 
                decimalLatitude = c(runif(99, -90,90), -15.78))
cc_cen(x, geod = FALSE)
#> Testing country centroids
#> Removed 0 records.
#>     species decimalLongitude decimalLatitude
#> 1         a       -24.540286     79.72702175
#> 2         b        82.508034     34.10992102
#> 3         c       -97.976705    -45.90471555
#> 4         d        68.205789    -66.40200358
#> 5         e        -3.817924      0.08306759
#> 6         f       127.392912    -57.53945095
#> 7         g        25.184732    -36.73280259
#> 8         h        19.377307     56.06828491
#> 9         i        90.560362     37.60082368
#> 10        j        85.515241    -64.67916068
#> 11        a        75.553597    -65.38246170
#> 12        b       169.318755    -36.59143777
#> 13        c      -111.226349    -76.59179932
#> 14        d        94.371032    -48.74526811
#> 15        e        13.766605     -3.55001507
#> 16        f       -91.794301    -55.01366664
#> 17        g       107.889132    -41.26403828
#> 18        h      -129.678058    -31.83712606
#> 19        i        81.988193      9.71552000
#> 20        j       -65.222899    -24.81625164
#> 21        a        11.088528    -75.59794261
#> 22        b      -106.980850     88.12445374
#> 23        c       122.140176    -56.61585805
#> 24        d       162.735221    -52.04470979
#> 25        e       -49.067580    -16.64256643
#> 26        f       125.710403    -56.82279613
#> 27        g        18.121502    -75.01007146
#> 28        h       -54.672605     41.50204967
#> 29        i       156.427640    -33.08538578
#> 30        j        -9.355723    -59.60229901
#> 31        a       -53.861160    -65.70265635
#> 32        b      -122.740720    -18.22197188
#> 33        c       -52.195652    -21.73356607
#> 34        d      -136.303920    -82.71649914
#> 35        e      -108.729664     -4.11257342
#> 36        f      -105.930325    -21.08554278
#> 37        g       169.859056     80.31470087
#> 38        h       171.194738    -42.54221078
#> 39        i        81.997513     49.11181917
#> 40        j       -11.201278    -41.11040797
#> 41        a       -76.354321     72.18550826
#> 42        b       167.746232     13.45268323
#> 43        c      -132.868262      1.47485135
#> 44        d        99.312831     29.65893673
#> 45        e        10.082179     11.58204323
#> 46        f      -169.282135     -5.15239783
#> 47        g      -135.700206     63.06844387
#> 48        h       -55.317376      6.38693583
#> 49        i      -175.720350     65.72211133
#> 50        j       -74.708059     15.77245581
#> 51        a       -18.332378     59.94066148
#> 52        b        14.740490    -15.64712289
#> 53        c        70.945107      5.76504513
#> 54        d        94.172769    -55.67092453
#> 55        e       -13.479744    -17.10087645
#> 56        f      -145.115057     34.91283601
#> 57        g       104.553893    -44.44426251
#> 58        h       103.401306    -47.83144122
#> 59        i       116.539294     56.70324341
#> 60        j       -78.670372     86.87132078
#> 61        a        28.233330     57.81847972
#> 62        b        58.255022      1.14387170
#> 63        c        57.842243    -26.56139087
#> 64        d       146.230038     41.92303784
#> 65        e         9.838763     70.81790982
#> 66        f        59.244696    -18.48911227
#> 67        g        -5.973934    -51.94022580
#> 68        h         6.132551     23.35053896
#> 69        i        26.064783    -89.02213935
#> 70        j        99.635091     68.13980492
#> 71        a      -120.902892    -58.09326848
#> 72        b      -120.304209      9.38256662
#> 73        c      -135.120211      9.60338661
#> 74        d       -10.632431    -38.66008749
#> 75        e       175.412471     70.49989379
#> 76        f       101.537278     74.97523627
#> 77        g        64.335952    -37.52975088
#> 78        h        60.471846    -52.93046292
#> 79        i       -90.023211     33.45416727
#> 80        j       156.895785    -48.72711550
#> 81        a       -70.497863     50.16336299
#> 82        b       154.071980    -48.19870084
#> 83        c       -59.368147    -23.19966675
#> 84        d       110.554255     82.33565656
#> 85        e       169.660988    -40.27902505
#> 86        f        -9.159452    -33.13391122
#> 87        g       -93.097858    -44.19622796
#> 88        h        26.889083    -45.22920360
#> 89        i       -62.752915     49.30981686
#> 90        j        94.221484     85.74092165
#> 91        a        35.110458    -44.48017687
#> 92        b        14.660403     -0.68555867
#> 93        c       158.572962     18.79565913
#> 94        d       -65.018972     32.93915430
#> 95        e       -92.072570     40.48200814
#> 96        f        73.778722     89.25335151
#> 97        g      -101.796291    -83.45878649
#> 98        h       112.085025      1.36314564
#> 99        i       174.429168    -55.97494877
#> 100       j       -47.920000    -15.78000000

if (FALSE) {
cc_inst(x, value = "flagged", buffer = 50000) #geod = T
}