Identify Coordinates in Vicinity of Country and Province Centroids

Removes or flags records within a radius around the geographic centroids of political countries and provinces. Poorly geo-referenced occurrence records in biological databases are often erroneously geo-referenced to centroids.

cc_cen(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  species = "species",
  buffer = 1000,
  geod = TRUE,
  test = "both",
  ref = NULL,
  verify = FALSE,
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

species

character string. The column with the species identity. Only required if verify = TRUE.

buffer

numerical. The buffer around each province or country centroid, where records should be flagged as problematic. Units depend on geod. Default = 1 kilometre.

geod

logical. If TRUE the radius around each capital is calculated based on a sphere, buffer is in meters and independent of latitude. If FALSE the radius is calculated assuming planar coordinates and varies slightly with latitude. Default = TRUE. See https://seethedatablog.wordpress.com/ for detail and credits.

test

a character string. Specifying the details of the test. One of c(“both”, “country”, “provinces”). If both tests for country and province centroids.

ref

SpatVector (geometry: polygons). Providing the geographic gazetteer. Can be any SpatVector (geometry: polygons), but the structure must be identical to countryref. Default = countryref.

verify

logical. If TRUE records are only flagged if they are the only record in a given species flagged close to a given reference. If FALSE, the distance is the only criterion

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = c(runif(99, -180, 180), -47.92), 
                decimalLatitude = c(runif(99, -90,90), -15.78))
cc_cen(x, geod = FALSE)
#> Testing country centroids
#> Removed 0 records.
#>     species decimalLongitude decimalLatitude
#> 1         a     -156.2253578      67.5656034
#> 2         b      100.2219199     -61.3883208
#> 3         c     -139.5872548     -53.2739998
#> 4         d      131.5682685     -77.5326966
#> 5         e      133.7184654      84.3564881
#> 6         f        2.0339628       4.2050509
#> 7         g     -131.0608130      12.1148272
#> 8         h      110.2733829     -89.5235873
#> 9         i      130.8556330      19.5969670
#> 10        j     -161.7770181      43.6411926
#> 11        a      -97.4889763     -58.3718258
#> 12        b      133.0827860     -71.8686850
#> 13        c     -145.7779916      32.8786629
#> 14        d     -142.1942122      -3.1216160
#> 15        e       53.0956691     -44.2224155
#> 16        f      -29.3090011       5.5282144
#> 17        g       88.0100211      44.2723160
#> 18        h      -95.2613324     -26.5149462
#> 19        i     -136.1406243     -19.9761169
#> 20        j      -95.7295601     -31.7614787
#> 21        a       17.7274432     -77.5460268
#> 22        b      143.7660237     -30.9681065
#> 23        c      -43.4745624      -5.3277207
#> 24        d      -42.5695690     -59.8440540
#> 25        e      -19.2087952     -28.2565048
#> 26        f      132.5173554      80.9579968
#> 27        g       91.9163675     -16.6417347
#> 28        h      117.2473945       4.6669322
#> 29        i      124.2672218      64.5911097
#> 30        j       56.0356755     -11.6788662
#> 31        a       40.2054675     -83.6010443
#> 32        b      114.1820791      20.1169781
#> 33        c      -37.9616437      47.5246436
#> 34        d      -49.7542407     -70.0582395
#> 35        e       87.0735769     -88.1120770
#> 36        f     -150.9624630      23.8240776
#> 37        g       97.8409962      84.0621238
#> 38        h      -24.0781410      35.6464547
#> 39        i      141.5940899      21.0744480
#> 40        j      141.6350174      46.5152212
#> 41        a      -81.4619969      -3.5265373
#> 42        b       41.2947139     -16.3478434
#> 43        c       -2.8978013      40.1072418
#> 44        d       13.1439221      40.4889186
#> 45        e       -5.9799856     -62.2490258
#> 46        f       30.3335079     -37.2966037
#> 47        g       88.6728121     -77.6489598
#> 48        h      155.0991328      63.4229205
#> 49        i     -162.4738472      47.9560343
#> 50        j       -2.0364407      69.7448290
#> 51        a     -158.1816757      67.1032922
#> 52        b      -31.4073338      56.4725780
#> 53        c       42.5154991      57.8092897
#> 54        d       -1.4445543     -87.4313558
#> 55        e       26.5516553      84.7149715
#> 56        f       32.4234012     -49.3886199
#> 57        g     -135.5567683      27.5550289
#> 58        h       38.1270899      52.9632007
#> 59        i       32.1897718     -83.0612885
#> 60        j      -48.7931517      62.7107894
#> 61        a      -41.4444776     -18.0225656
#> 62        b      161.8347184     -63.2157764
#> 63        c       83.8711107      64.9874230
#> 64        d      137.0615815      25.8333961
#> 65        e       18.2996541      64.3491468
#> 66        f     -114.0778556      20.6379128
#> 67        g       39.6409267     -32.5991825
#> 68        h       48.4417254     -75.1369527
#> 69        i     -129.4480625     -79.7902207
#> 70        j       66.6871602      82.8581211
#> 71        a      103.2294565      15.3850430
#> 72        b       85.6069045      11.9627584
#> 73        c       30.7700837      18.8443618
#> 74        d       -9.7777049      65.9544710
#> 75        e      -10.5292591      19.3749859
#> 76        f     -103.5932106      47.2203661
#> 77        g      -87.7181132      38.1657763
#> 78        h       35.5606880     -83.8078946
#> 79        i     -125.6270167      23.6197916
#> 80        j       51.5454481     -89.0827384
#> 81        a       36.3257203     -27.5516875
#> 82        b       -7.5430542      29.6656127
#> 83        c      129.5002690      53.6609385
#> 84        d       -0.4016887      -3.2829177
#> 85        e      -52.4056241      -0.8746865
#> 86        f     -139.1478797      38.0163853
#> 87        g        4.8053470     -52.0111857
#> 88        h      -98.2701266      68.1697189
#> 89        i      -96.7006340     -57.4967666
#> 90        j     -110.1365485       7.3756379
#> 91        a      160.3267142     -72.4020838
#> 92        b      -96.6616826       4.7629690
#> 93        c      123.7785761     -76.7638467
#> 94        d        9.6691793      41.6971650
#> 95        e       65.4842609       3.5041722
#> 96        f       22.5884200     -37.2055500
#> 97        g       65.1991335      17.6035621
#> 98        h      108.6452112      -3.4018686
#> 99        i      -36.0920617      28.4506648
#> 100       j      -47.9200000     -15.7800000

if (FALSE) {
cc_inst(x, value = "flagged", buffer = 50000) #geod = T
}