Identify Coordinates in Vicinity of Country and Province Centroids

Removes or flags records within a radius around the geographic centroids of political countries and provinces. Poorly geo-referenced occurrence records in biological databases are often erroneously geo-referenced to centroids.

cc_cen(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  species = "species",
  buffer = 1000,
  geod = TRUE,
  test = "both",
  ref = NULL,
  verify = FALSE,
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

species

character string. The column with the species identity. Only required if verify = TRUE.

buffer

numerical. The buffer around each province or country centroid, where records should be flagged as problematic. Units depend on geod. Default = 1 kilometre.

geod

logical. If TRUE the radius around each capital is calculated based on a sphere, buffer is in meters and independent of latitude. If FALSE the radius is calculated assuming planar coordinates and varies slightly with latitude. Default = TRUE. See https://seethedatablog.wordpress.com/ for detail and credits.

test

a character string. Specifying the details of the test. One of c(“both”, “country”, “provinces”). If both tests for country and province centroids.

ref

SpatVector (geometry: polygons). Providing the geographic gazetteer. Can be any SpatVector (geometry: polygons), but the structure must be identical to countryref. Default = countryref.

verify

logical. If TRUE records are only flagged if they are the only record in a given species flagged close to a given reference. If FALSE, the distance is the only criterion

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = c(runif(99, -180, 180), -47.92), 
                decimalLatitude = c(runif(99, -90,90), -15.78))
cc_cen(x, geod = FALSE)
#> Testing country centroids
#> Removed 0 records.
#>     species decimalLongitude decimalLatitude
#> 1         a       43.3251221      -85.327410
#> 2         b      -67.4142333      -54.586907
#> 3         c      132.2850138      -79.129179
#> 4         d      -12.0660204      -81.216970
#> 5         e      120.9253817       15.287198
#> 6         f     -176.2834117      -13.219126
#> 7         g       13.9283708       28.125786
#> 8         h     -143.5193474      -84.535738
#> 9         i      -37.3262268       86.074648
#> 10        j      159.0455479      -75.599221
#> 11        a     -123.1425760       75.761615
#> 12        b       60.8441720        2.238267
#> 13        c     -139.9882715      -27.072149
#> 14        d      -66.1957583      -44.119644
#> 15        e     -148.6221161       73.563668
#> 16        f       73.8197011      -53.210553
#> 17        g      103.7723201      -68.336631
#> 18        h       21.1804377       88.331687
#> 19        i      -52.9770649      -61.104907
#> 20        j      -32.5337450        4.931584
#> 21        a     -147.6511232      -49.749880
#> 22        b      -79.9051571      -55.669764
#> 23        c     -156.5449309       35.924276
#> 24        d      -85.3439927       40.654116
#> 25        e     -177.1949538       -2.742110
#> 26        f       96.9814179      -32.729260
#> 27        g      -17.6045893       27.133710
#> 28        h     -163.7141678       84.789843
#> 29        i     -148.4344342      -24.217575
#> 30        j       24.0866759      -16.003327
#> 31        a     -166.6066253      -56.237489
#> 32        b      -22.3687291       40.165018
#> 33        c      176.4899066       41.107912
#> 34        d     -137.7057198       34.008571
#> 35        e       39.3556132      -28.504212
#> 36        f       82.5292282       81.065308
#> 37        g      -54.1097140       34.950400
#> 38        h     -161.4973686       26.691468
#> 39        i     -146.3297220      -56.913106
#> 40        j     -128.8614936       28.946313
#> 41        a       -6.3079833      -25.236747
#> 42        b      137.9119530       -4.881650
#> 43        c     -155.9972245       88.077998
#> 44        d     -108.5975999      -39.346108
#> 45        e       95.6206151       63.458285
#> 46        f      122.8869143       35.924674
#> 47        g       57.6265194      -69.102234
#> 48        h       89.9802697       79.647093
#> 49        i      -14.7292088      -29.376726
#> 50        j      160.5636143       48.919001
#> 51        a     -148.8085086      -84.426442
#> 52        b      172.4930524       53.165329
#> 53        c      -86.8485534       54.245783
#> 54        d       27.1939758      -84.543753
#> 55        e     -131.1668702       85.443547
#> 56        f     -108.0451616      -61.134150
#> 57        g      -34.5436940      -79.070861
#> 58        h     -124.2437022       29.275685
#> 59        i       -0.3891275      -81.570918
#> 60        j     -133.6087700       85.332142
#> 61        a        6.2261748       -2.623356
#> 62        b      -56.0091216       30.488411
#> 63        c       22.9351591      -66.784413
#> 64        d      -93.3452249      -12.307710
#> 65        e       87.1568509       64.697737
#> 66        f       43.8927143       31.338636
#> 67        g       66.2226617      -41.950107
#> 68        h      -71.0166366      -30.434042
#> 69        i      -57.5856741      -47.494030
#> 70        j       81.3749727       19.539839
#> 71        a      139.1742939       60.592134
#> 72        b     -152.9286196        3.321249
#> 73        c      176.3897524      -10.505615
#> 74        d     -146.2635110       31.942146
#> 75        e      -31.0066326       88.953948
#> 76        f     -142.6848053      -49.650353
#> 77        g       41.0088820      -68.661017
#> 78        h     -133.1438924      -80.459701
#> 79        i       -3.6465412      -38.132002
#> 80        j      -42.7456565       -8.937189
#> 81        a       63.9899977      -46.205880
#> 82        b     -120.4405274       42.330769
#> 83        c      113.0968148      -39.503268
#> 84        d       40.7582630      -10.221528
#> 85        e     -101.7615544      -47.619287
#> 86        f       79.8449483      -13.486986
#> 87        g     -145.3039190       26.749782
#> 88        h     -120.6541390       -8.000856
#> 89        i       14.7918481       -4.775513
#> 90        j       -3.4255197       15.386868
#> 91        a      -63.3890215       41.609016
#> 92        b      -60.6506761       18.242724
#> 93        c      -16.4819367      -40.212161
#> 94        d     -111.2968529      -74.371146
#> 95        e      131.9515087       72.935747
#> 96        f       -2.2085307       69.630493
#> 97        g      -11.5392238      -69.560664
#> 98        h     -175.7382772      -18.244204
#> 99        i     -102.4398989       -4.715051
#> 100       j      -47.9200000      -15.780000

if (FALSE) {
cc_inst(x, value = "flagged", buffer = 50000) #geod = T
}