Identify Coordinates in Vicinity of Country and Province Centroids

Removes or flags records within a radius around the geographic centroids of political countries and provinces. Poorly geo-referenced occurrence records in biological databases are often erroneously geo-referenced to centroids.

cc_cen(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  species = "species",
  buffer = 1000,
  geod = TRUE,
  test = "both",
  ref = NULL,
  verify = FALSE,
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

species

character string. The column with the species identity. Only required if verify = TRUE.

buffer

numerical. The buffer around each province or country centroid, where records should be flagged as problematic. Units depend on geod. Default = 1 kilometre.

geod

logical. If TRUE the radius around each capital is calculated based on a sphere, buffer is in meters and independent of latitude. If FALSE the radius is calculated assuming planar coordinates and varies slightly with latitude. Default = TRUE. See https://seethedatablog.wordpress.com/ for detail and credits.

test

a character string. Specifying the details of the test. One of c(“both”, “country”, “provinces”). If both tests for country and province centroids.

ref

SpatVector (geometry: polygons). Providing the geographic gazetteer. Can be any SpatVector (geometry: polygons), but the structure must be identical to countryref. Default = countryref.

verify

logical. If TRUE records are only flagged if they are the only record in a given species flagged close to a given reference. If FALSE, the distance is the only criterion

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = c(runif(99, -180, 180), -47.92), 
                decimalLatitude = c(runif(99, -90,90), -15.78))
cc_cen(x, geod = FALSE)
#> Testing country centroids
#> Removed 0 records.
#>     species decimalLongitude decimalLatitude
#> 1         a      153.2116455        6.864898
#> 2         b      -51.8226165       47.954720
#> 3         c      -68.2488702       60.319737
#> 4         d     -108.6948895       43.755272
#> 5         e      -34.0823077      -64.819348
#> 6         f     -111.7825222       -7.225878
#> 7         g      166.7345825      -42.721110
#> 8         h      -63.0073081        3.636927
#> 9         i       36.0825526       40.382164
#> 10        j      143.3979929       65.215597
#> 11        a      102.3620755       14.978173
#> 12        b      -12.9838045       39.559371
#> 13        c      -76.7220657      -46.143035
#> 14        d      177.8193690      -60.593170
#> 15        e      135.4363048        1.332386
#> 16        f      110.8922297       64.762947
#> 17        g       10.2365116      -75.878917
#> 18        h       -4.4716681      -56.233204
#> 19        i      153.5159556      -88.307803
#> 20        j     -162.3381130      -43.709681
#> 21        a     -103.9014694       35.133923
#> 22        b       90.1316763      -58.426278
#> 23        c      150.5492411        6.825895
#> 24        d      -41.1457019      -16.947176
#> 25        e      -23.3667661      -78.804963
#> 26        f      136.6069115       25.966838
#> 27        g     -117.0455624       27.414280
#> 28        h      131.1437100      -62.125057
#> 29        i     -105.5756180      -86.607649
#> 30        j       -7.8471803      -65.040881
#> 31        a      -55.2874193      -67.416054
#> 32        b     -101.6612909       86.670071
#> 33        c      111.2040677        6.545787
#> 34        d       -0.2244037      -18.547419
#> 35        e     -165.7006621       45.443327
#> 36        f       74.6717838      -23.940947
#> 37        g       14.6035417       80.562820
#> 38        h       18.8107558      -13.480874
#> 39        i     -103.7970119       36.145834
#> 40        j      -57.9280171       57.414189
#> 41        a       71.6537663      -44.423177
#> 42        b     -132.0816602      -28.653639
#> 43        c      -84.6665907       48.725226
#> 44        d      -62.5748141      -34.814548
#> 45        e      141.8687836      -39.485085
#> 46        f      131.1008710       27.160935
#> 47        g       40.2689439       85.733633
#> 48        h     -178.0910937       -5.681523
#> 49        i      -19.0846786      -48.237542
#> 50        j      156.5031317      -58.133016
#> 51        a     -113.9326117      -10.747666
#> 52        b      -11.3288631      -60.497166
#> 53        c     -116.5829583       46.242145
#> 54        d       90.3296675       59.446412
#> 55        e      -28.9611129      -43.322378
#> 56        f     -146.3226828       82.334842
#> 57        g      102.3002021       48.734863
#> 58        h       13.0403307       -9.250320
#> 59        i      -81.3187999       61.336221
#> 60        j       35.1990217      -87.468863
#> 61        a      -79.1922553      -61.179945
#> 62        b       73.4607458       78.071584
#> 63        c     -148.2444813       88.941199
#> 64        d     -145.5815212       16.913262
#> 65        e       13.5157192      -50.312940
#> 66        f       49.6482331      -64.448286
#> 67        g       58.2604314      -26.217688
#> 68        h     -126.0606843      -41.447184
#> 69        i      -78.2644632       11.415550
#> 70        j       49.8688896       82.333769
#> 71        a       52.0224004       85.482581
#> 72        b     -128.9780015      -33.788102
#> 73        c      113.1941902      -62.816776
#> 74        d      -94.6671206       35.403059
#> 75        e     -159.3180939      -13.202561
#> 76        f      -41.0202210       48.719959
#> 77        g      134.7593075      -29.450772
#> 78        h       27.9417819       67.912607
#> 79        i      -23.8451789       36.028546
#> 80        j     -101.0077415       85.683665
#> 81        a      107.8696893      -77.415814
#> 82        b        7.6390677      -61.769609
#> 83        c      -80.7960719       39.277689
#> 84        d     -175.9583189      -52.828813
#> 85        e       26.4102495       71.686245
#> 86        f     -128.4799349      -36.095582
#> 87        g       25.1309165       62.143875
#> 88        h     -160.7145688       54.124677
#> 89        i      113.1334971      -74.375377
#> 90        j      -98.1919413       18.360013
#> 91        a      124.5072977      -41.546812
#> 92        b     -134.0921360      -84.116935
#> 93        c     -142.1453083      -29.685245
#> 94        d       64.6490993       55.761319
#> 95        e      -44.7382614      -45.529504
#> 96        f      156.2244005        6.362170
#> 97        g      167.8361663      -68.936692
#> 98        h      -11.0799654       29.894134
#> 99        i       42.1130488      -72.982267
#> 100       j      -47.9200000      -15.780000

if (FALSE) {
cc_inst(x, value = "flagged", buffer = 50000) #geod = T
}