Identify Coordinates in Vicinity of Country and Province Centroids

Removes or flags records within a radius around the geographic centroids of political countries and provinces. Poorly geo-referenced occurrence records in biological databases are often erroneously geo-referenced to centroids.

cc_cen(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  species = "species",
  buffer = 1000,
  geod = TRUE,
  test = "both",
  ref = NULL,
  verify = FALSE,
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

species

character string. The column with the species identity. Only required if verify = TRUE.

buffer

numerical. The buffer around each province or country centroid, where records should be flagged as problematic. Units depend on geod. Default = 1 kilometre.

geod

logical. If TRUE the radius around each capital is calculated based on a sphere, buffer is in meters and independent of latitude. If FALSE the radius is calculated assuming planar coordinates and varies slightly with latitude. Default = TRUE. See https://seethedatablog.wordpress.com/ for detail and credits.

test

a character string. Specifying the details of the test. One of c(“both”, “country”, “provinces”). If both tests for country and province centroids.

ref

SpatVector (geometry: polygons). Providing the geographic gazetteer. Can be any SpatVector (geometry: polygons), but the structure must be identical to countryref. Default = countryref.

verify

logical. If TRUE records are only flagged if they are the only record in a given species flagged close to a given reference. If FALSE, the distance is the only criterion

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = c(runif(99, -180, 180), -47.92), 
                decimalLatitude = c(runif(99, -90,90), -15.78))
cc_cen(x, geod = FALSE)
#> Testing country centroids
#> Removed 0 records.
#>     species decimalLongitude decimalLatitude
#> 1         a      -153.452436      66.7928371
#> 2         b      -121.412322      77.4572371
#> 3         c      -115.310235     -35.6886726
#> 4         d        13.072208      39.6091232
#> 5         e        88.669630      -1.0245222
#> 6         f       163.962105     -71.5857483
#> 7         g       -73.147144       3.3245233
#> 8         h      -164.903713       9.1055321
#> 9         i        10.511347      -7.6987463
#> 10        j       -76.145649      46.4598273
#> 11        a      -167.363721      -4.0161727
#> 12        b        -1.505993     -37.5809258
#> 13        c       -14.169930     -82.0025564
#> 14        d       113.377067      86.3692576
#> 15        e       178.285296     -31.6312673
#> 16        f       -49.525752     -22.6861717
#> 17        g       -34.211839      73.1959600
#> 18        h       -50.649835      75.7092411
#> 19        i       118.562884      15.6670870
#> 20        j      -170.739481      20.0809377
#> 21        a        14.327398     -40.6463960
#> 22        b        78.421118      -3.3575509
#> 23        c      -131.874512      70.4970497
#> 24        d        47.009511     -37.2986393
#> 25        e       114.530166     -23.5528331
#> 26        f       176.625250      57.9633180
#> 27        g      -152.419985       2.3088285
#> 28        h       -10.566351     -10.9176350
#> 29        i       173.877728     -24.3799943
#> 30        j        -9.322745      68.4544735
#> 31        a       -83.842115     -62.6036768
#> 32        b      -138.162647      60.5747157
#> 33        c        80.243053      53.6682696
#> 34        d        15.472038     -44.8563614
#> 35        e      -168.978160      54.3875943
#> 36        f       107.561181       8.6165886
#> 37        g        68.315886     -71.6852619
#> 38        h      -151.523441     -26.4535009
#> 39        i       154.235674      74.8903608
#> 40        j       -48.888612      33.8503592
#> 41        a         4.123992     -55.1262131
#> 42        b       109.738699     -54.0321921
#> 43        c      -154.300818     -48.5498845
#> 44        d      -118.528589      56.7514995
#> 45        e      -170.692485      23.6748191
#> 46        f      -100.281954      25.1430178
#> 47        g        46.556368     -10.5459226
#> 48        h       -11.673325     -24.5128436
#> 49        i        81.541509      77.9930414
#> 50        j       119.288990      78.5303937
#> 51        a        50.224167     -57.0606981
#> 52        b       -72.301509      27.7018741
#> 53        c      -102.986646     -74.5648421
#> 54        d      -156.908209       7.3215291
#> 55        e       166.212969      12.3964054
#> 56        f      -177.981890     -47.4780610
#> 57        g        17.332807      78.5791108
#> 58        h       163.496092       4.2795512
#> 59        i      -124.411722     -33.7600934
#> 60        j        74.201731      67.7438894
#> 61        a       -94.132081      57.7490731
#> 62        b       -83.593625     -43.9754824
#> 63        c        31.955098     -53.5357084
#> 64        d        48.767909       1.5951127
#> 65        e       -38.170149      70.1283787
#> 66        f        18.935734      47.1170079
#> 67        g       -35.814799       4.3841258
#> 68        h        51.630596     -75.0823860
#> 69        i      -139.685943       8.1666273
#> 70        j      -160.183978      -4.3677339
#> 71        a        50.877241     -19.3248348
#> 72        b        55.615336      67.4444457
#> 73        c      -137.951606      13.3391705
#> 74        d       -95.054928      65.3081693
#> 75        e       135.079938      54.8683441
#> 76        f        79.673849     -10.4916825
#> 77        g       -93.781718     -15.2792753
#> 78        h        76.518681      65.1325164
#> 79        i      -164.082037      25.8080376
#> 80        j       -50.556991     -45.1463655
#> 81        a         4.408181      22.9508035
#> 82        b       118.233990       0.9680698
#> 83        c      -178.228334      79.9608829
#> 84        d        76.462832      40.7012497
#> 85        e       -28.211715     -89.3393868
#> 86        f       -20.523350      49.2588493
#> 87        g       -61.946836     -27.8582797
#> 88        h      -171.489162      44.9140938
#> 89        i       140.825456      63.8990824
#> 90        j      -126.271127      46.2639408
#> 91        a       107.748805     -47.9580476
#> 92        b       130.217097      87.3644881
#> 93        c      -159.144603      26.2155363
#> 94        d      -161.192407      70.3799067
#> 95        e       117.180736       0.3928347
#> 96        f        41.122181     -17.5538970
#> 97        g        46.011879     -57.9369845
#> 98        h        95.324396      47.0750289
#> 99        i      -106.433356     -28.1989708
#> 100       j       -47.920000     -15.7800000

if (FALSE) {
cc_inst(x, value = "flagged", buffer = 50000) #geod = T
}