Identify Coordinates in Vicinity of Country and Province Centroids

Removes or flags records within a radius around the geographic centroids of political countries and provinces. Poorly geo-referenced occurrence records in biological databases are often erroneously geo-referenced to centroids.

cc_cen(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  species = "species",
  buffer = 1000,
  geod = TRUE,
  test = "both",
  ref = NULL,
  verify = FALSE,
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

species

character string. The column with the species identity. Only required if verify = TRUE.

buffer

numerical. The buffer around each province or country centroid, where records should be flagged as problematic. Units depend on geod. Default = 1 kilometre.

geod

logical. If TRUE the radius around each capital is calculated based on a sphere, buffer is in meters and independent of latitude. If FALSE the radius is calculated assuming planar coordinates and varies slightly with latitude. Default = TRUE. See https://seethedatablog.wordpress.com/ for detail and credits.

test

a character string. Specifying the details of the test. One of c(“both”, “country”, “provinces”). If both tests for country and province centroids.

ref

SpatVector (geometry: polygons). Providing the geographic gazetteer. Can be any SpatVector (geometry: polygons), but the structure must be identical to countryref. Default = countryref.

verify

logical. If TRUE records are only flagged if they are the only record in a given species flagged close to a given reference. If FALSE, the distance is the only criterion

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = c(runif(99, -180, 180), -47.92), 
                decimalLatitude = c(runif(99, -90,90), -15.78))
cc_cen(x, geod = FALSE)
#> Testing country centroids
#> Removed 0 records.
#>     species decimalLongitude decimalLatitude
#> 1         a      -125.485085     -43.4552326
#> 2         b      -167.530088     -49.7981863
#> 3         c       -79.245613      61.8436898
#> 4         d        24.289278      11.9969005
#> 5         e       -15.816251     -62.4832147
#> 6         f      -109.850633     -45.9034655
#> 7         g       -35.230244     -57.1827877
#> 8         h        32.114371     -36.9917212
#> 9         i       -62.347442     -16.2620752
#> 10        j        90.431376       1.9653963
#> 11        a       -81.136302     -50.9168605
#> 12        b       124.640089     -43.7029596
#> 13        c       178.299639      69.0415817
#> 14        d       -26.036299     -35.7772187
#> 15        e        71.133468      60.5950228
#> 16        f       133.407189     -83.6816967
#> 17        g      -149.493136     -79.9626386
#> 18        h      -155.890336      21.4091765
#> 19        i        92.749706      32.7728841
#> 20        j         5.403019     -86.9667677
#> 21        a       100.281545      13.4713228
#> 22        b        11.099456     -36.8470271
#> 23        c       158.881531     -68.9689192
#> 24        d        20.123546     -26.7919878
#> 25        e      -179.086972       7.5341486
#> 26        f        39.266499      45.3444002
#> 27        g       -33.055675      -2.1417699
#> 28        h        97.081486     -57.7178679
#> 29        i      -123.367366       3.9507672
#> 30        j       -49.551836     -29.5406048
#> 31        a      -125.547654     -71.8402163
#> 32        b       103.661822      48.1220524
#> 33        c         9.052422      50.6002658
#> 34        d       122.528301      49.4765331
#> 35        e       171.483650      65.2288474
#> 36        f       133.759851      31.9755576
#> 37        g       106.636622      20.9076859
#> 38        h        92.886514     -88.9397550
#> 39        i      -135.399478      84.7660538
#> 40        j      -124.231840      61.4735754
#> 41        a       134.593735      23.4352956
#> 42        b      -155.405339      59.7334204
#> 43        c        56.672705     -32.8915285
#> 44        d       -99.153687      30.6552559
#> 45        e       -85.748779     -75.5936154
#> 46        f      -156.124351      53.6238424
#> 47        g        87.435630       5.9719564
#> 48        h       163.727018      43.4357731
#> 49        i       126.213866     -42.9422498
#> 50        j       146.684274      50.2470966
#> 51        a       131.750747     -12.0277610
#> 52        b        54.847806      56.0474999
#> 53        c       165.171115      11.4138162
#> 54        d        44.263408      26.3812745
#> 55        e      -115.376681     -36.1203771
#> 56        f      -164.082493      50.3230750
#> 57        g        19.764433      48.4043070
#> 58        h       -32.876116     -43.0683563
#> 59        i        13.060474      69.0669121
#> 60        j       -43.344018     -78.1598795
#> 61        a      -147.085867      17.5439291
#> 62        b       -52.538990      86.6255276
#> 63        c        99.920482      78.3884763
#> 64        d       139.918393       0.4723651
#> 65        e       110.921577      56.7553404
#> 66        f       117.962817     -25.8603519
#> 67        g       175.522167     -43.6055730
#> 68        h      -163.363631      40.7220149
#> 69        i        57.372352     -33.2440948
#> 70        j        70.598487      -7.4488554
#> 71        a        51.770884     -46.0396112
#> 72        b       -90.831270      -8.6507685
#> 73        c       131.006643      16.1652422
#> 74        d        31.182628      -1.6947084
#> 75        e        98.024350      78.7171155
#> 76        f      -154.609495      24.6634646
#> 77        g       147.026728     -72.8446488
#> 78        h        81.324020     -71.7180562
#> 79        i        76.191560      40.3545526
#> 80        j       105.972076       6.4378895
#> 81        a      -140.527407      -2.2547755
#> 82        b       176.904202      66.3384314
#> 83        c       109.873070      62.5988595
#> 84        d      -141.497460       7.3078330
#> 85        e       179.850893     -14.2524842
#> 86        f       -30.600960      54.5528599
#> 87        g       177.639547      85.3075250
#> 88        h      -177.303027      63.4987880
#> 89        i       144.068786      57.2276419
#> 90        j      -121.580996      34.3412488
#> 91        a       -72.937206     -12.9294466
#> 92        b       -47.951809      49.0823588
#> 93        c        63.060634      -9.1449729
#> 94        d      -146.442395     -47.0421950
#> 95        e        21.550756     -57.1591421
#> 96        f       141.669098     -87.3465800
#> 97        g         3.122605     -62.5650984
#> 98        h        74.883905      34.3481893
#> 99        i       108.234428     -38.5254295
#> 100       j       -47.920000     -15.7800000

if (FALSE) {
cc_inst(x, value = "flagged", buffer = 50000) #geod = T
}