Identify Coordinates in Vicinity of Country and Province Centroids

Removes or flags records within a radius around the geographic centroids of political countries and provinces. Poorly geo-referenced occurrence records in biological databases are often erroneously geo-referenced to centroids.

cc_cen(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  species = "species",
  buffer = 1000,
  geod = TRUE,
  test = "both",
  ref = NULL,
  verify = FALSE,
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

species

character string. The column with the species identity. Only required if verify = TRUE.

buffer

numerical. The buffer around each province or country centroid, where records should be flagged as problematic. Units depend on geod. Default = 1 kilometre.

geod

logical. If TRUE the radius around each capital is calculated based on a sphere, buffer is in meters and independent of latitude. If FALSE the radius is calculated assuming planar coordinates and varies slightly with latitude. Default = TRUE. See https://seethedatablog.wordpress.com/ for detail and credits.

test

a character string. Specifying the details of the test. One of c(“both”, “country”, “provinces”). If both tests for country and province centroids.

ref

SpatVector (geometry: polygons). Providing the geographic gazetteer. Can be any SpatVector (geometry: polygons), but the structure must be identical to countryref. Default = countryref.

verify

logical. If TRUE records are only flagged if they are the only record in a given species flagged close to a given reference. If FALSE, the distance is the only criterion

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = c(runif(99, -180, 180), -47.92), 
                decimalLatitude = c(runif(99, -90,90), -15.78))
cc_cen(x, geod = FALSE)
#> Testing country centroids
#> Removed 0 records.
#>     species decimalLongitude decimalLatitude
#> 1         a        9.2128551      65.9006963
#> 2         b     -140.1613826      87.7205731
#> 3         c      -80.2337208      45.3598206
#> 4         d        2.5741909     -77.7244682
#> 5         e       24.4863043     -24.8937223
#> 6         f       81.4716492      32.8190875
#> 7         g     -122.1418976      33.3575805
#> 8         h      173.7045493     -80.4015144
#> 9         i      -45.6312450       8.5805302
#> 10        j     -165.9842476     -82.3582610
#> 11        a      -98.5613828      64.6435026
#> 12        b       -8.2911430     -74.3788686
#> 13        c      -90.2857729     -25.1738978
#> 14        d      -62.3951158     -33.1691671
#> 15        e       57.0703273      79.2591331
#> 16        f      176.3462239     -88.2489877
#> 17        g     -119.5639640      52.6720970
#> 18        h     -172.9047001     -71.0143510
#> 19        i      129.1808331      12.8681802
#> 20        j     -129.3396894      67.8563104
#> 21        a       21.9531841      83.2056269
#> 22        b     -124.3118335      33.0965821
#> 23        c     -119.8937666     -60.2853188
#> 24        d     -127.1241741     -43.6111809
#> 25        e     -132.9059781      38.7642851
#> 26        f      125.2395861     -78.4952578
#> 27        g      120.7209731      54.2947659
#> 28        h      -79.0666651      87.8109205
#> 29        i     -109.7471035     -79.3944430
#> 30        j       22.3346439      88.2234541
#> 31        a      115.8549044     -33.6219404
#> 32        b     -121.4945791      10.1003853
#> 33        c       99.7404114      88.3839290
#> 34        d     -125.9563969     -29.4293562
#> 35        e      -73.8203800      35.0053239
#> 36        f      174.1027850     -53.7588752
#> 37        g      153.9574273     -29.4234824
#> 38        h       75.5533902      26.6533342
#> 39        i       -0.5311928      13.2434133
#> 40        j       16.7939154     -39.0170450
#> 41        a      139.4523384       5.3783018
#> 42        b      165.2927465      25.6606104
#> 43        c       96.9858650      81.8522233
#> 44        d       80.8139771       0.5948523
#> 45        e       23.8462229     -27.5852566
#> 46        f      -59.8051495     -65.2840909
#> 47        g       40.7712614     -47.1074131
#> 48        h     -163.2375347      20.6141603
#> 49        i      175.9004796     -14.0424356
#> 50        j      120.1871984     -40.1454585
#> 51        a      177.0642694     -33.2699656
#> 52        b      -89.3141523      37.4378291
#> 53        c      175.9134994     -61.8752709
#> 54        d       69.2789732      26.9970965
#> 55        e      -43.4092899      35.2424933
#> 56        f     -150.9445111     -76.9085066
#> 57        g      -29.1400090      78.8665298
#> 58        h     -145.4367059     -48.4108653
#> 59        i      115.3735736     -18.0152533
#> 60        j      114.0117881     -49.5507080
#> 61        a      176.9159731      83.6897528
#> 62        b       67.0742468      81.4637672
#> 63        c       17.2780207     -38.6266072
#> 64        d       83.1884420     -61.1299825
#> 65        e     -151.9367294     -56.9476641
#> 66        f      155.3180946     -46.8417174
#> 67        g       61.0613495      10.6138338
#> 68        h        1.5120227     -21.3586327
#> 69        i       65.7399439      11.5950878
#> 70        j      -75.0186676      83.0127265
#> 71        a       93.1356222     -85.3676403
#> 72        b     -177.5905105      14.9620235
#> 73        c       78.4120031       7.0231456
#> 74        d     -132.5084293      26.1515314
#> 75        e       45.2866365     -17.8780362
#> 76        f     -109.6799577      17.7072519
#> 77        g      -26.7264689      68.3485136
#> 78        h     -160.7725755     -38.4208609
#> 79        i      -94.8125170      81.0857194
#> 80        j      -69.8297076     -61.8170147
#> 81        a      -54.3437282      82.9693415
#> 82        b      136.9883520      33.0047185
#> 83        c       10.0207667      25.6087061
#> 84        d     -121.2511668      -9.8727734
#> 85        e       57.5057923     -11.5458965
#> 86        f      -76.5767723     -25.5865159
#> 87        g     -101.0817899      80.4744293
#> 88        h      124.5170254     -63.7161254
#> 89        i       74.6889662      84.1500839
#> 90        j      -40.3862008     -24.5071590
#> 91        a      126.8348103     -17.1342168
#> 92        b      137.3192753      -2.9579266
#> 93        c     -170.6405212     -72.6027294
#> 94        d       44.4954094      82.3946006
#> 95        e      -41.3803835     -19.9368608
#> 96        f     -100.9380075     -50.3038360
#> 97        g        1.5876139      55.5190501
#> 98        h      -37.9824607     -53.1238690
#> 99        i       92.6808832     -29.4766551
#> 100       j      -47.9200000     -15.7800000

if (FALSE) {
cc_inst(x, value = "flagged", buffer = 50000) #geod = T
}