Identify Coordinates in Vicinity of Country and Province Centroids

Removes or flags records within a radius around the geographic centroids of political countries and provinces. Poorly geo-referenced occurrence records in biological databases are often erroneously geo-referenced to centroids.

cc_cen(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  species = "species",
  buffer = 1000,
  geod = TRUE,
  test = "both",
  ref = NULL,
  verify = FALSE,
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

species

character string. The column with the species identity. Only required if verify = TRUE.

buffer

numerical. The buffer around each province or country centroid, where records should be flagged as problematic. Units depend on geod. Default = 1 kilometre.

geod

logical. If TRUE the radius around each capital is calculated based on a sphere, buffer is in meters and independent of latitude. If FALSE the radius is calculated assuming planar coordinates and varies slightly with latitude. Default = TRUE. See https://seethedatablog.wordpress.com/ for detail and credits.

test

a character string. Specifying the details of the test. One of c(“both”, “country”, “provinces”). If both tests for country and province centroids.

ref

SpatVector (geometry: polygons). Providing the geographic gazetteer. Can be any SpatVector (geometry: polygons), but the structure must be identical to countryref. Default = countryref.

verify

logical. If TRUE records are only flagged if they are the only record in a given species flagged close to a given reference. If FALSE, the distance is the only criterion

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = c(runif(99, -180, 180), -47.92), 
                decimalLatitude = c(runif(99, -90,90), -15.78))
cc_cen(x, geod = FALSE)
#> Testing country centroids
#> Removed 0 records.
#>     species decimalLongitude decimalLatitude
#> 1         a       84.6841484       19.590602
#> 2         b       -4.5783034       78.727483
#> 3         c       50.4432095       42.417119
#> 4         d     -131.2876604       59.585105
#> 5         e     -145.4462985       48.321454
#> 6         f       11.1756372       68.620256
#> 7         g       36.7617774      -74.001147
#> 8         h      -26.6767418       -7.871401
#> 9         i       82.3167042      -45.762354
#> 10        j      124.5595376      -89.833977
#> 11        a        6.1558889       31.845548
#> 12        b      145.9699936      -46.896850
#> 13        c      133.5544765       72.069284
#> 14        d       58.6272515      -14.316038
#> 15        e       19.6275736      -75.527300
#> 16        f     -113.4620847      -85.691264
#> 17        g       18.1197555       80.750763
#> 18        h     -109.7546354      -47.779098
#> 19        i     -156.3633843       30.333825
#> 20        j       27.7541096        5.390269
#> 21        a      -47.5875755       61.789700
#> 22        b      -56.6794901      -30.910732
#> 23        c       99.6632637      -53.606110
#> 24        d     -156.1075065       16.291033
#> 25        e       -0.6940288       62.657576
#> 26        f       32.8054360      -84.272661
#> 27        g     -146.7665176      -74.012641
#> 28        h     -166.1417375      -88.822421
#> 29        i      171.4493688       51.429155
#> 30        j       11.5988789      -58.770251
#> 31        a      -75.7391417      -88.603604
#> 32        b      -18.8295308      -16.700986
#> 33        c      177.8741202       85.727976
#> 34        d      131.9029939      -67.479290
#> 35        e      -39.0577164      -63.540194
#> 36        f     -136.4322140      -22.024228
#> 37        g      127.9086225       44.172135
#> 38        h     -133.9252846      -79.505020
#> 39        i       -9.9493036       20.557687
#> 40        j      176.7365222      -16.475285
#> 41        a      -40.0745931        8.729361
#> 42        b     -126.0258088      -49.522436
#> 43        c      103.8091915      -19.222265
#> 44        d      139.3642513      -65.406471
#> 45        e      114.7064485      -23.014499
#> 46        f       21.2822491      -26.064636
#> 47        g     -144.3316837       22.427619
#> 48        h      139.6198341      -56.040248
#> 49        i     -151.9885522      -22.436492
#> 50        j      -96.0148176      -50.577861
#> 51        a     -149.9745497      -87.240771
#> 52        b       96.7651148       32.289553
#> 53        c     -120.6251073       31.235176
#> 54        d     -106.4172795       62.857147
#> 55        e      136.9131321      -23.863175
#> 56        f      148.9527268      -47.195573
#> 57        g      -61.9206159      -65.983337
#> 58        h     -137.9338099      -59.768225
#> 59        i     -102.9047031      -64.684131
#> 60        j       10.3070003      -20.374079
#> 61        a     -164.2027851      -23.144831
#> 62        b       24.7239865        1.364365
#> 63        c       95.4267064      -17.712412
#> 64        d      -18.3966010       -8.776076
#> 65        e      102.1513966      -38.110704
#> 66        f     -168.7883138      -41.961019
#> 67        g       55.3374684       27.460454
#> 68        h       48.7065084       69.163278
#> 69        i      -19.5221929       68.294085
#> 70        j      108.5171976       12.309993
#> 71        a     -140.8844963      -75.562654
#> 72        b      -21.0203436       88.986194
#> 73        c      139.6783130       53.335893
#> 74        d     -118.7368462        5.912293
#> 75        e      -36.4814624      -30.357862
#> 76        f       48.2842445      -87.652511
#> 77        g      -27.4886998       -4.527607
#> 78        h      -15.5511753       56.011127
#> 79        i      -60.7526788       60.562239
#> 80        j     -164.8569326       82.346842
#> 81        a       17.2132777       25.730368
#> 82        b     -101.1046130       73.551687
#> 83        c     -156.8220302       10.148948
#> 84        d      -65.2570807      -38.824078
#> 85        e      155.5501935      -84.404516
#> 86        f      105.1589851       83.074192
#> 87        g       15.6021389       -1.567878
#> 88        h      162.9218668      -80.456593
#> 89        i      -15.8181105      -16.416822
#> 90        j     -137.3742411       -5.132535
#> 91        a       85.6014553        1.881816
#> 92        b      -58.6239285       82.413984
#> 93        c      -23.6727333      -66.150949
#> 94        d      113.0127479      -69.124382
#> 95        e      -53.6663948       73.584391
#> 96        f       35.4674468       54.558779
#> 97        g        8.6783629      -23.723221
#> 98        h      117.6284932      -62.452436
#> 99        i     -108.0336117      -55.950157
#> 100       j      -47.9200000      -15.780000

if (FALSE) {
cc_inst(x, value = "flagged", buffer = 50000) #geod = T
}