Identify Coordinates in Vicinity of Country and Province Centroids

Removes or flags records within a radius around the geographic centroids of political countries and provinces. Poorly geo-referenced occurrence records in biological databases are often erroneously geo-referenced to centroids.

cc_cen(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  species = "species",
  buffer = 1000,
  geod = TRUE,
  test = "both",
  ref = NULL,
  verify = FALSE,
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

species

character string. The column with the species identity. Only required if verify = TRUE.

buffer

numerical. The buffer around each province or country centroid, where records should be flagged as problematic. Units depend on geod. Default = 1 kilometre.

geod

logical. If TRUE the radius around each capital is calculated based on a sphere, buffer is in meters and independent of latitude. If FALSE the radius is calculated assuming planar coordinates and varies slightly with latitude. Default = TRUE. See https://seethedatablog.wordpress.com/ for detail and credits.

test

a character string. Specifying the details of the test. One of c(“both”, “country”, “provinces”). If both tests for country and province centroids.

ref

SpatVector (geometry: polygons). Providing the geographic gazetteer. Can be any SpatVector (geometry: polygons), but the structure must be identical to countryref. Default = countryref.

verify

logical. If TRUE records are only flagged if they are the only record in a given species flagged close to a given reference. If FALSE, the distance is the only criterion

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = c(runif(99, -180, 180), -47.92), 
                decimalLatitude = c(runif(99, -90,90), -15.78))
cc_cen(x, geod = FALSE)
#> Testing country centroids
#> Removed 0 records.
#>     species decimalLongitude decimalLatitude
#> 1         a     -152.3778955      77.4079512
#> 2         b      -55.4955156      20.6006228
#> 3         c       42.0136769     -71.4137607
#> 4         d      -56.6899794     -71.3654122
#> 5         e      -10.0483458     -80.8953412
#> 6         f     -128.1393378     -86.6887675
#> 7         g      -21.0874112      66.2285264
#> 8         h       52.9581996     -16.3003378
#> 9         i      -63.8559570      -7.6424894
#> 10        j       17.2319697     -13.8158335
#> 11        a      140.0720337     -88.8598842
#> 12        b      119.9272635      58.7033761
#> 13        c       20.6757118      88.2451458
#> 14        d     -111.6404766     -49.8044823
#> 15        e      157.4242668     -74.0755166
#> 16        f      153.3434420     -62.0286172
#> 17        g      -66.3459826     -63.6301353
#> 18        h     -100.9380699     -18.9171920
#> 19        i      116.5070991     -30.6856769
#> 20        j       80.1118815      29.5483388
#> 21        a      139.7271284      39.1088299
#> 22        b     -105.2308232     -85.9519383
#> 23        c       51.2161153      24.1925289
#> 24        d      -66.4547828      66.8998648
#> 25        e     -118.2174180     -36.3434809
#> 26        f       51.8824508      13.4789206
#> 27        g       35.1565415     -32.4222778
#> 28        h      -87.2546580     -57.1646022
#> 29        i      -69.4582365      42.8361060
#> 30        j     -107.0328653      -5.3298789
#> 31        a        5.9217308     -83.5571334
#> 32        b       56.4684962     -60.0043907
#> 33        c        0.3051574      74.8013418
#> 34        d     -138.0483735      -0.9700496
#> 35        e        7.0255128      13.8126293
#> 36        f       -6.8667458      33.9310789
#> 37        g       26.4843927     -43.7662934
#> 38        h      -12.7791458      64.3464984
#> 39        i      -10.7590858      55.5288102
#> 40        j      -82.0679991      -5.4104757
#> 41        a     -160.6508186     -11.2470789
#> 42        b       23.8339436     -29.8093604
#> 43        c      166.0367812      66.1458040
#> 44        d      162.6009364     -28.8182907
#> 45        e       12.9817143      -5.1945904
#> 46        f      -58.6555522      81.1042349
#> 47        g       80.0525167      42.9043623
#> 48        h      -13.2292418      29.4052399
#> 49        i      -60.9104536     -67.8580568
#> 50        j      -14.5417751     -63.2585276
#> 51        a       37.7108782     -54.2500952
#> 52        b      138.6483227      32.3349842
#> 53        c       64.6171974       2.8310873
#> 54        d     -177.5774807      74.2933247
#> 55        e      -19.2830012      12.1282332
#> 56        f        5.9834883     -43.8510893
#> 57        g      129.1296212      51.1140960
#> 58        h       23.7427589     -41.3081110
#> 59        i       -8.1651239      77.0911961
#> 60        j     -161.9310484      69.6759734
#> 61        a      -36.0354821     -54.5197013
#> 62        b      113.9281139     -44.4331300
#> 63        c     -110.4887636     -25.1797448
#> 64        d      101.5153918     -72.3553514
#> 65        e      -41.3133123     -63.9058450
#> 66        f      -43.3581220      61.5359830
#> 67        g       37.1500676      43.5107161
#> 68        h      116.4938934      15.7173349
#> 69        i     -165.7573678      18.8598396
#> 70        j      153.9048231     -12.0360017
#> 71        a      114.3706094     -87.0086260
#> 72        b      -90.8324958     -21.5317355
#> 73        c      -19.6988171     -89.6061180
#> 74        d     -166.6754358     -24.4133486
#> 75        e      115.9787309     -24.6235522
#> 76        f     -156.8040544      12.2919973
#> 77        g       80.1187365     -79.6594545
#> 78        h     -102.4781008      33.2287006
#> 79        i      -53.1704145     -16.3086011
#> 80        j      -68.3382974      30.1775145
#> 81        a      -23.4542791     -72.3844230
#> 82        b       10.8355973       3.6238787
#> 83        c      159.5108420      55.1604665
#> 84        d      -13.3718471      51.5381553
#> 85        e     -170.5353367      67.1104498
#> 86        f      137.9933309     -58.4149058
#> 87        g     -126.3031933      81.1670765
#> 88        h     -100.4865575      48.9813384
#> 89        i      153.5880368     -53.8991555
#> 90        j      -13.5424665      19.8539382
#> 91        a      -62.1372920       6.7020257
#> 92        b       57.0035038      67.4906715
#> 93        c      123.8121071       5.2863049
#> 94        d      -42.2106867     -10.1619987
#> 95        e     -134.6138725       9.3702640
#> 96        f       25.0679824      85.3127688
#> 97        g      -24.0051808      12.5475841
#> 98        h      -84.8283060     -62.4725229
#> 99        i     -112.0637405      28.5305052
#> 100       j      -47.9200000     -15.7800000

if (FALSE) {
cc_inst(x, value = "flagged", buffer = 50000) #geod = T
}