Identify Coordinates in Vicinity of Country and Province Centroids

Removes or flags records within a radius around the geographic centroids of political countries and provinces. Poorly geo-referenced occurrence records in biological databases are often erroneously geo-referenced to centroids.

cc_cen(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  species = "species",
  buffer = 1000,
  geod = TRUE,
  test = "both",
  ref = NULL,
  verify = FALSE,
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

species

character string. The column with the species identity. Only required if verify = TRUE.

buffer

numerical. The buffer around each province or country centroid, where records should be flagged as problematic. Units depend on geod. Default = 1 kilometre.

geod

logical. If TRUE the radius around each capital is calculated based on a sphere, buffer is in meters and independent of latitude. If FALSE the radius is calculated assuming planar coordinates and varies slightly with latitude. Default = TRUE. See https://seethedatablog.wordpress.com/ for detail and credits.

test

a character string. Specifying the details of the test. One of c(“both”, “country”, “provinces”). If both tests for country and province centroids.

ref

SpatVector (geometry: polygons). Providing the geographic gazetteer. Can be any SpatVector (geometry: polygons), but the structure must be identical to countryref. Default = countryref.

verify

logical. If TRUE records are only flagged if they are the only record in a given species flagged close to a given reference. If FALSE, the distance is the only criterion

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = c(runif(99, -180, 180), -47.92), 
                decimalLatitude = c(runif(99, -90,90), -15.78))
cc_cen(x, geod = FALSE)
#> Testing country centroids
#> Removed 0 records.
#>     species decimalLongitude decimalLatitude
#> 1         a      -28.9400766     -39.2195186
#> 2         b      -20.5230493      89.7310824
#> 3         c      -59.9544258      69.1750430
#> 4         d      147.1826585      13.8673118
#> 5         e       82.7881636       5.3311329
#> 6         f      152.6542253      54.0535913
#> 7         g       59.0269713     -12.9534951
#> 8         h     -160.2686388     -42.4158188
#> 9         i      117.0313600     -41.4853882
#> 10        j       30.7468795      14.5535183
#> 11        a     -167.0503629     -31.6730522
#> 12        b       17.2193352      -6.9894891
#> 13        c       56.1215925     -61.2162757
#> 14        d     -174.1115621      39.4694411
#> 15        e     -138.9884788      13.3896217
#> 16        f     -150.8678992     -23.5213688
#> 17        g       77.5832090      48.6809745
#> 18        h      -95.3671420      -5.0227804
#> 19        i      -25.6273186     -67.4449433
#> 20        j       -7.8884572     -25.1981155
#> 21        a      149.4289245      61.4020290
#> 22        b      179.8567394     -42.5501638
#> 23        c      -14.6962035     -52.5869096
#> 24        d     -164.6729369       3.7849481
#> 25        e       11.7449396      17.1951113
#> 26        f       -0.6414888      -4.0920655
#> 27        g       10.8529324       0.6123704
#> 28        h       36.1452875      28.8560542
#> 29        i      -53.6143805     -12.7892040
#> 30        j      -42.1000541      79.8614717
#> 31        a       73.1365374      68.0003319
#> 32        b     -168.1233439     -53.6907843
#> 33        c     -138.0891825     -80.5332953
#> 34        d      -34.3276105     -33.5426767
#> 35        e     -143.7518159     -47.7228144
#> 36        f       39.5640377      28.0288145
#> 37        g       95.3025988     -87.6884166
#> 38        h      117.7617558      61.3498253
#> 39        i     -177.8166161      54.0664847
#> 40        j     -123.3911570     -67.9119094
#> 41        a      102.0499597     -88.7803622
#> 42        b      126.4736724      12.2349644
#> 43        c      148.3283273     -64.9058048
#> 44        d     -116.5133655     -30.4419731
#> 45        e     -155.0810323      72.3700442
#> 46        f      -69.9828088      -9.2828328
#> 47        g     -114.0848973      -9.6880216
#> 48        h      -84.2223254      70.9511461
#> 49        i      159.2705247      17.6237966
#> 50        j     -118.5980315     -27.9052602
#> 51        a      164.2340368      23.3525869
#> 52        b       22.9085972     -87.0831363
#> 53        c     -126.6452612      72.3230382
#> 54        d      -11.3409579      -7.1746754
#> 55        e       17.5518993      79.9555095
#> 56        f       66.1196782     -86.4901332
#> 57        g     -142.0437923     -62.8457975
#> 58        h       51.7451480     -64.0465179
#> 59        i      -66.2936007     -46.6964830
#> 60        j     -134.8601282      80.3258120
#> 61        a      -71.5307047     -50.2749503
#> 62        b      -48.7062774     -11.1518086
#> 63        c       30.3402723       2.5919315
#> 64        d      108.4668973      21.3189092
#> 65        e     -149.8136407     -78.8276608
#> 66        f      -22.5193384     -49.1955827
#> 67        g      116.0477033     -85.5013909
#> 68        h      171.5143103      17.8293400
#> 69        i      -47.3461822      14.3866138
#> 70        j       99.9934773      13.7835863
#> 71        a     -125.4310295     -12.3279883
#> 72        b     -130.8277946      29.1158060
#> 73        c      114.6624270      82.3589888
#> 74        d      -82.5514489      36.6066807
#> 75        e       20.9217440      10.5006358
#> 76        f        3.8884046     -27.5700413
#> 77        g     -148.6209470      19.7267191
#> 78        h      177.0328951       1.1391400
#> 79        i     -110.1217936     -89.0839803
#> 80        j     -145.6747487       7.7506828
#> 81        a       68.7252126       6.9885505
#> 82        b     -121.8154951     -66.7435633
#> 83        c      144.8164972     -62.3695047
#> 84        d       62.2975239      57.2933081
#> 85        e      113.9837766     -36.1287467
#> 86        f      158.9330172      31.3698404
#> 87        g      162.0650472     -59.6545671
#> 88        h     -150.6364902      65.1637656
#> 89        i     -114.9581780     -39.4013642
#> 90        j      115.0830011     -84.7815212
#> 91        a     -136.7655912      18.1707268
#> 92        b      -81.8870773      85.9122056
#> 93        c      -25.5880261      78.8410337
#> 94        d     -130.1731834     -39.7362216
#> 95        e     -135.8393225      61.8837311
#> 96        f      -23.0312324      65.8701395
#> 97        g      171.4439192      56.6023713
#> 98        h       61.0760197     -42.9220089
#> 99        i     -144.5701652      55.4194125
#> 100       j      -47.9200000     -15.7800000

if (FALSE) {
cc_inst(x, value = "flagged", buffer = 50000) #geod = T
}