Identify Coordinates in Vicinity of Country and Province Centroids

Removes or flags records within a radius around the geographic centroids of political countries and provinces. Poorly geo-referenced occurrence records in biological databases are often erroneously geo-referenced to centroids.

cc_cen(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  species = "species",
  buffer = 1000,
  geod = TRUE,
  test = "both",
  ref = NULL,
  verify = FALSE,
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

species

character string. The column with the species identity. Only required if verify = TRUE.

buffer

numerical. The buffer around each province or country centroid, where records should be flagged as problematic. Units depend on geod. Default = 1 kilometre.

geod

logical. If TRUE the radius around each capital is calculated based on a sphere, buffer is in meters and independent of latitude. If FALSE the radius is calculated assuming planar coordinates and varies slightly with latitude. Default = TRUE. See https://seethedatablog.wordpress.com/ for detail and credits.

test

a character string. Specifying the details of the test. One of c(“both”, “country”, “provinces”). If both tests for country and province centroids.

ref

SpatVector (geometry: polygons). Providing the geographic gazetteer. Can be any SpatVector (geometry: polygons), but the structure must be identical to countryref. Default = countryref.

verify

logical. If TRUE records are only flagged if they are the only record in a given species flagged close to a given reference. If FALSE, the distance is the only criterion

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = c(runif(99, -180, 180), -47.92), 
                decimalLatitude = c(runif(99, -90,90), -15.78))
cc_cen(x, geod = FALSE)
#> Testing country centroids
#> Removed 0 records.
#>     species decimalLongitude decimalLatitude
#> 1         a       132.061248     -42.4567724
#> 2         b       -19.152738      37.4891265
#> 3         c       -73.792451      -8.0729713
#> 4         d        71.060038       4.8296366
#> 5         e        26.355005     -15.5274388
#> 6         f        94.931524      68.2848012
#> 7         g       128.903126      69.8395657
#> 8         h       -35.668345       0.6705683
#> 9         i      -144.159842      79.3891893
#> 10        j       -35.752163     -19.8631330
#> 11        a       106.457926     -42.5531368
#> 12        b       160.446586     -64.6409843
#> 13        c        48.558562      61.0315196
#> 14        d        73.131486       6.6652452
#> 15        e       -63.674883     -20.6565099
#> 16        f       -80.945084     -58.5405067
#> 17        g      -146.183637     -26.9231944
#> 18        h       -43.524209      -1.9742308
#> 19        i      -145.249207      85.8620231
#> 20        j       109.007809      -7.1553118
#> 21        a        38.224494      36.7505695
#> 22        b        16.430870      29.5979671
#> 23        c       -16.308549     -20.9254182
#> 24        d       133.562816     -52.4137334
#> 25        e        83.100398      21.9890072
#> 26        f      -179.950810     -61.8082137
#> 27        g        29.652820     -54.5662375
#> 28        h      -175.675961     -45.8338737
#> 29        i      -153.765717       2.7097138
#> 30        j        46.255397      39.9560902
#> 31        a       -55.263854     -80.8342880
#> 32        b       152.907983      23.2148059
#> 33        c        39.072918      -0.3806713
#> 34        d        94.526388      24.9442638
#> 35        e      -152.386767      59.4681492
#> 36        f       -75.143316      36.5354234
#> 37        g       132.214265     -25.1487751
#> 38        h       179.119155     -33.4137661
#> 39        i       172.835946     -14.4336065
#> 40        j       -45.493786      40.6772554
#> 41        a       -74.043412     -85.7817170
#> 42        b        74.741620     -37.9890041
#> 43        c        26.110230      50.3463952
#> 44        d       117.657129      43.1007713
#> 45        e       -23.879736       3.9290577
#> 46        f      -134.892281     -67.7041890
#> 47        g         9.436922     -47.4693020
#> 48        h      -135.373238      -6.6129929
#> 49        i      -163.571673      58.5305638
#> 50        j       -28.266192      22.6848208
#> 51        a       -22.969003      49.1391573
#> 52        b        65.259329     -75.6311447
#> 53        c       141.487387     -45.8262562
#> 54        d        94.137829     -62.0718790
#> 55        e       166.201660      24.2714334
#> 56        f      -142.813231      19.2213231
#> 57        g       -43.773682      82.7364543
#> 58        h        80.146489     -29.8666469
#> 59        i      -167.792826     -86.2165220
#> 60        j        -3.696307     -35.8873279
#> 61        a       172.398689     -52.4417242
#> 62        b        19.173821       3.8640152
#> 63        c       161.024534      60.9006185
#> 64        d       -33.067055      49.3216647
#> 65        e       -23.007586      78.2825107
#> 66        f      -101.831006     -23.2212450
#> 67        g       -16.756287     -38.1733956
#> 68        h       -15.245861      50.4028772
#> 69        i       142.401031      19.1750964
#> 70        j        27.012398      -4.2150133
#> 71        a      -155.096494     -62.9075364
#> 72        b      -112.470141     -89.3408980
#> 73        c        54.447200     -49.3907321
#> 74        d       -86.532662     -80.1212782
#> 75        e        89.960030     -57.5535253
#> 76        f        97.314636      41.3880661
#> 77        g      -169.236842     -70.0522142
#> 78        h        94.577431      78.9225098
#> 79        i        52.409644     -20.0227006
#> 80        j      -166.586325     -82.5418337
#> 81        a        77.044580      89.6316593
#> 82        b       110.125267      65.3259719
#> 83        c      -132.195475      58.6759088
#> 84        d       -23.052707      61.5943813
#> 85        e       -60.578107     -50.7179688
#> 86        f        92.732433     -16.3006974
#> 87        g        98.750356     -74.4974724
#> 88        h        25.066458     -40.0618010
#> 89        i      -129.994439      34.4192058
#> 90        j       137.142126      41.9327970
#> 91        a      -176.166696     -48.8651409
#> 92        b        75.022278      42.4001825
#> 93        c        -3.225703     -26.0874257
#> 94        d       -53.857439     -12.8360042
#> 95        e       112.834829      80.1406530
#> 96        f       -75.407173      38.5583468
#> 97        g      -122.895762     -31.4517117
#> 98        h       -49.972946      30.2464135
#> 99        i       -44.637111      84.7321010
#> 100       j       -47.920000     -15.7800000

if (FALSE) {
cc_inst(x, value = "flagged", buffer = 50000) #geod = T
}