Identify Coordinates in Vicinity of Country and Province Centroids

Removes or flags records within a radius around the geographic centroids of political countries and provinces. Poorly geo-referenced occurrence records in biological databases are often erroneously geo-referenced to centroids.

cc_cen(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  species = "species",
  buffer = 1000,
  geod = TRUE,
  test = "both",
  ref = NULL,
  verify = FALSE,
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

species

character string. The column with the species identity. Only required if verify = TRUE.

buffer

numerical. The buffer around each province or country centroid, where records should be flagged as problematic. Units depend on geod. Default = 1 kilometre.

geod

logical. If TRUE the radius around each capital is calculated based on a sphere, buffer is in meters and independent of latitude. If FALSE the radius is calculated assuming planar coordinates and varies slightly with latitude. Default = TRUE. See https://seethedatablog.wordpress.com/ for detail and credits.

test

a character string. Specifying the details of the test. One of c(“both”, “country”, “provinces”). If both tests for country and province centroids.

ref

SpatVector (geometry: polygons). Providing the geographic gazetteer. Can be any SpatVector (geometry: polygons), but the structure must be identical to countryref. Default = countryref.

verify

logical. If TRUE records are only flagged if they are the only record in a given species flagged close to a given reference. If FALSE, the distance is the only criterion

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = c(runif(99, -180, 180), -47.92), 
                decimalLatitude = c(runif(99, -90,90), -15.78))
cc_cen(x, geod = FALSE)
#> Testing country centroids
#> Removed 0 records.
#>     species decimalLongitude decimalLatitude
#> 1         a       141.251321       46.917863
#> 2         b        40.369808      -73.946320
#> 3         c       120.511227      -22.188387
#> 4         d        62.618052       45.239823
#> 5         e      -136.264671      -37.674450
#> 6         f       166.676737       87.765910
#> 7         g       176.416313       37.296296
#> 8         h      -122.585243       65.775467
#> 9         i      -160.175777       -4.743583
#> 10        j        -1.313123      -27.296162
#> 11        a         7.591688       67.845119
#> 12        b        46.188688       14.663589
#> 13        c       152.215046      -65.379560
#> 14        d       107.992689      -40.798022
#> 15        e        43.388671       34.737896
#> 16        f       142.893549       23.106688
#> 17        g      -107.779059       48.176670
#> 18        h        60.979837       78.088772
#> 19        i        99.389157      -45.800683
#> 20        j      -103.900838      -84.012862
#> 21        a       -21.101248       72.879711
#> 22        b        -6.770962       71.936882
#> 23        c      -167.206461       86.538412
#> 24        d       -74.732632      -61.342910
#> 25        e       -13.952730       52.273410
#> 26        f      -125.616163       62.176288
#> 27        g      -112.737548       86.344664
#> 28        h        93.254200       -2.138746
#> 29        i       -31.653596       64.398412
#> 30        j      -126.136862      -27.793002
#> 31        a      -158.924359      -48.208376
#> 32        b      -161.482878      -69.575931
#> 33        c      -177.873785      -39.467684
#> 34        d        21.687300       68.857140
#> 35        e      -108.202159      -42.937304
#> 36        f      -161.809366       -9.485971
#> 37        g        20.401661       42.637874
#> 38        h       123.243444       -2.723110
#> 39        i       156.851895      -39.626081
#> 40        j      -108.258325       42.178344
#> 41        a       172.649697       37.623690
#> 42        b      -114.039291      -88.148232
#> 43        c       -17.471592       51.018396
#> 44        d       -32.924744      -32.081327
#> 45        e      -155.840520       58.863853
#> 46        f        80.173934      -34.382174
#> 47        g       -65.499246       78.746039
#> 48        h      -151.922426        9.462066
#> 49        i       173.337534      -29.348518
#> 50        j        26.491371      -33.611463
#> 51        a       -62.722587      -64.824931
#> 52        b      -113.962240      -12.080629
#> 53        c      -160.930345       12.374035
#> 54        d        -6.387953      -74.916091
#> 55        e      -147.527917       62.386814
#> 56        f       -84.669973       58.276676
#> 57        g      -161.019319       56.139875
#> 58        h      -165.216939       39.308462
#> 59        i       -79.134538       87.988521
#> 60        j        13.924423       54.006664
#> 61        a         1.533268       44.133641
#> 62        b       108.288321       27.191970
#> 63        c       178.203647      -25.305281
#> 64        d        85.824852       71.897374
#> 65        e        17.302071      -39.518492
#> 66        f        -6.268465      -78.064961
#> 67        g      -146.190000      -72.941553
#> 68        h      -158.381803      -66.353186
#> 69        i       124.784554      -85.533079
#> 70        j       -76.209593        7.634840
#> 71        a       -61.489268      -40.858366
#> 72        b      -169.461478       36.023910
#> 73        c      -140.746685      -43.327019
#> 74        d       -12.130706      -80.561955
#> 75        e        86.189838       43.227109
#> 76        f      -121.624083       59.304764
#> 77        g      -179.809591       -2.357013
#> 78        h      -101.521779       21.231597
#> 79        i       152.659813      -41.342278
#> 80        j        34.786103       26.619448
#> 81        a        62.878460      -47.838509
#> 82        b       -21.471399       49.429528
#> 83        c       170.405647       13.937843
#> 84        d       -68.767533      -86.060446
#> 85        e        32.291914       34.040899
#> 86        f       -27.683566       87.524801
#> 87        g       -78.887218      -18.145581
#> 88        h      -133.176195       81.520176
#> 89        i       119.068743        6.334218
#> 90        j        28.422350       71.882481
#> 91        a       -59.432938      -65.428766
#> 92        b       113.457361       -5.193539
#> 93        c        -7.983557      -23.952919
#> 94        d        38.186406      -21.699428
#> 95        e       -72.465044      -44.781962
#> 96        f       -33.238247       79.397257
#> 97        g       -59.649206      -55.581655
#> 98        h       -52.759523       69.245630
#> 99        i        88.549676       41.989671
#> 100       j       -47.920000      -15.780000

if (FALSE) {
cc_inst(x, value = "flagged", buffer = 50000) #geod = T
}