Flag Coordinates in Vicinity of Country and Province Centroids

Flags records within a radius around the geographic centroids of political countries and provinces. Poorly geo-referenced occurrence records in biological databases are often erroneously geo-referenced to centroids.

cc_cen(x, lon = "decimallongitude", lat = "decimallatitude", buffer = 0.1,
  test = "both", ref = NULL, value = "clean", verbose = TRUE)

Arguments

x
a data.frame. Containing geographical coordinates and species names.

lon
a character string. The column with the longitude coordinates. Default = “decimallongitude”.

lat
a character string. The column with the longitude coordinates. Default = “decimallatitude”.

buffer
numerical. The buffer around each province or country centroid, where records should be flagged as problematic, in decimal degrees. Default = 0.1.

test
a character string. Specifying the details of the test. One of c(“both”, “country”, “provinces”). If both tests for country and province centroids.

ref
a SpatialPointsDataframe. Providing the geographic gazetteer. Can be any SpatialPointsDataframe, but the structure must be identical to countryref. Default = countryref

value
a character string. Defining the output value. See value.

verbose
logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic. Default = “clean”.

Details

Note: the buffer radius is in degrees, thus will differ slightly between different latitudes.

Note

See https://github.com/azizka/CoordinateCleaner/wiki for more details and tutorials.

Examples

x <- data.frame(species = letters[1:10],
                decimallongitude = runif(100, -180, 180),
                decimallatitude = runif(100, -90,90))

cc_cen(x)

#> Testing country centroids

#> Flagged 0 records.

#>     species decimallongitude decimallatitude
#> 1         a       -54.252343       29.896562
#> 2         b       161.034576      -53.297878
#> 3         c      -102.204009        8.746459
#> 4         d      -168.446625       58.950255
#> 5         e      -127.686296      -67.039623
#> 6         f       127.578200      -42.986499
#> 7         g      -103.266248      -38.084935
#> 8         h      -104.288134      -87.341577
#> 9         i      -165.772552       63.670271
#> 10        j       160.118929      -15.471391
#> 11        a       -91.825925       45.182492
#> 12        b       101.204125       39.422512
#> 13        c       -76.234619        2.763954
#> 14        d       135.128848       85.470526
#> 15        e       -73.529966      -25.060039
#> 16        f       174.069147      -59.328809
#> 17        g        32.341520       50.965321
#> 18        h        93.297017      -84.433699
#> 19        i       120.987110       52.095665
#> 20        j        94.615009      -60.011338
#> 21        a       -29.782827      -84.831738
#> 22        b      -130.293057       51.471210
#> 23        c      -150.895814       58.555415
#> 24        d        56.153745       83.743275
#> 25        e        36.721389      -21.844588
#> 26        f        56.518499      -58.610461
#> 27        g       -61.445821       18.675072
#> 28        h       172.610719       54.983472
#> 29        i        77.467008      -83.397706
#> 30        j       134.146910       41.923920
#> 31        a       173.982149      -51.250922
#> 32        b      -101.317323      -87.119434
#> 33        c        59.230823      -66.851536
#> 34        d       -39.756946       33.491359
#> 35        e      -163.417090       25.549428
#> 36        f        42.089241      -31.079066
#> 37        g        35.450998      -20.253243
#> 38        h       -33.532693       37.350071
#> 39        i       128.998134       35.577977
#> 40        j         6.365227       76.618218
#> 41        a       172.545628       -7.376810
#> 42        b      -173.874352       17.271756
#> 43        c        62.441217      -60.298022
#> 44        d       -46.342842       12.043761
#> 45        e       150.483829       71.653927
#> 46        f        64.072114       17.005011
#> 47        g        59.454886       59.704185
#> 48        h        92.174792       16.813506
#> 49        i        15.421376       50.214714
#> 50        j       -93.856285      -18.401104
#> 51        a         3.201686       62.978899
#> 52        b       -29.784826       43.532205
#> 53        c        81.701587      -32.797761
#> 54        d        49.566799      -69.897560
#> 55        e       -37.292416      -71.802834
#> 56        f       165.413739       54.017595
#> 57        g       -72.483109      -21.609586
#> 58        h      -161.927580      -80.513825
#> 59        i        27.427471       87.571566
#> 60        j      -101.553909       18.751314
#> 61        a      -134.691743      -63.219110
#> 62        b       157.734967        6.976831
#> 63        c       108.459046      -67.357939
#> 64        d        92.899303       83.372716
#> 65        e        11.723459      -81.551697
#> 66        f        16.849719      -60.946347
#> 67        g      -145.466459       78.567737
#> 68        h       -40.194089       84.722645
#> 69        i      -117.953320       37.849329
#> 70        j        68.661305       69.316367
#> 71        a        63.075061       85.837557
#> 72        b       160.666145      -83.718064
#> 73        c      -109.360972      -11.023692
#> 74        d       168.709501       26.004557
#> 75        e       -40.645340       89.069833
#> 76        f        54.123804      -33.360728
#> 77        g       113.254633       64.058754
#> 78        h      -154.452682        7.260783
#> 79        i         9.658917       67.215123
#> 80        j        94.850937        2.805412
#> 81        a       -23.260809       66.502674
#> 82        b        18.890041       64.174067
#> 83        c      -106.548967      -27.777027
#> 84        d      -168.830634      -89.981336
#> 85        e       169.094541      -53.067646
#> 86        f      -115.699286       80.139688
#> 87        g       100.185403      -39.359968
#> 88        h       138.855887       68.568159
#> 89        i       121.120650       83.000930
#> 90        j        37.932639      -89.276678
#> 91        a       146.476605        9.522037
#> 92        b      -167.072469      -50.537525
#> 93        c      -132.689336       26.410886
#> 94        d      -146.149068       40.986368
#> 95        e        70.770117       67.065740
#> 96        f       -33.937660      -21.243322
#> 97        g      -156.370810       70.694227
#> 98        h      -134.462656       61.896651
#> 99        i       157.438878       41.357018
#> 100       j      -102.103117      -83.085304

cc_cen(x, value = "flagged")

#> Testing country centroids

#> Flagged 0 records.

#>    1    2    3    4    5    6    7    8    9   10   11   12   13   14   15   16 
#> TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE 
#>   17   18   19   20   21   22   23   24   25   26   27   28   29   30   31   32 
#> TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE 
#>   33   34   35   36   37   38   39   40   41   42   43   44   45   46   47   48 
#> TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE 
#>   49   50   51   52   53   54   55   56   57   58   59   60   61   62   63   64 
#> TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE 
#>   65   66   67   68   69   70   71   72   73   74   75   76   77   78   79   80 
#> TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE 
#>   81   82   83   84   85   86   87   88   89   90   91   92   93   94   95   96 
#> TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE 
#>   97   98   99  100 
#> TRUE TRUE TRUE TRUE