Identify Records with Identical lat/lon

Removes or flags records with equal latitude and longitude coordinates, either exact or absolute. Equal coordinates can often indicate data entry errors.

cc_equ(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  test = "absolute",
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

test

character string. Defines if coordinates are compared exactly (“identical”) or on the absolute scale (i.e. -1 = 1, “absolute”). Default is to “absolute”.

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = runif(100, -180, 180), 
                decimalLatitude = runif(100, -90,90))

cc_equ(x)
#> Testing equal lat/lon
#> Removed 0 records.
#>     species decimalLongitude decimalLatitude
#> 1         a      -151.415451      -11.253416
#> 2         b       144.502861      -48.491601
#> 3         c       135.767846       82.726697
#> 4         d      -104.168691      -46.736979
#> 5         e        52.860919      -50.371945
#> 6         f       -70.111782      -54.483703
#> 7         g        50.492195      -34.389347
#> 8         h       133.571764      -39.734287
#> 9         i       133.917605       86.016183
#> 10        j       107.061767      -80.896210
#> 11        a       -50.070511       16.804385
#> 12        b       -23.929240       14.922960
#> 13        c       116.760752       57.991744
#> 14        d       156.082862      -58.846521
#> 15        e        36.143937       52.257159
#> 16        f       -40.133287       66.809304
#> 17        g       119.662586        6.259581
#> 18        h      -123.709550       37.556865
#> 19        i        90.092860        5.008395
#> 20        j      -113.721954       -4.266972
#> 21        a       159.307998       32.987480
#> 22        b         5.683480      -80.967272
#> 23        c      -179.301113       33.524595
#> 24        d       -85.017268       37.556694
#> 25        e        60.123422       -2.427849
#> 26        f        21.581088       50.193705
#> 27        g      -112.916811       30.398769
#> 28        h       -68.829315       39.802771
#> 29        i        15.507126      -81.990808
#> 30        j       -53.461344      -83.420709
#> 31        a      -109.310446       53.678671
#> 32        b       164.422379      -67.128663
#> 33        c      -143.091774       74.717958
#> 34        d      -132.262539      -27.010990
#> 35        e       -21.582143      -61.135707
#> 36        f       119.548652       14.612050
#> 37        g        34.626181       57.948112
#> 38        h      -171.526838      -33.105645
#> 39        i       151.202750      -50.776912
#> 40        j       100.888267      -66.904899
#> 41        a       132.622628       46.859516
#> 42        b        46.910960      -89.666162
#> 43        c       -80.782058       53.667882
#> 44        d       162.904850       26.340849
#> 45        e       161.839572       37.642658
#> 46        f       -48.441686       70.191672
#> 47        g       -70.220982       53.545113
#> 48        h       127.405795       36.772214
#> 49        i       167.242919        4.353729
#> 50        j       109.335658       34.040880
#> 51        a        22.847756      -83.732760
#> 52        b        74.526925       73.279184
#> 53        c      -149.618700       41.607676
#> 54        d        17.492810       35.556252
#> 55        e        51.421125      -43.548108
#> 56        f       -67.247012       13.148124
#> 57        g       106.547560       46.616052
#> 58        h      -108.051916      -24.303666
#> 59        i      -137.702004       69.005264
#> 60        j       -59.688010       69.945417
#> 61        a      -115.514276      -55.209122
#> 62        b       116.682088       58.127905
#> 63        c       161.868887       14.924340
#> 64        d       161.075567      -18.976938
#> 65        e       -28.959315        6.869066
#> 66        f       -91.136295        6.876811
#> 67        g      -111.451780      -13.092532
#> 68        h      -165.697507      -48.400512
#> 69        i        23.844832        8.881796
#> 70        j      -114.758824       62.901195
#> 71        a       -26.633993       67.916165
#> 72        b       -43.314458      -31.071117
#> 73        c       166.144234      -42.894452
#> 74        d        99.895763       59.392828
#> 75        e       -25.824791      -37.897729
#> 76        f       111.520154       22.709035
#> 77        g        63.256754      -68.183927
#> 78        h       -62.527429      -53.023664
#> 79        i       -72.194509       31.446138
#> 80        j         3.245852      -55.840570
#> 81        a       106.691325       34.247388
#> 82        b       -16.812987      -83.613943
#> 83        c      -179.140389      -54.357578
#> 84        d        -5.889271       62.799559
#> 85        e        48.552592      -69.313223
#> 86        f       147.599150      -69.278452
#> 87        g      -154.885253      -54.011868
#> 88        h        43.891788       42.232013
#> 89        i        42.282465       78.883957
#> 90        j       144.546092      -32.105520
#> 91        a       149.684533        2.142019
#> 92        b        34.561438      -27.165433
#> 93        c        64.044930      -29.593231
#> 94        d       -83.552121       30.653976
#> 95        e       -18.708383       30.460046
#> 96        f       -63.198925      -85.538199
#> 97        g        16.365208       50.736668
#> 98        h       -22.373626       -5.643252
#> 99        i       140.701109      -19.076264
#> 100       j       -24.357786       49.202652
cc_equ(x, value = "flagged")
#> Testing equal lat/lon
#> Flagged 0 records.
#>   [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [16] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [31] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [46] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [61] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [76] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [91] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE