Identify Records with Identical lat/lon

Removes or flags records with equal latitude and longitude coordinates, either exact or absolute. Equal coordinates can often indicate data entry errors.

cc_equ(
  x,
  lon = "decimallongitude",
  lat = "decimallatitude",
  test = "absolute",
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimallongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimallatitude”.

test

character string. Defines if coordinates are compared exactly (“identical”) or on the absolute scale (i.e. -1 = 1, “absolute”). Default is to “absolute”.

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimallongitude = runif(100, -180, 180), 
                decimallatitude = runif(100, -90,90))

cc_equ(x)
#> Testing equal lat/lon
#> Removed 0 records.
#>     species decimallongitude decimallatitude
#> 1         a        55.770211    -79.52210391
#> 2         b       -18.285695     11.02525029
#> 3         c        49.700320    -46.84665543
#> 4         d        62.475717    -41.33497998
#> 5         e       149.340694     67.25637940
#> 6         f        41.982840     89.34478525
#> 7         g       121.765106    -52.91695842
#> 8         h        93.653583    -36.93329960
#> 9         i      -139.060379     28.80441402
#> 10        j       112.094433     78.44060174
#> 11        a        89.376628    -41.39567351
#> 12        b        78.950057     75.01515043
#> 13        c        34.544327     19.92752728
#> 14        d      -173.190495     -0.08896187
#> 15        e        42.240436    -13.91471875
#> 16        f      -105.564476    -44.08057194
#> 17        g        79.587529    -47.06819056
#> 18        h        33.121691    -75.13167896
#> 19        i       -20.063226     11.57409953
#> 20        j       -27.868633      7.28677563
#> 21        a       -62.663265    -24.49732454
#> 22        b       -40.281434    -12.49674263
#> 23        c        27.929797    -65.87863977
#> 24        d        -3.332245    -47.22149427
#> 25        e       -82.118995    -10.64553645
#> 26        f       -58.424803    -75.33996615
#> 27        g      -175.162821    -50.54183178
#> 28        h        51.360973     58.78595650
#> 29        i      -163.854921    -30.86868520
#> 30        j       155.076249     23.30935133
#> 31        a       -67.811130    -31.45164963
#> 32        b       111.306115     33.01905815
#> 33        c        42.482453    -72.84357111
#> 34        d        -7.680424    -54.82025900
#> 35        e       -30.128770    -18.45172266
#> 36        f        69.782446     25.53415243
#> 37        g       141.309296     81.50985871
#> 38        h       100.581440     16.51961581
#> 39        i        98.721087     86.22999413
#> 40        j       165.226513      0.86278328
#> 41        a       175.821616    -29.73247049
#> 42        b       -91.881171      5.21399753
#> 43        c       -66.838832    -18.51070426
#> 44        d        88.260755     48.57002841
#> 45        e      -114.727082     -9.33370176
#> 46        f       152.555672     30.43787340
#> 47        g       162.932714    -52.96762421
#> 48        h       119.920603     12.45613256
#> 49        i      -145.708393     57.99094987
#> 50        j        72.132936     78.79919867
#> 51        a      -125.340430    -33.09072518
#> 52        b       130.203798     -4.02911631
#> 53        c       -44.868720    -25.56399565
#> 54        d       -71.050380    -10.24916111
#> 55        e       -47.174258      6.20712674
#> 56        f       -69.600817     14.06851251
#> 57        g      -134.228460    -40.26544901
#> 58        h        70.656618    -47.52529317
#> 59        i        52.392578      2.50342730
#> 60        j       -61.512510    -81.42507115
#> 61        a      -106.215885      7.78896177
#> 62        b       -65.582032    -78.24408090
#> 63        c        69.488890     20.04689929
#> 64        d       148.575312      6.24685352
#> 65        e       -53.294868    -88.90321171
#> 66        f        28.471367    -71.08226214
#> 67        g       119.768710    -14.80139595
#> 68        h       132.102528    -40.81394273
#> 69        i       150.431236    -48.88527414
#> 70        j      -157.090699     87.87456410
#> 71        a      -103.929342    -39.24690319
#> 72        b        56.175215     -2.19950850
#> 73        c       -71.879612     71.13093333
#> 74        d      -172.016475     71.30588536
#> 75        e      -176.883274      3.20529170
#> 76        f       116.924216     -1.42080216
#> 77        g        29.405260     80.93318839
#> 78        h         6.233633     -2.83488270
#> 79        i       -87.101791    -21.44430898
#> 80        j       -53.252502    -33.97761175
#> 81        a        63.520742     64.29268290
#> 82        b       133.704154    -52.58824628
#> 83        c       136.778478    -63.15715848
#> 84        d      -160.507077    -74.63332603
#> 85        e      -124.286246    -36.40051033
#> 86        f       -72.950185     46.54988141
#> 87        g       -26.335140    -41.43429149
#> 88        h       178.108289      7.37514402
#> 89        i        64.754709     84.11624470
#> 90        j       170.555501    -24.66087953
#> 91        a      -160.814171    -17.82805420
#> 92        b      -134.116811     -7.52638226
#> 93        c        61.436221    -76.30705426
#> 94        d      -171.306432    -38.48745058
#> 95        e        92.053574     67.56993039
#> 96        f        18.746793      9.80201701
#> 97        g      -152.547891    -78.65321215
#> 98        h       -51.536705    -81.85346557
#> 99        i      -127.866479    -77.39496959
#> 100       j        99.728258     69.56634499
cc_equ(x, value = "flagged")
#> Testing equal lat/lon
#> Flagged 0 records.
#>   [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [16] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [31] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [46] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [61] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [76] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [91] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE