Identify Records with Identical lat/lon

Removes or flags records with equal latitude and longitude coordinates, either exact or absolute. Equal coordinates can often indicate data entry errors.

cc_equ(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  test = "absolute",
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

test

character string. Defines if coordinates are compared exactly (“identical”) or on the absolute scale (i.e. -1 = 1, “absolute”). Default is to “absolute”.

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = runif(100, -180, 180), 
                decimalLatitude = runif(100, -90,90))

cc_equ(x)
#> Testing equal lat/lon
#> Removed 0 records.
#>     species decimalLongitude decimalLatitude
#> 1         a       -98.577081      46.8754520
#> 2         b       -76.211211       2.9564808
#> 3         c       -23.293345      -7.9108657
#> 4         d       120.943936     -36.7792211
#> 5         e        36.551308     -28.6576259
#> 6         f      -146.066412      15.3505676
#> 7         g        56.021911     -11.8934499
#> 8         h         8.128677      47.5101549
#> 9         i        37.054842      71.9909839
#> 10        j       -74.165315      23.1063198
#> 11        a        -4.850806      86.2410361
#> 12        b       -84.566300     -74.7110831
#> 13        c       171.800352     -61.4479993
#> 14        d       -54.838301     -15.9160516
#> 15        e       -15.796458      61.1496201
#> 16        f       150.596788      49.2194950
#> 17        g       139.483255     -60.4252907
#> 18        h       -58.877532      79.1444132
#> 19        i       161.976726     -78.7554027
#> 20        j       -56.290915      82.8551414
#> 21        a      -176.331324     -75.2381622
#> 22        b      -132.511852     -76.7210704
#> 23        c       152.841184     -85.6853529
#> 24        d       -95.906929      81.2345364
#> 25        e       145.849418     -50.5576851
#> 26        f        61.941934      11.0966810
#> 27        g      -161.117964     -55.9653698
#> 28        h        64.665704     -37.7659941
#> 29        i        30.747225      39.0325425
#> 30        j        30.969899      58.2717821
#> 31        a       -69.740312     -68.9339009
#> 32        b       -81.385343      87.3986814
#> 33        c        91.914875      31.6303827
#> 34        d       -94.576078      65.3315713
#> 35        e      -154.381768     -15.6874850
#> 36        f       141.819593      -8.7012316
#> 37        g       137.351610     -30.4354324
#> 38        h       155.411347      70.4936720
#> 39        i        -7.270781     -24.6954490
#> 40        j        45.124976     -84.3258485
#> 41        a      -175.358545      -0.4218264
#> 42        b       149.547137     -26.4094487
#> 43        c       -78.275112      17.6148861
#> 44        d        86.483857      80.5040813
#> 45        e       -80.206934     -38.1212722
#> 46        f      -171.856641      18.8544901
#> 47        g       128.331262     -82.5090245
#> 48        h        -2.829939      59.4162560
#> 49        i      -164.299230      38.0129642
#> 50        j       160.231420      12.1372246
#> 51        a       -31.335301      71.2096825
#> 52        b        62.701434     -44.3177973
#> 53        c        56.537889      15.6911540
#> 54        d       -22.474516      78.2175748
#> 55        e      -131.463015      25.9620524
#> 56        f        62.443166      31.5592257
#> 57        g      -127.400960     -57.4180847
#> 58        h       -62.516327     -10.5921125
#> 59        i       175.629797      -8.3643549
#> 60        j       -84.037333      89.1032349
#> 61        a       123.210753     -31.3901096
#> 62        b       -68.431248     -47.3270717
#> 63        c       -41.305404     -52.0509513
#> 64        d        76.836621      46.7388790
#> 65        e       175.987628     -49.5441662
#> 66        f        75.821738      45.4545520
#> 67        g       -93.333721     -50.2001068
#> 68        h       -64.826572      47.8992409
#> 69        i      -161.138121     -39.9540170
#> 70        j       -71.424357      19.2046482
#> 71        a       -39.413378     -17.0207407
#> 72        b       -79.423063      36.6960506
#> 73        c        76.452360     -44.6516475
#> 74        d      -115.358558     -12.1607934
#> 75        e      -157.592508      56.1935325
#> 76        f        -3.915331     -30.4546867
#> 77        g       -71.741131      79.5910704
#> 78        h        -1.513553       2.2768059
#> 79        i       167.522680     -73.9310176
#> 80        j       -41.241076     -41.1745961
#> 81        a       -86.618163      27.4585646
#> 82        b        90.310158     -13.3952165
#> 83        c        69.528297      40.3961247
#> 84        d      -101.856899     -57.2081588
#> 85        e       -37.049953      -4.5312082
#> 86        f       -74.325057      53.9172969
#> 87        g      -111.867989     -79.5300093
#> 88        h       -37.389009     -78.3128756
#> 89        i       -24.976763      58.9118611
#> 90        j       -46.698502      12.3822980
#> 91        a       156.400883      -0.5041476
#> 92        b       -62.657430      42.0715236
#> 93        c        35.282209      -8.3128278
#> 94        d       163.253146     -82.7372799
#> 95        e      -112.652460      37.5880736
#> 96        f       147.566585      49.0706472
#> 97        g       169.533414      33.2208199
#> 98        h      -120.825573      78.5716279
#> 99        i       -55.838024      89.5183257
#> 100       j      -161.987742     -61.1341690
cc_equ(x, value = "flagged")
#> Testing equal lat/lon
#> Flagged 0 records.
#>   [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [16] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [31] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [46] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [61] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [76] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [91] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE