Identify Records with Identical lat/lon

Removes or flags records with equal latitude and longitude coordinates, either exact or absolute. Equal coordinates can often indicate data entry errors.

cc_equ(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  test = "absolute",
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

test

character string. Defines if coordinates are compared exactly (“identical”) or on the absolute scale (i.e. -1 = 1, “absolute”). Default is to “absolute”.

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = runif(100, -180, 180), 
                decimalLatitude = runif(100, -90,90))

cc_equ(x)
#> Testing equal lat/lon
#> Removed 0 records.
#>     species decimalLongitude decimalLatitude
#> 1         a         7.067998      -31.140357
#> 2         b       -40.012499       51.581162
#> 3         c        29.794232      -52.008655
#> 4         d       121.775542      -58.208518
#> 5         e      -140.180905      -10.372002
#> 6         f      -141.232906       20.923197
#> 7         g        22.173575       57.591763
#> 8         h        87.391824      -34.169824
#> 9         i      -108.000733       12.259310
#> 10        j       -75.668319      -63.242491
#> 11        a      -149.813402       -7.745388
#> 12        b       -79.283197       20.581268
#> 13        c       120.284959       53.717831
#> 14        d        71.697161       32.189330
#> 15        e       -96.107136       29.890827
#> 16        f       171.703602      -31.293160
#> 17        g       -11.357096      -64.230726
#> 18        h        -5.234903      -87.499633
#> 19        i      -110.237129        4.483792
#> 20        j        35.999650       46.529266
#> 21        a        -4.032601       49.079288
#> 22        b        95.463863        3.937135
#> 23        c       -97.805737      -75.284288
#> 24        d       -92.305320      -22.011439
#> 25        e       -92.877452       75.852555
#> 26        f      -102.326610      -29.654220
#> 27        g       -58.640388       13.639229
#> 28        h      -103.241918       54.679508
#> 29        i        34.618591      -41.687583
#> 30        j      -114.820691      -33.030836
#> 31        a         5.688872       79.889213
#> 32        b       -87.837082       88.815539
#> 33        c       -60.915945       52.725149
#> 34        d      -134.292637       19.786802
#> 35        e       130.889732       18.015793
#> 36        f        12.516426       85.574038
#> 37        g        31.178553       63.559490
#> 38        h         1.258421       12.552191
#> 39        i        15.968919      -44.047360
#> 40        j       -94.930019       71.985941
#> 41        a      -141.245341      -71.906222
#> 42        b      -173.374424      -54.822654
#> 43        c       147.305610       53.247721
#> 44        d      -115.287843       34.455921
#> 45        e      -123.101383       70.147599
#> 46        f      -174.378551      -68.932723
#> 47        g       134.569191      -50.907478
#> 48        h        57.725809      -49.878388
#> 49        i        55.472187        4.020881
#> 50        j       161.698057      -46.075831
#> 51        a       -99.343844       89.756216
#> 52        b       -66.511785      -44.600733
#> 53        c       149.024464       62.448137
#> 54        d        49.938085       44.563177
#> 55        e        66.269455      -79.073497
#> 56        f        23.557310      -13.148828
#> 57        g       -70.944617      -32.085492
#> 58        h        21.196142      -65.725263
#> 59        i       103.784749       41.205686
#> 60        j        88.600421       73.724869
#> 61        a      -153.501938      -43.165449
#> 62        b        95.205488       20.192915
#> 63        c       135.077767        2.606729
#> 64        d       -41.588880       -7.974020
#> 65        e      -115.078937       -1.997116
#> 66        f       173.511828       -6.355198
#> 67        g       177.641825       78.585737
#> 68        h        55.787863       11.114320
#> 69        i        75.566018      -38.333761
#> 70        j      -143.943517       21.078160
#> 71        a       -44.969018       68.432985
#> 72        b       131.366790      -28.758243
#> 73        c      -142.687256      -84.036772
#> 74        d        -5.252239      -60.778378
#> 75        e       -28.671767      -19.089609
#> 76        f       -90.056311      -24.021353
#> 77        g        13.265668      -70.884105
#> 78        h      -114.637500      -17.278523
#> 79        i       106.666295       48.566390
#> 80        j       162.256439      -41.461220
#> 81        a       -75.868874       36.964144
#> 82        b        80.446851       11.349737
#> 83        c       151.075997       28.385182
#> 84        d       -62.243311       27.012205
#> 85        e      -168.993351       24.393144
#> 86        f        26.931006       31.102094
#> 87        g        76.665531      -86.232703
#> 88        h        83.577529       -1.085133
#> 89        i       103.055995      -80.313835
#> 90        j      -152.548724      -43.350132
#> 91        a      -174.898099      -77.242885
#> 92        b       -24.770758      -44.716792
#> 93        c        57.221283       85.003382
#> 94        d        50.654542       40.561380
#> 95        e       -62.431995      -73.046803
#> 96        f      -135.350054      -16.328699
#> 97        g       -11.870695      -41.859277
#> 98        h        95.637561       -1.070021
#> 99        i       -39.538527      -23.144166
#> 100       j      -151.611994      -68.700889
cc_equ(x, value = "flagged")
#> Testing equal lat/lon
#> Flagged 0 records.
#>   [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [16] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [31] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [46] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [61] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [76] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [91] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE