Identify Records with Identical lat/lon

Removes or flags records with equal latitude and longitude coordinates, either exact or absolute. Equal coordinates can often indicate data entry errors.

cc_equ(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  test = "absolute",
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

test

character string. Defines if coordinates are compared exactly (“identical”) or on the absolute scale (i.e. -1 = 1, “absolute”). Default is to “absolute”.

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = runif(100, -180, 180), 
                decimalLatitude = runif(100, -90,90))

cc_equ(x)
#> Testing equal lat/lon
#> Removed 0 records.
#>     species decimalLongitude decimalLatitude
#> 1         a       -21.443994     -32.2069896
#> 2         b      -138.388930     -76.1378496
#> 3         c       -97.463741     -35.7503665
#> 4         d         4.353422      61.6578750
#> 5         e        34.272854     -44.3398872
#> 6         f        15.942519     -30.6383643
#> 7         g       -54.733904      84.0399759
#> 8         h      -177.700583      64.5769729
#> 9         i       -17.267472     -10.1307232
#> 10        j        81.173501     -26.4698494
#> 11        a       145.907976      14.3191328
#> 12        b      -146.801340     -18.2533159
#> 13        c       -45.344956      14.0177701
#> 14        d        -3.828756     -31.3442715
#> 15        e       178.211704      54.4339357
#> 16        f       -83.982362     -47.6442698
#> 17        g       -54.084352      85.9207901
#> 18        h        -2.255115      56.3294117
#> 19        i      -138.902206     -58.9076233
#> 20        j      -162.235633       3.4776144
#> 21        a       -89.085531      36.7524576
#> 22        b        85.811623     -51.0350291
#> 23        c      -141.365998     -72.9599915
#> 24        d      -119.335799     -68.3533873
#> 25        e        97.514934     -64.8653123
#> 26        f       106.873360      47.2538299
#> 27        g      -149.886631      85.4010597
#> 28        h        -6.886240      85.3980624
#> 29        i      -124.435871      24.7701286
#> 30        j       -50.880435      -3.6052348
#> 31        a       -95.979815      70.4292248
#> 32        b        72.671844     -30.4898774
#> 33        c      -135.327283     -88.2957122
#> 34        d      -119.817668      23.1623950
#> 35        e       -29.674475     -61.2433937
#> 36        f        -3.735339     -22.3402483
#> 37        g       -97.217437      -1.0672597
#> 38        h      -127.551593     -80.9592689
#> 39        i        -5.688143     -67.7462873
#> 40        j        81.988236       4.0950041
#> 41        a        77.169545      73.4845019
#> 42        b       176.416039       0.8752822
#> 43        c       -72.993391     -23.9434359
#> 44        d       120.035341      78.7990628
#> 45        e        55.482796      45.7086770
#> 46        f       104.280969      54.7942397
#> 47        g       -42.331899      58.4259382
#> 48        h       145.324291     -74.8321164
#> 49        i       -53.809235      37.7555408
#> 50        j       -95.699083      18.6537701
#> 51        a       -15.753797      29.6571966
#> 52        b       -42.482927      36.4570884
#> 53        c       -45.313768     -32.4908227
#> 54        d       -44.773120      45.1869764
#> 55        e       118.560285      70.7841692
#> 56        f      -139.867381     -57.7322448
#> 57        g        52.752284      65.6236278
#> 58        h        50.050823      85.7979262
#> 59        i       -16.833127      -1.6113316
#> 60        j       137.323757     -85.9440646
#> 61        a       -71.761617     -27.6279080
#> 62        b      -132.728916      11.7537075
#> 63        c      -170.471952      60.8371719
#> 64        d       -16.959186     -35.9898126
#> 65        e        78.372673      83.5841210
#> 66        f      -148.702625       0.3463618
#> 67        g       121.230412     -83.4827420
#> 68        h       -78.635626     -89.8450385
#> 69        i        17.076737     -58.6589477
#> 70        j      -160.969737      27.0021697
#> 71        a       -75.518071     -27.1958289
#> 72        b      -164.619428     -10.6544738
#> 73        c       155.430075      63.6434647
#> 74        d        28.192897      38.4423983
#> 75        e       140.735322     -79.0585257
#> 76        f        59.667215      69.8276373
#> 77        g      -136.262829     -13.5709391
#> 78        h       166.545543      89.1282951
#> 79        i       -60.501695      17.9238885
#> 80        j      -136.976990     -21.8277051
#> 81        a       -31.477259       6.3950900
#> 82        b       125.104939      23.7844771
#> 83        c       -43.282137     -26.3274336
#> 84        d       118.443592      -9.6715889
#> 85        e       149.369699     -19.4092302
#> 86        f       -22.319910     -28.4394060
#> 87        g       100.412196     -43.9109107
#> 88        h        75.981483      51.5419611
#> 89        i      -122.246277     -20.3134588
#> 90        j      -179.931615     -68.8216105
#> 91        a        25.814490     -88.1804989
#> 92        b       -57.167124     -56.1821252
#> 93        c        -6.619289      78.4333415
#> 94        d        13.638119      35.0976204
#> 95        e        59.725772      73.3399986
#> 96        f      -116.702997     -22.9246048
#> 97        g        10.826452      25.8140060
#> 98        h        80.608198      66.9484571
#> 99        i       -33.505688     -66.7910743
#> 100       j        32.961723      88.1299990
cc_equ(x, value = "flagged")
#> Testing equal lat/lon
#> Flagged 0 records.
#>   [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [16] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [31] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [46] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [61] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [76] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [91] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE