Identify Records with Identical lat/lon

Removes or flags records with equal latitude and longitude coordinates, either exact or absolute. Equal coordinates can often indicate data entry errors.

cc_equ(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  test = "absolute",
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

test

character string. Defines if coordinates are compared exactly (“identical”) or on the absolute scale (i.e. -1 = 1, “absolute”). Default is to “absolute”.

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = runif(100, -180, 180), 
                decimalLatitude = runif(100, -90,90))

cc_equ(x)
#> Testing equal lat/lon
#> Removed 0 records.
#>     species decimalLongitude decimalLatitude
#> 1         a      -170.914329       20.314134
#> 2         b        -5.865779      -33.168120
#> 3         c        92.095937       26.296219
#> 4         d       -65.564197        2.597345
#> 5         e      -135.871283      -70.264061
#> 6         f      -165.570930       70.081548
#> 7         g       140.279684        4.665817
#> 8         h        39.202921      -20.923978
#> 9         i       139.920535       65.734680
#> 10        j       -10.790137       25.908307
#> 11        a      -147.996578      -65.544432
#> 12        b        30.250760      -40.251407
#> 13        c        59.374095       35.391548
#> 14        d      -154.335114       25.800202
#> 15        e      -174.531938      -42.843698
#> 16        f        85.741481      -69.158848
#> 17        g      -145.933767       49.086220
#> 18        h        96.726386       10.241764
#> 19        i      -117.337206       -4.935083
#> 20        j        98.820137       35.610103
#> 21        a        -7.864723       48.102306
#> 22        b        57.773431       61.776235
#> 23        c        57.959811       47.894799
#> 24        d       -27.376205      -60.320877
#> 25        e        12.879760       44.896221
#> 26        f       176.505362      -19.704844
#> 27        g        72.648320       61.291218
#> 28        h       -63.061037      -41.125158
#> 29        i       -31.620980       -4.337077
#> 30        j        93.128818      -46.323183
#> 31        a       -10.794635       87.287762
#> 32        b        68.582206      -89.257009
#> 33        c        31.982941       47.438354
#> 34        d       137.656931      -19.895303
#> 35        e       -76.298891       33.298778
#> 36        f       103.719533       60.076043
#> 37        g        69.682044       12.998193
#> 38        h        91.424333      -37.772588
#> 39        i       110.292139      -84.873407
#> 40        j       173.659014       86.360968
#> 41        a       116.337390      -32.461841
#> 42        b       158.148662       30.716337
#> 43        c       142.197744      -16.338451
#> 44        d        65.376169      -12.817737
#> 45        e       102.821003      -17.517848
#> 46        f      -113.177139      -10.775776
#> 47        g       -32.187383       20.575544
#> 48        h       146.667439       86.482062
#> 49        i       -42.313094      -47.328401
#> 50        j       132.155417      -65.418496
#> 51        a       -32.516897      -61.664384
#> 52        b      -137.727064       68.268362
#> 53        c       -57.406569       65.078534
#> 54        d       -81.860973       32.649620
#> 55        e        39.395605        6.966558
#> 56        f       156.276102       83.673751
#> 57        g        92.679943       32.041544
#> 58        h       117.262381      -35.887421
#> 59        i       -17.036910       47.712938
#> 60        j        76.768809      -15.354107
#> 61        a      -110.947993      -67.656164
#> 62        b        16.327885      -19.318702
#> 63        c       119.495442       -5.977033
#> 64        d       -20.416168      -49.673614
#> 65        e       -70.706522      -65.442358
#> 66        f        36.424399       24.081767
#> 67        g       145.070960       17.729742
#> 68        h      -149.059675       52.813728
#> 69        i        90.167349      -30.282617
#> 70        j      -149.191528      -66.945877
#> 71        a       -52.235835      -63.289786
#> 72        b       102.613411      -11.992849
#> 73        c        44.141046        3.402691
#> 74        d      -148.093220      -18.703800
#> 75        e      -162.169980       12.246167
#> 76        f       -71.496141       62.288736
#> 77        g      -118.376929       38.141954
#> 78        h       101.317017       74.850300
#> 79        i        28.937891       38.169834
#> 80        j       144.288270      -39.060300
#> 81        a       162.214324        8.145668
#> 82        b        19.312511       20.178911
#> 83        c       -73.960892       20.995970
#> 84        d        -4.717733       40.884609
#> 85        e      -141.518125        2.102226
#> 86        f       162.218262       48.021711
#> 87        g        77.949923        4.513444
#> 88        h      -120.744118      -16.064868
#> 89        i      -145.630249       39.170197
#> 90        j      -117.540845       -6.113687
#> 91        a       104.471930       11.683776
#> 92        b       -60.061722       85.759938
#> 93        c        95.403095      -87.567383
#> 94        d       108.516405       21.333058
#> 95        e        61.954664       83.294891
#> 96        f       167.848319       -3.733921
#> 97        g      -113.979021      -72.901324
#> 98        h       -67.567136       -7.524197
#> 99        i       -43.449123      -51.679457
#> 100       j       -78.663616       -3.627533
cc_equ(x, value = "flagged")
#> Testing equal lat/lon
#> Flagged 0 records.
#>   [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [16] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [31] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [46] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [61] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [76] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [91] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE