Identify Records with Identical lat/lon

Removes or flags records with equal latitude and longitude coordinates, either exact or absolute. Equal coordinates can often indicate data entry errors.

cc_equ(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  test = "absolute",
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

test

character string. Defines if coordinates are compared exactly (“identical”) or on the absolute scale (i.e. -1 = 1, “absolute”). Default is to “absolute”.

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = runif(100, -180, 180), 
                decimalLatitude = runif(100, -90,90))

cc_equ(x)
#> Testing equal lat/lon
#> Removed 0 records.
#>     species decimalLongitude decimalLatitude
#> 1         a       -49.353993       33.843599
#> 2         b        55.039798       75.415094
#> 3         c       154.319800       76.879884
#> 4         d       160.331452       13.519849
#> 5         e       -76.211289      -27.676687
#> 6         f       100.642114       71.604536
#> 7         g       158.063867      -46.945302
#> 8         h       -94.931383       44.149357
#> 9         i       101.198091      -78.654503
#> 10        j       122.482210       80.376346
#> 11        a       -56.363044       65.826243
#> 12        b        98.679914      -61.886513
#> 13        c       150.989568       29.086364
#> 14        d       -27.324109       45.938814
#> 15        e        56.102347       52.755404
#> 16        f       154.332149      -56.890227
#> 17        g        80.731586       40.470439
#> 18        h       171.338074      -61.830713
#> 19        i       104.868592       76.821231
#> 20        j        -2.515202      -31.203622
#> 21        a        67.470749       85.216090
#> 22        b      -127.228539       30.770821
#> 23        c       126.560540      -15.220105
#> 24        d        29.010161      -43.648187
#> 25        e        97.843563       33.456519
#> 26        f      -140.168980       48.211279
#> 27        g      -110.047894       30.884270
#> 28        h        22.848140      -63.729545
#> 29        i       150.762206       -7.940610
#> 30        j       -34.864861        7.248514
#> 31        a       -74.136061       -2.957316
#> 32        b       -18.937786      -19.972901
#> 33        c        71.415874       85.466365
#> 34        d       -78.906360        1.923973
#> 35        e       -52.401704      -24.682882
#> 36        f      -172.745890      -37.824887
#> 37        g       166.641017      -20.062082
#> 38        h        41.689752      -30.055175
#> 39        i       -27.680369       18.247006
#> 40        j      -143.460621       42.217380
#> 41        a       -48.040266       27.533546
#> 42        b      -145.066999      -57.179298
#> 43        c       -97.438696        3.687479
#> 44        d        41.025304       11.079336
#> 45        e      -151.208567       15.021800
#> 46        f      -131.172584      -12.787439
#> 47        g       -82.081566       62.084458
#> 48        h        -5.095919       45.280676
#> 49        i       -12.138551      -19.456765
#> 50        j        76.455004      -39.619679
#> 51        a       115.569051      -66.329484
#> 52        b       104.237901      -28.784753
#> 53        c      -127.517588        3.599802
#> 54        d        -4.140729      -20.660010
#> 55        e      -154.290403      -17.911205
#> 56        f       120.017541      -34.412117
#> 57        g      -151.750736        8.587217
#> 58        h       -10.238200      -48.527285
#> 59        i      -110.199503       45.590379
#> 60        j       -93.656604      -73.142059
#> 61        a        25.313742        9.085066
#> 62        b       165.645964      -78.283084
#> 63        c       144.794312      -50.948305
#> 64        d        -8.340784      -53.292351
#> 65        e        47.535664      -88.886427
#> 66        f        -5.926929      -51.070777
#> 67        g      -166.072593       67.048826
#> 68        h      -159.410307        4.896518
#> 69        i        76.276213      -11.344972
#> 70        j      -135.342114      -28.049721
#> 71        a      -162.791522       41.908533
#> 72        b      -139.745871      -71.381840
#> 73        c       -30.679374       86.643663
#> 74        d      -128.605616      -42.162393
#> 75        e       -58.357262       77.012913
#> 76        f      -156.976607       50.719259
#> 77        g        56.773758      -55.576667
#> 78        h         9.699518      -59.684491
#> 79        i       110.661022       60.169913
#> 80        j        63.887353       75.821450
#> 81        a       -64.874182       35.098270
#> 82        b       106.810347       52.845938
#> 83        c       -80.210423       35.978937
#> 84        d        71.068013      -82.272476
#> 85        e        37.344177        4.520920
#> 86        f      -158.397525       24.454698
#> 87        g       -16.677685       21.679892
#> 88        h       -84.393214       76.477448
#> 89        i        69.737601        4.136685
#> 90        j      -138.015032       35.143491
#> 91        a      -106.908483      -27.191729
#> 92        b       121.605668       14.112729
#> 93        c        19.561757       37.416381
#> 94        d        88.039110       16.074821
#> 95        e        86.396019       44.532513
#> 96        f        69.597634       48.268666
#> 97        g        15.804005      -25.137584
#> 98        h      -151.548829       62.031158
#> 99        i        42.628228      -67.728520
#> 100       j      -179.662695       88.635309
cc_equ(x, value = "flagged")
#> Testing equal lat/lon
#> Flagged 0 records.
#>   [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [16] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [31] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [46] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [61] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [76] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [91] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE