Identify Records with Identical lat/lon

Removes or flags records with equal latitude and longitude coordinates, either exact or absolute. Equal coordinates can often indicate data entry errors.

cc_equ(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  test = "absolute",
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

test

character string. Defines if coordinates are compared exactly (“identical”) or on the absolute scale (i.e. -1 = 1, “absolute”). Default is to “absolute”.

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = runif(100, -180, 180), 
                decimalLatitude = runif(100, -90,90))

cc_equ(x)
#> Testing equal lat/lon
#> Removed 0 records.
#>     species decimalLongitude decimalLatitude
#> 1         a       118.101237      13.2416072
#> 2         b         7.339973       5.7871603
#> 3         c      -125.554462       1.8948065
#> 4         d       175.350915      12.5884784
#> 5         e      -115.032980     -46.8809103
#> 6         f       -88.391457     -71.8604320
#> 7         g       136.632326      70.8779520
#> 8         h       -76.714125      34.9441540
#> 9         i      -103.338787      48.1650459
#> 10        j       -99.003498      40.7286072
#> 11        a         6.509617     -23.1340333
#> 12        b       -24.003959      47.1655490
#> 13        c       -63.342677      10.5021904
#> 14        d       142.551610      33.7939544
#> 15        e        22.131448      60.7690679
#> 16        f       -80.184497       2.9927587
#> 17        g       -23.454728     -43.0400768
#> 18        h       131.507630      65.7694183
#> 19        i      -171.968876      62.1557411
#> 20        j         6.088380     -10.8158129
#> 21        a       172.812205     -36.9192058
#> 22        b       134.975225      22.9992920
#> 23        c        11.158100     -58.4950040
#> 24        d         8.205268     -40.4283899
#> 25        e       149.801643      43.4207043
#> 26        f        -9.681953      82.0473136
#> 27        g       -81.181052      20.4064303
#> 28        h       172.381877     -68.9866778
#> 29        i       -16.419075      -9.3880524
#> 30        j        -2.186880     -66.7272716
#> 31        a        35.473468      42.9465088
#> 32        b       119.449215      42.3890092
#> 33        c        58.132786     -52.9683599
#> 34        d       115.099601     -36.8871950
#> 35        e       165.115735      65.7664215
#> 36        f      -141.575127     -33.4085014
#> 37        g        14.325212      21.8230176
#> 38        h       -36.401809      67.2123579
#> 39        i        77.818966      29.6001490
#> 40        j      -140.497818     -31.9196902
#> 41        a        61.641244     -20.4815139
#> 42        b       130.024333     -70.8187039
#> 43        c      -123.891914      20.4711209
#> 44        d      -106.672857      34.1228433
#> 45        e      -155.640026      33.8471763
#> 46        f      -145.724442      74.7200466
#> 47        g      -120.864237      38.2847951
#> 48        h        34.746079      42.4722076
#> 49        i      -152.482925      -5.1015307
#> 50        j        -7.017957      18.9598345
#> 51        a        48.605946       9.4231586
#> 52        b        -7.343421      79.5743276
#> 53        c        91.679455     -65.5606679
#> 54        d       175.755893      25.3306471
#> 55        e        18.688980     -88.2987082
#> 56        f       -33.272805      87.4408331
#> 57        g        72.471192      48.3434636
#> 58        h        86.865687      37.8010893
#> 59        i      -140.077625     -56.9778699
#> 60        j       -13.284168     -81.2921675
#> 61        a        61.534423      77.2193697
#> 62        b      -179.438103     -76.0056616
#> 63        c       -92.941334      11.8938196
#> 64        d       112.382624       3.1797070
#> 65        e      -124.944121     -61.0337401
#> 66        f      -152.556069      22.2988734
#> 67        g       -38.676276      29.6668952
#> 68        h       -60.395754     -29.8831246
#> 69        i        44.375141      46.3855853
#> 70        j       -68.541612     -32.2126380
#> 71        a        -8.188345     -66.3621686
#> 72        b       -95.432647     -34.1233428
#> 73        c       -62.201635     -64.1490272
#> 74        d       -99.307092     -46.9239901
#> 75        e       152.171712      73.0230519
#> 76        f       -10.384020      12.1913201
#> 77        g       -54.624503     -38.2518036
#> 78        h      -120.223821     -76.3045292
#> 79        i       -65.768156      42.9978841
#> 80        j      -147.541212      77.5066458
#> 81        a       116.176584      45.7241954
#> 82        b       -16.864578     -26.2431694
#> 83        c       -46.356353      47.7994698
#> 84        d      -146.869084      11.7715501
#> 85        e       121.221195      12.8624313
#> 86        f       135.841468      68.3357141
#> 87        g      -146.191518      63.9550576
#> 88        h      -161.717935     -81.8305101
#> 89        i        46.370065     -43.8187665
#> 90        j       -46.780288      -3.4629600
#> 91        a      -116.116215      82.9303477
#> 92        b       151.101783     -69.6925954
#> 93        c        57.450613      40.9843931
#> 94        d        38.289927      35.5849995
#> 95        e        45.259038      26.5348813
#> 96        f       -95.469844      58.1950404
#> 97        g       -61.240961     -16.0662440
#> 98        h      -172.128958      -0.7320322
#> 99        i       -40.194722      81.9309643
#> 100       j       -14.561127      53.0961414
cc_equ(x, value = "flagged")
#> Testing equal lat/lon
#> Flagged 0 records.
#>   [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [16] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [31] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [46] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [61] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [76] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [91] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE