Identify Records with Identical lat/lon

Removes or flags records with equal latitude and longitude coordinates, either exact or absolute. Equal coordinates can often indicate data entry errors.

cc_equ(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  test = "absolute",
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

test

character string. Defines if coordinates are compared exactly (“identical”) or on the absolute scale (i.e. -1 = 1, “absolute”). Default is to “absolute”.

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = runif(100, -180, 180), 
                decimalLatitude = runif(100, -90,90))

cc_equ(x)
#> Testing equal lat/lon
#> Removed 0 records.
#>     species decimalLongitude decimalLatitude
#> 1         a      -137.134515      70.6161888
#> 2         b       120.275773     -15.9838791
#> 3         c        13.671534      88.2783105
#> 4         d        56.963092     -65.4334120
#> 5         e        66.287994      61.4524768
#> 6         f       106.306862      86.3075691
#> 7         g        -7.492576     -58.5345897
#> 8         h       143.136303       8.0265326
#> 9         i        45.932599     -81.2751428
#> 10        j      -127.444711     -36.7260947
#> 11        a       178.951225     -58.9832174
#> 12        b       112.323153       0.7174794
#> 13        c       138.462190      78.3067949
#> 14        d      -177.442150      12.5368211
#> 15        e        96.764122       0.0742663
#> 16        f       135.080346     -86.6743002
#> 17        g       167.179141      89.3928894
#> 18        h       -78.894920      74.2216810
#> 19        i        15.688571      74.2471867
#> 20        j       138.722813      33.4469157
#> 21        a       108.129789     -32.4051277
#> 22        b       132.130352      75.5954715
#> 23        c      -153.288429      86.5692351
#> 24        d       138.327362     -44.8839931
#> 25        e       -41.895959     -66.1134888
#> 26        f      -114.703241      16.8767608
#> 27        g       101.182700      47.1102005
#> 28        h        26.928365     -32.7107363
#> 29        i       175.817366      78.4888148
#> 30        j        30.485523     -41.5140539
#> 31        a        -3.599508      64.4908327
#> 32        b        33.585688     -75.3102883
#> 33        c        15.681043     -37.5394879
#> 34        d      -132.350022      11.4319062
#> 35        e       -53.173668      26.2277761
#> 36        f        20.673589      14.2032751
#> 37        g      -136.014378      51.9801978
#> 38        h       -65.959665      53.1546435
#> 39        i       108.354478      39.0861345
#> 40        j       -41.491459     -72.0766020
#> 41        a        68.920380     -58.8621190
#> 42        b       -30.919377       3.1526509
#> 43        c        -5.638690      87.8581824
#> 44        d       -36.259036     -63.4968473
#> 45        e      -133.922789      78.1840775
#> 46        f       -95.624682     -35.7182062
#> 47        g       125.891889      -2.3633052
#> 48        h       -88.716672      78.1023184
#> 49        i        17.005070      78.2750407
#> 50        j       173.645316      22.7490478
#> 51        a      -164.400259      11.4727209
#> 52        b        20.409740      28.1581158
#> 53        c        60.925822     -32.1119869
#> 54        d      -177.696131     -63.7649848
#> 55        e       172.818059      -0.4341991
#> 56        f       -89.809935     -49.0952678
#> 57        g      -106.421206      49.1244367
#> 58        h      -106.593850     -18.5734501
#> 59        i        31.782071      70.4939913
#> 60        j       -21.454420      86.6194817
#> 61        a       159.758938     -58.4883754
#> 62        b      -160.715717     -17.7554208
#> 63        c      -128.660988     -19.0903090
#> 64        d       -81.141170      10.4475476
#> 65        e       -13.295113     -66.8011481
#> 66        f       128.519428     -52.2017856
#> 67        g      -103.796183     -78.9602505
#> 68        h      -115.551107      76.5275749
#> 69        i       138.306473      20.9762637
#> 70        j        85.160106     -41.3069278
#> 71        a       112.753244      79.1673902
#> 72        b      -134.300310      74.3237910
#> 73        c       143.282544       2.4830569
#> 74        d       -87.855123     -52.1667853
#> 75        e       -68.040170       5.6211141
#> 76        f       -32.881125     -81.7482272
#> 77        g       -10.887077      82.0212732
#> 78        h      -113.238275      55.7263323
#> 79        i        53.500625      27.4202444
#> 80        j       141.303919      61.1563307
#> 81        a        27.871296      51.1926794
#> 82        b       -18.981273      73.8352816
#> 83        c       -71.677089     -59.6311330
#> 84        d       147.524307     -20.0525403
#> 85        e       121.541131     -15.6351767
#> 86        f        91.623864      46.1678302
#> 87        g       -92.686066     -45.7230174
#> 88        h         1.437979      64.6085913
#> 89        i      -160.815058      66.8761489
#> 90        j         4.889433      76.0118802
#> 91        a       -94.371619       0.5243623
#> 92        b       158.972723      30.6141490
#> 93        c      -105.860925       6.6344837
#> 94        d       -31.566638     -10.0377673
#> 95        e       -10.688943     -41.2484096
#> 96        f       171.955579     -15.9020581
#> 97        g       -12.284664      19.9484540
#> 98        h       -40.373452       7.0247201
#> 99        i      -167.681052      -7.3736021
#> 100       j        11.170553     -14.2067727
cc_equ(x, value = "flagged")
#> Testing equal lat/lon
#> Flagged 0 records.
#>   [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [16] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [31] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [46] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [61] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [76] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [91] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE