Identify Records with Identical lat/lon

Removes or flags records with equal latitude and longitude coordinates, either exact or absolute. Equal coordinates can often indicate data entry errors.

cc_equ(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  test = "absolute",
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

test

character string. Defines if coordinates are compared exactly (“identical”) or on the absolute scale (i.e. -1 = 1, “absolute”). Default is to “absolute”.

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = runif(100, -180, 180), 
                decimalLatitude = runif(100, -90,90))

cc_equ(x)
#> Testing equal lat/lon
#> Removed 0 records.
#>     species decimalLongitude decimalLatitude
#> 1         a     -154.4434414      -68.624477
#> 2         b      146.7415011      -29.715366
#> 3         c      124.6513824       16.865389
#> 4         d      -84.1847915      -26.390948
#> 5         e      -73.6902048       21.823520
#> 6         f     -124.8306224       89.455540
#> 7         g       46.5217789      -89.040464
#> 8         h      -30.6587591       54.788179
#> 9         i      117.0342762       37.377330
#> 10        j     -147.8194687       53.569475
#> 11        a      167.9320325       24.788254
#> 12        b      170.3868202       77.077833
#> 13        c        0.5996425       74.505626
#> 14        d      -12.6431271      -10.609744
#> 15        e        5.6668429       80.268793
#> 16        f      -15.6884450       64.687202
#> 17        g       10.0990159      -85.321401
#> 18        h     -146.6272809      -24.685752
#> 19        i     -104.2166551       89.531157
#> 20        j     -125.7985030       -5.809713
#> 21        a     -142.0945754      -69.608782
#> 22        b      153.4697542       37.837861
#> 23        c      -42.0397230       31.045619
#> 24        d      -28.4195076      -36.950465
#> 25        e       16.9222997      -11.967028
#> 26        f       -0.2991997       54.189878
#> 27        g      -36.9763905      -34.445979
#> 28        h       59.4389972      -33.829392
#> 29        i     -167.1921531      -86.667740
#> 30        j      -12.1235011       17.607000
#> 31        a      179.8170810      -23.081467
#> 32        b      100.9238352      -10.309965
#> 33        c      -40.6192093       18.600143
#> 34        d      151.0669765        7.351474
#> 35        e      171.6999597       10.705925
#> 36        f      -82.0552303        1.136661
#> 37        g     -171.3170894       -2.482544
#> 38        h      107.3870979       31.810227
#> 39        i      -49.9757809      -11.914417
#> 40        j      -18.3910948       50.181023
#> 41        a       22.0837765       -3.985085
#> 42        b      -75.7176791       80.728303
#> 43        c       79.2654445      -54.524744
#> 44        d     -117.1575344       50.517748
#> 45        e      145.2097742       31.046385
#> 46        f      -55.2988806       34.258374
#> 47        g       36.8537102      -79.305077
#> 48        h      -29.2609970      -12.159065
#> 49        i      -17.6419525      -47.290267
#> 50        j      166.4881367       44.088186
#> 51        a        4.6440192       42.651321
#> 52        b      154.9645509       61.186496
#> 53        c      160.5354870       34.369716
#> 54        d      -50.4165095       34.832181
#> 55        e      -95.5320693      -47.830685
#> 56        f      -24.1274017       78.346193
#> 57        g      144.7466062      -67.204758
#> 58        h      153.6186612       22.564333
#> 59        i      100.3164476       -2.306239
#> 60        j       30.8895496       71.102147
#> 61        a     -126.1128665        4.244100
#> 62        b      157.4882572      -76.451117
#> 63        c     -177.5230351      -28.198160
#> 64        d      -90.0061164      -68.740981
#> 65        e     -133.0406708      -40.537644
#> 66        f      -61.1717427       -5.103734
#> 67        g      -83.9118083       62.074262
#> 68        h      167.8189678      -84.984186
#> 69        i      -76.7055963      -86.306439
#> 70        j      -92.7955957       18.004001
#> 71        a      -87.1646013       50.006442
#> 72        b     -101.1618624       -2.070714
#> 73        c      -50.1171404      -50.148909
#> 74        d      115.4165535       88.141994
#> 75        e       97.0817642       30.491677
#> 76        f      -31.1455616       60.028405
#> 77        g     -158.7919176      -82.189434
#> 78        h       11.5658661      -37.812492
#> 79        i     -113.6160977      -51.825504
#> 80        j      149.7010917       69.054895
#> 81        a      -62.5796813       28.870014
#> 82        b      140.2662592      -68.527979
#> 83        c      144.8120418      -35.220669
#> 84        d      -88.4886762      -74.566385
#> 85        e       59.9261027       40.589487
#> 86        f     -169.8570724      -28.360097
#> 87        g      103.9331332       33.176217
#> 88        h      -62.3591194       41.355001
#> 89        i       -6.9068556       83.218082
#> 90        j     -105.7000647       31.612191
#> 91        a     -176.3743600       12.833971
#> 92        b       82.1576194       11.485738
#> 93        c       54.2867606      -58.167019
#> 94        d      -57.7882081       53.903204
#> 95        e       -2.3537120      -49.216422
#> 96        f     -166.3826377       76.878256
#> 97        g      163.2594107       24.398285
#> 98        h      165.7098992       24.281023
#> 99        i     -179.6463409       43.230057
#> 100       j       71.3606013      -79.750122
cc_equ(x, value = "flagged")
#> Testing equal lat/lon
#> Flagged 0 records.
#>   [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [16] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [31] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [46] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [61] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [76] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [91] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE