Identify Records with Identical lat/lon

Removes or flags records with equal latitude and longitude coordinates, either exact or absolute. Equal coordinates can often indicate data entry errors.

cc_equ(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  test = "absolute",
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

test

character string. Defines if coordinates are compared exactly (“identical”) or on the absolute scale (i.e. -1 = 1, “absolute”). Default is to “absolute”.

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = runif(100, -180, 180), 
                decimalLatitude = runif(100, -90,90))

cc_equ(x)
#> Testing equal lat/lon
#> Removed 0 records.
#>     species decimalLongitude decimalLatitude
#> 1         a     -110.8507273       81.208792
#> 2         b       46.1280464       40.518869
#> 3         c      118.5750284       66.798499
#> 4         d       71.0291084       52.915868
#> 5         e      109.2990929      -59.124545
#> 6         f      109.3254984      -59.689108
#> 7         g     -128.0221715       53.681888
#> 8         h        0.5808218      -13.767728
#> 9         i        8.4651760      -38.978263
#> 10        j       78.2990080      -66.415011
#> 11        a     -153.8479496       87.732991
#> 12        b       80.0972671      -84.288637
#> 13        c       88.4783750       -2.899143
#> 14        d       67.2660494       56.921154
#> 15        e       60.8190553       60.541561
#> 16        f     -144.4158746      -41.474344
#> 17        g      -53.2318812      -72.796393
#> 18        h     -100.2631840       81.965992
#> 19        i      -86.6597671      -13.356509
#> 20        j       39.1622731      -66.181510
#> 21        a       81.3421239       19.676495
#> 22        b      -59.1284798       73.509292
#> 23        c      173.2701483       55.085606
#> 24        d      -55.2665576       -1.346190
#> 25        e       63.6651836       50.846416
#> 26        f      178.1240717       16.696238
#> 27        g      136.5242032      -35.384617
#> 28        h     -164.6470693       43.694409
#> 29        i      114.9016101      -44.315326
#> 30        j     -172.6411063      -24.996541
#> 31        a      162.0476022      -13.137087
#> 32        b     -154.5656732      -48.336579
#> 33        c       -3.4928098       33.098872
#> 34        d     -165.5415092      -86.259768
#> 35        e     -121.0424142       26.061497
#> 36        f      -14.6875966      -48.564113
#> 37        g      -51.3846040      -51.599464
#> 38        h      -86.7702372       84.466131
#> 39        i       64.2400701      -12.268588
#> 40        j      154.0636755      -70.333632
#> 41        a     -133.7294225      -65.026468
#> 42        b     -119.3266032      -39.150755
#> 43        c      -27.3316554      -78.107235
#> 44        d      177.2226743        8.132724
#> 45        e     -169.0004356      -24.631698
#> 46        f       69.7285600      -37.686060
#> 47        g       24.5903352      -40.034294
#> 48        h       17.9477217      -67.667464
#> 49        i      -92.1769685       22.212782
#> 50        j       43.8901949       -3.792957
#> 51        a       17.0098911        1.571676
#> 52        b     -126.3412157      -11.099845
#> 53        c       65.3577527      -75.392185
#> 54        d      125.8852652      -24.171920
#> 55        e      -59.8569907      -86.122065
#> 56        f     -138.8178834       15.989847
#> 57        g       84.1538976       62.688914
#> 58        h       11.7812466       73.212225
#> 59        i     -149.4955577      -76.185481
#> 60        j     -177.8351607      -44.305333
#> 61        a     -121.1537250       44.401748
#> 62        b       17.0924256       88.529503
#> 63        c       36.9747729      -71.510477
#> 64        d        7.9111751      -32.651890
#> 65        e     -152.7214542      -55.078489
#> 66        f       54.7153130       72.312219
#> 67        g       71.6206334       47.301995
#> 68        h      125.9991039      -10.980893
#> 69        i       13.2193874       29.362502
#> 70        j     -143.6218285      -52.802870
#> 71        a      154.9806302       86.534318
#> 72        b      142.7286654       82.577692
#> 73        c       99.7065517       45.811479
#> 74        d      -76.5336437       -4.717282
#> 75        e      -31.3663293      -88.810203
#> 76        f     -143.0825520       34.874556
#> 77        g       28.2868416       75.138695
#> 78        h      178.6861631      -22.532355
#> 79        i       -7.8845410       23.372369
#> 80        j     -107.0352606       42.721382
#> 81        a     -156.4523691       40.984637
#> 82        b      -40.5563186      -27.464705
#> 83        c       87.5143005      -60.210952
#> 84        d      100.2814392       41.166374
#> 85        e      -75.1477790      -43.906480
#> 86        f       96.0914794       89.603598
#> 87        g      -90.3031851      -31.052468
#> 88        h      -63.2881495      -83.937827
#> 89        i      107.8125381       86.061922
#> 90        j      142.5486298       46.903075
#> 91        a      111.0082637      -58.327512
#> 92        b      160.6917680       61.943354
#> 93        c     -127.9990954       -4.481069
#> 94        d        7.8433113      -10.085049
#> 95        e     -131.4960761       83.375890
#> 96        f       51.0717475      -74.465161
#> 97        g      116.5864790       74.731925
#> 98        h     -106.7669941      -73.347937
#> 99        i     -138.8630825       43.038222
#> 100       j       17.7527337       37.417892
cc_equ(x, value = "flagged")
#> Testing equal lat/lon
#> Flagged 0 records.
#>   [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [16] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [31] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [46] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [61] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [76] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [91] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE