Identify Records with Identical lat/lon

Removes or flags records with equal latitude and longitude coordinates, either exact or absolute. Equal coordinates can often indicate data entry errors.

cc_equ(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  test = "absolute",
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

test

character string. Defines if coordinates are compared exactly (“identical”) or on the absolute scale (i.e. -1 = 1, “absolute”). Default is to “absolute”.

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = runif(100, -180, 180), 
                decimalLatitude = runif(100, -90,90))

cc_equ(x)
#> Testing equal lat/lon
#> Removed 0 records.
#>     species decimalLongitude decimalLatitude
#> 1         a       87.5870253      -5.8997175
#> 2         b      170.5658620      81.6144062
#> 3         c       74.7004756      45.9039820
#> 4         d      -18.6133429     -61.3827914
#> 5         e      -71.9769675     -80.5048926
#> 6         f      120.5138758      81.8732313
#> 7         g       68.7495920       6.4118112
#> 8         h     -129.3272321      -1.5701454
#> 9         i       39.4598025      42.9775389
#> 10        j     -149.1331878      82.2938624
#> 11        a     -148.9050279      -1.9474387
#> 12        b      173.3596625       0.4329361
#> 13        c       83.6049222     -58.4870134
#> 14        d     -109.8295603      87.5770032
#> 15        e      137.5494313     -74.9479146
#> 16        f     -162.1896985     -25.3239824
#> 17        g       98.6580907     -13.9077271
#> 18        h     -130.2353012      28.8209275
#> 19        i      113.1923828      20.6993626
#> 20        j      130.0211036      10.9886269
#> 21        a      155.2831183      32.3442825
#> 22        b      -55.4115225     -35.4180374
#> 23        c     -179.3071495      60.7524779
#> 24        d      -67.7608152     -66.5465642
#> 25        e     -149.0079920     -37.8817277
#> 26        f      133.6357134      10.1451166
#> 27        g      -78.5855200     -78.6580705
#> 28        h      128.7493168     -76.1683804
#> 29        i     -111.6220624      72.3022993
#> 30        j     -120.5504782      30.7036025
#> 31        a       61.2331112      17.3052077
#> 32        b     -169.6642624     -40.3489644
#> 33        c      -26.6757296     -62.9159514
#> 34        d      139.2949306     -16.7554694
#> 35        e     -127.0250126     -45.1392273
#> 36        f      -60.3609959      65.5537541
#> 37        g     -132.3444237      86.5977500
#> 38        h       52.0246416     -11.8626817
#> 39        i       93.6491186      77.8662856
#> 40        j     -104.2412655       5.4807074
#> 41        a     -142.9408311      72.6083783
#> 42        b     -164.1705737     -57.7157937
#> 43        c      126.0833160     -54.1578078
#> 44        d       82.7641068      32.3049712
#> 45        e     -159.0406007      16.9138401
#> 46        f     -156.8717479     -47.6744900
#> 47        g      -96.5112452     -64.1207545
#> 48        h       50.3060374     -49.4620453
#> 49        i      -24.5593492      29.9464066
#> 50        j      120.0663137     -59.7421242
#> 51        a      140.1463969      -5.1475256
#> 52        b       25.3180129     -30.4691850
#> 53        c     -165.9723452      32.6234661
#> 54        d     -128.5802528     -53.2383832
#> 55        e      127.4029136      66.0097327
#> 56        f       79.8682008      26.4696444
#> 57        g      -66.2973526      81.8056579
#> 58        h     -165.6914504      70.2042160
#> 59        i      -26.0239170      61.0281611
#> 60        j       -8.6208151      36.0417122
#> 61        a      -72.9266553     -40.0781174
#> 62        b       40.2984333      55.5963856
#> 63        c      174.4252348     -74.9306575
#> 64        d      -42.8295073      -7.1790956
#> 65        e     -147.3631921     -45.5970415
#> 66        f       92.9317771     -66.9917093
#> 67        g     -171.7958149      20.6742823
#> 68        h      -79.5896236     -39.5459561
#> 69        i       23.7017539     -87.7203621
#> 70        j      -82.4983760     -80.7862070
#> 71        a       87.7080708     -20.9720570
#> 72        b     -147.3883681     -26.0502991
#> 73        c      114.2334812      41.8463947
#> 74        d       90.0041169       9.8366779
#> 75        e     -108.3015127      25.3915848
#> 76        f       88.1919934      12.5921272
#> 77        g      135.3798611      65.5816402
#> 78        h      -93.9979173     -58.2480787
#> 79        i       68.4599512     -20.6198047
#> 80        j      152.0692275     -26.8084807
#> 81        a      -89.6246459      58.2434775
#> 82        b        1.4774030      88.0720052
#> 83        c       53.2586730       9.6262843
#> 84        d        0.7949514      31.2861351
#> 85        e     -117.8265810      81.6717723
#> 86        f      171.2744120      79.5067888
#> 87        g      -64.7964743      20.8973625
#> 88        h      156.7559276     -53.4696082
#> 89        i      103.7288446     -22.6412108
#> 90        j     -127.2698704     -49.1207287
#> 91        a      -46.5171995      36.1867218
#> 92        b     -123.9467938      26.3655519
#> 93        c     -179.9730701     -28.9687487
#> 94        d      134.0635227     -56.1829005
#> 95        e      125.4197766      -2.3080969
#> 96        f       99.9849815      22.9576491
#> 97        g       36.4923844      62.0286286
#> 98        h      -97.9334346      -1.8000688
#> 99        i      -56.2082818     -18.3362701
#> 100       j       26.5494350      60.7003842
cc_equ(x, value = "flagged")
#> Testing equal lat/lon
#> Flagged 0 records.
#>   [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [16] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [31] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [46] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [61] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [76] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [91] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE