Flag Records in the Vicinity of Biodiversity Institutions

Flag records assigned to the location of zoos, botanical gardens, herbaria, universities and museums, based on a global database of ~10,000 such biodiversity institutions. Coordinates from these locations can be related to data-entry errors, false automated geo-reference or individuals in captivity/horticulture.

cc_inst(x, lon = "decimallongitude", lat = "decimallatitude",
  buffer = 100, geod = TRUE, ref = NULL, value = "clean",
  verbose = TRUE)

Arguments

x
a data.frame. Containing geographical coordinates and species names.

lon
a character string. The column with the longitude coordinates. Default = “decimallongitude”.

lat
a character string. The column with the longitude coordinates. Default = “decimallatitude”.

buffer
numerical. The buffer around each province or country centroid, where records should be flagged as problematic, in decimal degrees. Default = 100m.

geod
logical. If TRUE the radius around each institution is calculated based on a sphere, buffer is in meters and independent of latitude. If FALSE the radius is calculated assuming planar coordinates and varies slightly with latitude, in this case buffer is in degrees. DEfault = T.

ref
a SpatialPointsDataframe. Providing the geographic gazetteer. Can be any SpatialPointsDataframe, but the structure must be identical to institutions. Default = institutions

value
a character string. Defining the output value. See value.

verbose
logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic. Default = “clean”.

Details

Note: the buffer radius is in degrees, thus will differ slightly between different latitudes.

Note

See https://azizka.github.io/CoordinateCleaner for more details and tutorials.

See also

Examples

x <- data.frame(species = letters[1:10],
                decimallongitude = runif(100, -180, 180),
                decimallatitude = runif(100, -90,90))

cc_inst(x, value = "flagged", buffer = 500000)#large buffer for demonstration

#> Testing biodiversity institutions

#> Flagged 36 records.

#>     1     2     3     4     5     6     7     8     9    10    11    12    13 
#>  TRUE  TRUE FALSE  TRUE  TRUE  TRUE  TRUE FALSE  TRUE  TRUE  TRUE  TRUE  TRUE 
#>    14    15    16    17    18    19    20    21    22    23    24    25    26 
#>  TRUE  TRUE  TRUE  TRUE FALSE FALSE  TRUE FALSE  TRUE FALSE  TRUE  TRUE FALSE 
#>    27    28    29    30    31    32    33    34    35    36    37    38    39 
#> FALSE FALSE  TRUE  TRUE FALSE FALSE FALSE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE 
#>    40    41    42    43    44    45    46    47    48    49    50    51    52 
#> FALSE FALSE  TRUE  TRUE FALSE  TRUE FALSE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE 
#>    53    54    55    56    57    58    59    60    61    62    63    64    65 
#> FALSE  TRUE  TRUE  TRUE FALSE  TRUE  TRUE  TRUE FALSE FALSE FALSE  TRUE FALSE 
#>    66    67    68    69    70    71    72    73    74    75    76    77    78 
#>  TRUE  TRUE FALSE  TRUE  TRUE FALSE  TRUE  TRUE  TRUE FALSE FALSE  TRUE  TRUE 
#>    79    80    81    82    83    84    85    86    87    88    89    90    91 
#>  TRUE  TRUE  TRUE  TRUE FALSE  TRUE  TRUE FALSE  TRUE FALSE FALSE FALSE FALSE 
#>    92    93    94    95    96    97    98    99   100 
#>  TRUE  TRUE FALSE  TRUE  TRUE FALSE  TRUE  TRUE  TRUE