Identify Geographic Outliers in Species Distributions

Removes out or flags records that are outliers in geographic space according to the method defined via the method argument. Geographic outliers often represent erroneous coordinates, for example due to data entry errors, imprecise geo-references, individuals in horticulture/captivity.

cc_outl(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  species = "species",
  method = "quantile",
  mltpl = 5,
  tdi = 1000,
  value = "clean",
  sampling_thresh = 0,
  verbose = TRUE,
  min_occs = 7,
  thinning = FALSE,
  thinning_res = 0.5
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

species

character string. The column with the species name. Default = “species”.

method

character string. Defining the method for outlier selection. See details. One of “distance”, “quantile”, “mad”. Default = “quantile”.

mltpl

numeric. The multiplier of the interquartile range (method == 'quantile') or median absolute deviation (method == 'mad')to identify outliers. See details. Default = 5.

tdi

numeric. The minimum absolute distance (method == 'distance') of a record to all other records of a species to be identified as outlier, in km. See details. Default = 1000.

value

character string. Defining the output value. See value.

sampling_thresh

numeric. Cut off threshold for the sampling correction. Indicates the quantile of sampling in which outliers should be ignored. For instance, if sampling_thresh == 0.25, records in the 25 (no sampling correction).

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

min_occs

Minimum number of geographically unique datapoints needed for a species to be tested. This is necessary for reliable outlier estimation. Species with fewer than min_occs records will not be tested and the output value will be 'TRUE'. Default is to 7. If method == 'distance', consider a lower threshold.

thinning

forces a raster approximation for the distance calculation. This is routinely used for species with more than 10,000 records for computational reasons, but can be enforced for smaller datasets, which is recommended when sampling is very uneven.

thinning_res

The resolution for the spatial thinning in decimal degrees. Default = 0.5.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Details

The method for outlier identification depends on the method argument. If “quantile”: a boxplot method is used and records are flagged as outliers if their mean distance to all other records of the same species is larger than mltpl * the interquartile range of the mean distance of all records of this species. If “mad”: the median absolute deviation is used. In this case a record is flagged as outlier, if the mean distance to all other records of the same species is larger than the median of the mean distance of all points plus/minus the mad of the mean distances of all records of the species * mltpl. If “distance”: records are flagged as outliers, if the minimum distance to the next record of the species is > tdi. For species with records from > 10000 unique locations a random sample of 1000 records is used for the distance matrix calculation. The test skips species with fewer than min_occs, geographically unique records.

The likelihood of occurrence records being erroneous outliers is linked to the sampling effort in any given location. To account for this, the sampling_cor option fetches the number of occurrence records available from www.gbif.org, per country as a proxy of sampling effort. The outlier test (the mean distance) for each records is than weighted by the log transformed number of records per square kilometre in this country. See for https://besjournals.onlinelibrary.wiley.com/doi/full/10.1111/2041-210X.13152 an example and further explanation of the outlier test.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

See also

Examples

x <- data.frame(species = letters[1:10],
                decimalLongitude = runif(100, -180, 180),
                decimalLatitude = runif(100, -90,90))

cc_outl(x)
#> Testing geographic outliers
#> Removed 0 records.
#>     species decimalLongitude decimalLatitude
#> 1         a        -8.230974      48.7642668
#> 2         b         4.612023      -4.7377379
#> 3         c       -51.358622      46.2159614
#> 4         d        75.105655      46.1287029
#> 5         e       -30.079838     -78.1868978
#> 6         f       -36.611195      49.0670829
#> 7         g       -56.292351      48.4129285
#> 8         h       137.167025     -18.4734905
#> 9         i         4.537803     -68.5561443
#> 10        j       -79.518388      -7.3408701
#> 11        a       154.801149      82.1186186
#> 12        b      -158.173676     -85.1740926
#> 13        c       126.180196     -56.9526648
#> 14        d        75.611047      53.2221060
#> 15        e       -69.266879     -10.4790212
#> 16        f        20.558099     -61.3437205
#> 17        g       125.409125     -24.4374830
#> 18        h        91.064979     -73.2680543
#> 19        i        66.970221      71.5161119
#> 20        j       -19.705859      76.6349830
#> 21        a        84.949363      77.1295472
#> 22        b      -153.958400      62.3476728
#> 23        c        26.073241     -86.6822082
#> 24        d      -155.030985     -16.5248553
#> 25        e       -15.514142      -5.2133869
#> 26        f      -115.389807     -71.4231095
#> 27        g       110.674567     -87.2044911
#> 28        h       171.476649     -83.7874458
#> 29        i       -49.938124     -27.0558878
#> 30        j       -89.636079      36.5615360
#> 31        a       135.011715     -32.9373510
#> 32        b       -92.585476       4.7496801
#> 33        c        88.276676     -86.7935873
#> 34        d        21.895710      -7.3517153
#> 35        e       122.043861     -34.8273298
#> 36        f        49.048115     -13.7405318
#> 37        g      -170.673810      64.6616091
#> 38        h        57.427655      21.3400372
#> 39        i        99.782955      45.3443741
#> 40        j      -122.040649       6.1580299
#> 41        a      -162.630842     -13.3641816
#> 42        b       114.734256     -63.5755617
#> 43        c       -89.764045      86.0048503
#> 44        d       121.467209       6.1445182
#> 45        e         8.134242     -65.0570907
#> 46        f      -174.465996      30.2943364
#> 47        g       126.740299      56.1612229
#> 48        h      -156.447955      -0.0836713
#> 49        i      -152.583839      -4.8843788
#> 50        j       113.324557      87.7355928
#> 51        a       110.006148      65.1985549
#> 52        b        29.154302      58.4088169
#> 53        c         2.623620      69.1038702
#> 54        d       127.152988      32.2513817
#> 55        e      -121.571320      41.9305361
#> 56        f      -164.632318     -34.8504423
#> 57        g       150.097074     -66.1525701
#> 58        h      -103.951279     -54.8149194
#> 59        i        22.577567     -52.3518973
#> 60        j       155.010865      79.3259066
#> 61        a        66.043705      86.0651009
#> 62        b        76.689250     -20.8667924
#> 63        c        88.469219     -18.5887876
#> 64        d      -107.327500     -53.3658621
#> 65        e       -40.639346     -20.8005798
#> 66        f       -79.142436      40.9065398
#> 67        g       -47.219889     -19.6125143
#> 68        h        86.947131     -77.4004860
#> 69        i      -123.536319     -73.7293517
#> 70        j       146.067302       5.5928810
#> 71        a       179.853848       8.3526256
#> 72        b      -160.532072       4.5055731
#> 73        c       123.710390     -26.9456694
#> 74        d       120.548948     -75.7143405
#> 75        e       122.305287     -44.5025468
#> 76        f        31.800903      -6.2975188
#> 77        g       127.303510     -35.6153485
#> 78        h        89.208413      59.1936329
#> 79        i       135.956098     -19.6052035
#> 80        j       -38.376440     -65.5944476
#> 81        a       -87.274036     -66.5209836
#> 82        b        77.801910     -30.0982997
#> 83        c       -34.478991     -70.9998527
#> 84        d       -88.876535      42.0302678
#> 85        e      -111.361224      -6.5953259
#> 86        f       161.796173     -41.6224380
#> 87        g       -85.763883     -89.5706554
#> 88        h       110.983703     -88.4774067
#> 89        i       104.512907      49.8169140
#> 90        j       113.208814      49.7768902
#> 91        a        84.851420      -6.9397186
#> 92        b        77.056075      76.5544834
#> 93        c        40.230633      41.1692447
#> 94        d       -76.847080     -42.7420489
#> 95        e       132.415484       1.5586267
#> 96        f       -49.450961     -33.1200895
#> 97        g        93.382639     -41.0997033
#> 98        h       -63.934327     -34.4776047
#> 99        i       164.623691      84.4345927
#> 100       j      -167.217681      50.8126400
cc_outl(x, method = "quantile", value = "flagged")
#> Testing geographic outliers
#> Flagged 0 records.
#>   [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [16] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [31] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [46] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [61] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [76] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [91] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
cc_outl(x, method = "distance", value = "flagged", tdi = 10000)
#> Testing geographic outliers
#> Flagged 0 records.
#>   [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [16] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [31] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [46] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [61] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [76] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [91] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
cc_outl(x, method = "distance", value = "flagged", tdi = 1000)
#> Testing geographic outliers
#> Flagged 86 records.
#>   [1] FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE
#>  [13] FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE  TRUE FALSE
#>  [25] FALSE FALSE  TRUE  TRUE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE
#>  [37] FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [49] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [61]  TRUE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE
#>  [73] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [85] FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [97] FALSE FALSE FALSE FALSE