Identify Geographic Outliers in Species Distributions

Removes out or flags records that are outliers in geographic space according to the method defined via the method argument. Geographic outliers often represent erroneous coordinates, for example due to data entry errors, imprecise geo-references, individuals in horticulture/captivity.

cc_outl(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  species = "species",
  method = "quantile",
  mltpl = 5,
  tdi = 1000,
  value = "clean",
  sampling_thresh = 0,
  verbose = TRUE,
  min_occs = 7,
  thinning = FALSE,
  thinning_res = 0.5
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

species

character string. The column with the species name. Default = “species”.

method

character string. Defining the method for outlier selection. See details. One of “distance”, “quantile”, “mad”. Default = “quantile”.

mltpl

numeric. The multiplier of the interquartile range (method == 'quantile') or median absolute deviation (method == 'mad')to identify outliers. See details. Default = 5.

tdi

numeric. The minimum absolute distance (method == 'distance') of a record to all other records of a species to be identified as outlier, in km. See details. Default = 1000.

value

character string. Defining the output value. See value.

sampling_thresh

numeric. Cut off threshold for the sampling correction. Indicates the quantile of sampling in which outliers should be ignored. For instance, if sampling_thresh == 0.25, records in the 25 (no sampling correction).

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

min_occs

Minimum number of geographically unique datapoints needed for a species to be tested. This is necessary for reliable outlier estimation. Species with fewer than min_occs records will not be tested and the output value will be 'TRUE'. Default is to 7. If method == 'distance', consider a lower threshold.

thinning

forces a raster approximation for the distance calculation. This is routinely used for species with more than 10,000 records for computational reasons, but can be enforced for smaller datasets, which is recommended when sampling is very uneven.

thinning_res

The resolution for the spatial thinning in decimal degrees. Default = 0.5.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Details

The method for outlier identification depends on the method argument. If “quantile”: a boxplot method is used and records are flagged as outliers if their mean distance to all other records of the same species is larger than mltpl * the interquartile range of the mean distance of all records of this species. If “mad”: the median absolute deviation is used. In this case a record is flagged as outlier, if the mean distance to all other records of the same species is larger than the median of the mean distance of all points plus/minus the mad of the mean distances of all records of the species * mltpl. If “distance”: records are flagged as outliers, if the minimum distance to the next record of the species is > tdi. For species with records from > 10000 unique locations a random sample of 1000 records is used for the distance matrix calculation. The test skips species with fewer than min_occs, geographically unique records.

The likelihood of occurrence records being erroneous outliers is linked to the sampling effort in any given location. To account for this, the sampling_cor option fetches the number of occurrence records available from www.gbif.org, per country as a proxy of sampling effort. The outlier test (the mean distance) for each records is than weighted by the log transformed number of records per square kilometre in this country. See for https://besjournals.onlinelibrary.wiley.com/doi/full/10.1111/2041-210X.13152 an example and further explanation of the outlier test.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

See also

Examples

x <- data.frame(species = letters[1:10],
                decimalLongitude = runif(100, -180, 180),
                decimalLatitude = runif(100, -90,90))

cc_outl(x)
#> Testing geographic outliers
#> Removed 0 records.
#>     species decimalLongitude decimalLatitude
#> 1         a        85.177964      34.3687002
#> 2         b       138.764854      56.3569539
#> 3         c       140.662535      48.2629894
#> 4         d       118.502708      37.9813192
#> 5         e       -27.372084      87.8338827
#> 6         f       -47.252599      60.0217443
#> 7         g       -11.918477     -79.9431982
#> 8         h       159.175764      72.7541191
#> 9         i      -120.740449     -54.8649574
#> 10        j       151.978560     -50.4147778
#> 11        a        37.934600     -39.0327446
#> 12        b       -73.046005     -82.5273070
#> 13        c       -69.755088      73.6416461
#> 14        d       177.551983     -52.5952466
#> 15        e        64.580859      22.5397754
#> 16        f       -30.554799      72.3398797
#> 17        g        40.835836     -77.4267824
#> 18        h        80.687683     -42.6218007
#> 19        i       148.459817      23.7644124
#> 20        j      -132.724203      14.7551885
#> 21        a        57.404053     -42.0951076
#> 22        b        59.280303      -0.1118544
#> 23        c      -160.044816      73.3487235
#> 24        d       111.367012     -56.1701753
#> 25        e      -129.641197     -41.3823463
#> 26        f       -12.182931       9.3714678
#> 27        g      -149.810127      50.7465432
#> 28        h        83.024734     -54.9143462
#> 29        i         2.670467     -47.2241267
#> 30        j        40.624985     -55.6754225
#> 31        a       109.731583     -53.6216183
#> 32        b      -124.704890     -83.6942029
#> 33        c       -45.177921     -46.4453538
#> 34        d       -81.303699      45.9584952
#> 35        e       171.698218     -36.3094825
#> 36        f       -97.379219      77.9268648
#> 37        g        60.045732      33.4266141
#> 38        h       -63.237613      56.3072691
#> 39        i        80.173089     -61.7613545
#> 40        j      -148.606440     -57.2949083
#> 41        a        -9.135653     -13.9843585
#> 42        b      -135.502291      68.0000325
#> 43        c       -85.344778     -82.5360399
#> 44        d       -61.433440     -60.5229702
#> 45        e       -15.947328     -21.9233993
#> 46        f       -13.268510      40.1544532
#> 47        g       -97.070316     -40.8865820
#> 48        h       165.565770      24.2091967
#> 49        i        -0.060070     -75.2113112
#> 50        j        30.095895     -39.0360041
#> 51        a       111.400657     -28.7522753
#> 52        b       -91.784810      52.7170701
#> 53        c        72.825226       1.9409017
#> 54        d       -49.869035      58.9112287
#> 55        e       -47.198813      32.3762637
#> 56        f       130.421939     -16.9314278
#> 57        g      -159.999079      67.6712134
#> 58        h       -33.385758     -83.1572964
#> 59        i      -176.753502     -41.2231946
#> 60        j        45.515478       5.7480090
#> 61        a       -68.800392     -75.8208440
#> 62        b       -82.479068       1.6584866
#> 63        c        37.566631     -66.2413769
#> 64        d        16.373476      12.7436258
#> 65        e      -172.701878     -51.3455589
#> 66        f        69.521869      32.4494403
#> 67        g        93.871927      47.3725791
#> 68        h       105.487079      53.5368346
#> 69        i      -149.393522      73.1651759
#> 70        j      -122.301518      31.0134668
#> 71        a        96.179672      19.4113581
#> 72        b       159.922618      13.8410458
#> 73        c      -148.535799     -16.6139302
#> 74        d       -84.565282     -49.7369731
#> 75        e       -86.558050      17.9208879
#> 76        f        57.192895      46.6040021
#> 77        g        26.519707      77.5697986
#> 78        h       117.573868     -26.2385579
#> 79        i       104.791173       1.2636332
#> 80        j      -117.085318      80.6758026
#> 81        a       -39.830792      17.8621316
#> 82        b        53.337896      69.1745543
#> 83        c        47.580547     -82.1935969
#> 84        d      -146.671598     -74.0209134
#> 85        e        39.394537      65.6187705
#> 86        f        93.652110      41.2874954
#> 87        g        -9.869951      43.4541523
#> 88        h       -74.054955     -61.8638147
#> 89        i        -4.504213     -34.9571479
#> 90        j       127.852316      38.4517069
#> 91        a       -38.490161     -36.6235821
#> 92        b       -74.786671      50.7006264
#> 93        c       -50.558720      55.0191254
#> 94        d        54.951649     -87.0309168
#> 95        e        91.544457      51.1490461
#> 96        f       -92.503747     -12.3479823
#> 97        g      -154.216269      75.2100428
#> 98        h       -22.055928     -16.2136065
#> 99        i       175.558734      24.8868521
#> 100       j       -58.962320     -14.5553569
cc_outl(x, method = "quantile", value = "flagged")
#> Testing geographic outliers
#> Flagged 0 records.
#>   [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [16] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [31] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [46] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [61] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [76] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [91] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
cc_outl(x, method = "distance", value = "flagged", tdi = 10000)
#> Testing geographic outliers
#> Flagged 0 records.
#>   [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [16] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [31] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [46] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [61] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [76] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [91] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
cc_outl(x, method = "distance", value = "flagged", tdi = 1000)
#> Testing geographic outliers
#> Flagged 96 records.
#>   [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE
#>  [13] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [25] FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE
#>  [37] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [49] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE
#>  [61] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [73] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [85] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [97]  TRUE FALSE FALSE FALSE