Identify Geographic Outliers in Species Distributions

Removes out or flags records that are outliers in geographic space according to the method defined via the method argument. Geographic outliers often represent erroneous coordinates, for example due to data entry errors, imprecise geo-references, individuals in horticulture/captivity.

cc_outl(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  species = "species",
  method = "quantile",
  mltpl = 5,
  tdi = 1000,
  value = "clean",
  sampling_thresh = 0,
  verbose = TRUE,
  min_occs = 7,
  thinning = FALSE,
  thinning_res = 0.5
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

species

character string. The column with the species name. Default = “species”.

method

character string. Defining the method for outlier selection. See details. One of “distance”, “quantile”, “mad”. Default = “quantile”.

mltpl

numeric. The multiplier of the interquartile range (method == 'quantile') or median absolute deviation (method == 'mad')to identify outliers. See details. Default = 5.

tdi

numeric. The minimum absolute distance (method == 'distance') of a record to all other records of a species to be identified as outlier, in km. See details. Default = 1000.

value

character string. Defining the output value. See value.

sampling_thresh

numeric. Cut off threshold for the sampling correction. Indicates the quantile of sampling in which outliers should be ignored. For instance, if sampling_thresh == 0.25, records in the 25 (no sampling correction).

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

min_occs

Minimum number of geographically unique datapoints needed for a species to be tested. This is necessary for reliable outlier estimation. Species with fewer than min_occs records will not be tested and the output value will be 'TRUE'. Default is to 7. If method == 'distance', consider a lower threshold.

thinning

forces a raster approximation for the distance calculation. This is routinely used for species with more than 10,000 records for computational reasons, but can be enforced for smaller datasets, which is recommended when sampling is very uneven.

thinning_res

The resolution for the spatial thinning in decimal degrees. Default = 0.5.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Details

The method for outlier identification depends on the method argument. If “quantile”: a boxplot method is used and records are flagged as outliers if their mean distance to all other records of the same species is larger than mltpl * the interquartile range of the mean distance of all records of this species. If “mad”: the median absolute deviation is used. In this case a record is flagged as outlier, if the mean distance to all other records of the same species is larger than the median of the mean distance of all points plus/minus the mad of the mean distances of all records of the species * mltpl. If “distance”: records are flagged as outliers, if the minimum distance to the next record of the species is > tdi. For species with records from > 10000 unique locations a random sample of 1000 records is used for the distance matrix calculation. The test skips species with fewer than min_occs, geographically unique records.

The likelihood of occurrence records being erroneous outliers is linked to the sampling effort in any given location. To account for this, the sampling_cor option fetches the number of occurrence records available from www.gbif.org, per country as a proxy of sampling effort. The outlier test (the mean distance) for each records is than weighted by the log transformed number of records per square kilometre in this country. See for https://besjournals.onlinelibrary.wiley.com/doi/full/10.1111/2041-210X.13152 an example and further explanation of the outlier test.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

See also

Examples

x <- data.frame(species = letters[1:10],
                decimalLongitude = runif(100, -180, 180),
                decimalLatitude = runif(100, -90,90))

cc_outl(x)
#> Testing geographic outliers
#> Removed 0 records.
#>     species decimalLongitude decimalLatitude
#> 1         a        18.597290        7.188174
#> 2         b      -135.315991      -60.904963
#> 3         c      -140.941943        6.960009
#> 4         d        98.225580      -47.876787
#> 5         e      -139.887154      -72.867304
#> 6         f       -65.279072       66.299996
#> 7         g        -4.627747      -81.800146
#> 8         h        64.728577      -75.534912
#> 9         i        93.187829      -22.949090
#> 10        j      -116.510109      -44.816708
#> 11        a      -129.773251       23.066472
#> 12        b      -179.489836      -61.114162
#> 13        c        63.289222      -86.593483
#> 14        d      -152.891493       67.705287
#> 15        e       -57.009754       56.302349
#> 16        f       -86.734273      -26.019868
#> 17        g      -109.773952       49.025058
#> 18        h      -130.367952       59.592745
#> 19        i       160.883723       45.691650
#> 20        j       -60.983761      -43.020464
#> 21        a       113.545302       14.092612
#> 22        b       -96.908877      -89.557140
#> 23        c        13.010052       41.833096
#> 24        d        43.800031       15.106734
#> 25        e       114.922518       33.967958
#> 26        f       161.672773      -22.593768
#> 27        g       157.695492      -82.009619
#> 28        h        59.796044      -65.831821
#> 29        i         4.751828      -51.568565
#> 30        j      -177.151798       26.802975
#> 31        a      -158.411334      -39.973567
#> 32        b        52.861622       14.194764
#> 33        c       148.829632       62.680858
#> 34        d       110.785509       61.708124
#> 35        e        24.430448      -89.501074
#> 36        f         9.313974      -26.147366
#> 37        g       -41.276161       13.781685
#> 38        h       114.996467      -41.178477
#> 39        i      -137.185240      -53.769323
#> 40        j       154.233265       85.426460
#> 41        a         3.195561      -15.866165
#> 42        b        13.619637       35.393344
#> 43        c        20.959598      -13.960250
#> 44        d       -70.723122      -65.836083
#> 45        e      -158.672314        4.098278
#> 46        f        33.796723       30.264012
#> 47        g       122.816166       37.498953
#> 48        h      -105.598692       73.493083
#> 49        i        38.881849       19.434686
#> 50        j       -49.283989       74.735130
#> 51        a       116.815169       -7.720564
#> 52        b       -96.320078       45.507518
#> 53        c       113.846961       64.720198
#> 54        d      -116.260055      -67.459563
#> 55        e       -55.736659      -69.765642
#> 56        f      -106.254238       21.042821
#> 57        g      -126.396682      -24.594416
#> 58        h        18.672666      -21.138258
#> 59        i       -61.397010       11.596217
#> 60        j        20.180191       87.058892
#> 61        a      -177.063769      -42.920942
#> 62        b       143.914378       22.211367
#> 63        c      -104.690137       52.810521
#> 64        d        55.605140       32.116673
#> 65        e        33.182508      -38.605258
#> 66        f        18.164302      -48.197385
#> 67        g       178.559168       75.932492
#> 68        h       160.201319      -47.912270
#> 69        i       -46.503341       67.216383
#> 70        j        93.178875       47.692968
#> 71        a       -37.018130       14.310356
#> 72        b       -50.709585      -59.246542
#> 73        c        41.013322      -46.317807
#> 74        d       -59.208943      -40.977662
#> 75        e       174.718747       83.834763
#> 76        f       -50.709365       13.005318
#> 77        g      -114.019173       -3.495776
#> 78        h       -18.573675       50.794224
#> 79        i      -102.009120       82.409654
#> 80        j         5.615097        9.804099
#> 81        a       174.336783      -65.381108
#> 82        b       179.340911       51.257683
#> 83        c        36.535543       -9.651480
#> 84        d       -94.052382       80.809489
#> 85        e      -169.904017      -53.456008
#> 86        f        16.071933       32.592503
#> 87        g        23.469209       14.769930
#> 88        h       -45.326364      -43.363585
#> 89        i      -165.998848      -74.866601
#> 90        j       116.427764      -23.105011
#> 91        a      -113.782691      -22.446435
#> 92        b       -92.160618      -89.615541
#> 93        c       153.113181      -87.988237
#> 94        d       -32.564997       -0.176867
#> 95        e      -141.084588       31.664085
#> 96        f      -171.820101       81.870697
#> 97        g      -114.228758       84.026435
#> 98        h      -103.610421      -49.930205
#> 99        i         7.184872      -60.649896
#> 100       j      -165.878409       26.703903
cc_outl(x, method = "quantile", value = "flagged")
#> Testing geographic outliers
#> Flagged 0 records.
#>   [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [16] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [31] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [46] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [61] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [76] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [91] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
cc_outl(x, method = "distance", value = "flagged", tdi = 10000)
#> Testing geographic outliers
#> Flagged 1 records.
#>   [1]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [13]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [25]  TRUE FALSE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [37]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [49]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [61]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [73]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [85]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [97]  TRUE  TRUE  TRUE  TRUE
cc_outl(x, method = "distance", value = "flagged", tdi = 1000)
#> Testing geographic outliers
#> Flagged 94 records.
#>   [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [13]  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE
#>  [25] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [37] FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [49] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE
#>  [61] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [73] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [85] FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE FALSE FALSE FALSE
#>  [97] FALSE FALSE FALSE FALSE