Flag Geographic Outliers in Species Distributions

Flags records that are outliers in geographic space according to the method defined via the method argument. Geographic outliers often represent erroneous coordinates, for example due to data entry errors, imprecise geo-references, individuals in horticulture/captivity.

cc_outl(x, lon = "decimallongitude", lat = "decimallatitude",
  species = "species", method = "quantile", mltpl = 3, tdi = 1000,
  value = "clean", verbose = TRUE)

Arguments

x
a data.frame. Containing geographical coordinates and species names.

lon
a character string. The column with the longitude coordinates. Default = “decimallongitude”.

lat
a character string. The column with the longitude coordinates. Default = “decimallatitude”.

species
a character string. The column with the species name. Default = “species”.

method
a character string. Defining the method for outlier selection. See details. One of “distance”, “quantile”, “mad”. Default = “quantile”.

mltpl
numeric. The multiplier of the interquartile range (method == 'quantile') or median absolute deviation (method == 'mad')to identify outliers. See details. Default = 3.

tdi
numeric. The minimum absolute distance (method == 'distance') of a record to all other records of a species to be identified as outlier, in km. See details. Default = 1000.

value
a character string. Defining the output value. See value.

verbose
logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic. Default = “clean”.

Details

The method for outlier identification depends on the method argument. If “outlier”: a boxplot method is used and records are flagged as outliers if their mean distance to all other records of the same species is larger than mltpl * the interquartile range of the mean distance of all records of this species. If “mad”: the median absolute deviation is used. In this case a record is flagged as outlier, if the mean distance to all other records of the same species is larger than the median of the mean distance of all points plus/minus the mad of the mean distances of all records of the species * mltpl. If “distance”: records are flagged as outliers, if the minimum distance to the next record of the species is > tdi.

Note

See https://github.com/azizka/CoordinateCleaner/wiki for more details and tutorials.

Examples

x <- data.frame(species = letters[1:10],
                decimallongitude = runif(100, -180, 180),
                decimallatitude = runif(100, -90,90))

cc_outl(x)

#> Testing geographic outliers

#> Flagged 2 records.

#>     species decimallongitude decimallatitude
#> 1         a       -94.183290     -51.0073273
#> 2         b       -31.795210     -68.3925429
#> 3         c        88.585803     -59.3525413
#> 4         d        80.442199       3.8751993
#> 6         f        -9.366995      69.8170535
#> 7         g        28.245974      24.5603874
#> 8         h       -93.026260     -30.7115547
#> 9         i      -138.823682      18.4000342
#> 10        j       -72.595932      48.9890761
#> 11        a        -5.716982     -52.1756473
#> 12        b        24.266275      15.1755059
#> 13        c        91.423119       7.5736460
#> 14        d        16.353085     -73.3686043
#> 15        e        93.484066     -23.5186515
#> 16        f      -103.331713      81.4372962
#> 17        g       122.135409      80.9440634
#> 18        h       -33.589060      -9.0401721
#> 19        i       134.546328     -43.2327162
#> 20        j       137.576354     -86.5968894
#> 21        a      -141.231828      39.8529086
#> 22        b        26.835974      81.5057369
#> 23        c       178.572017     -10.7598053
#> 24        d       126.199236       0.7222283
#> 25        e       -22.444180     -82.1597403
#> 26        f       -62.485227      76.0454020
#> 27        g       -59.051211      76.2821201
#> 28        h       -93.861426      19.0165885
#> 29        i       -28.306395     -89.2652434
#> 30        j         4.369259      28.2341586
#> 31        a      -113.239308     -62.8327890
#> 32        b       -62.495812     -54.1510800
#> 33        c       114.146511     -46.8651007
#> 34        d        89.545679      29.0833141
#> 35        e        11.427073     -46.4780121
#> 36        f       166.165693     -77.1210471
#> 37        g      -142.543306      36.8409470
#> 38        h        85.364520     -33.2285562
#> 39        i       125.092793     -22.3424649
#> 40        j        51.745951      56.7175935
#> 41        a       -96.315409     -53.8992060
#> 42        b       149.842644     -35.6525664
#> 43        c       -27.144880     -74.2103564
#> 44        d      -110.540338     -57.0023769
#> 45        e        72.558121     -52.7756625
#> 46        f       177.329056      63.8927098
#> 47        g        75.208503     -42.6762186
#> 48        h         6.294657      80.1601694
#> 49        i        75.143537      52.6923846
#> 50        j       -44.949428      59.6847406
#> 51        a       -88.281444      21.4159005
#> 52        b       178.708657     -10.1279469
#> 53        c         8.242330      83.6326104
#> 54        d        77.691390     -60.8042333
#> 55        e      -179.102726     -56.5823177
#> 56        f       -74.531242      20.3582555
#> 57        g      -114.906240      -3.3327939
#> 58        h       -33.266199     -18.6951784
#> 59        i      -101.881480     -32.6238466
#> 60        j      -154.316639      60.9075016
#> 61        a       -88.807831     -36.1433654
#> 62        b        54.892628      20.8911305
#> 63        c        84.399878      -6.8213096
#> 64        d       114.304732      79.0570721
#> 65        e       -69.843478     -88.8675504
#> 66        f       138.689125     -27.1030527
#> 67        g        38.361187      11.9288507
#> 68        h       137.166901       5.2122714
#> 69        i        64.771428      24.6627719
#> 70        j      -128.540282     -26.8567417
#> 71        a        95.567609      76.3508155
#> 72        b        45.707987      -4.6725799
#> 73        c        82.921846      50.0641411
#> 74        d       -27.466412      53.7725764
#> 76        f        96.857293      73.7718939
#> 77        g        31.444804      -9.8032039
#> 78        h       143.484392     -37.8793209
#> 79        i       133.316834     -35.4040675
#> 80        j       -36.249755      75.4243826
#> 81        a       167.442880      78.4282311
#> 82        b        -1.443224      30.8966988
#> 83        c       -12.024894     -23.1714062
#> 84        d      -141.211969      12.0836585
#> 85        e       149.147298      -3.8522805
#> 86        f      -128.230758      12.1947812
#> 87        g       -84.532611     -38.4230088
#> 88        h      -126.553303      46.7278431
#> 89        i       -30.870474     -48.8761608
#> 90        j      -140.778118     -13.2290292
#> 91        a       -91.051939      -7.8147489
#> 92        b        32.631520      37.7832166
#> 93        c       -61.823773     -38.2577439
#> 94        d        83.869388      49.1508473
#> 95        e        82.045881      13.4212114
#> 96        f        16.460432     -86.0767627
#> 97        g        84.148674      49.1447665
#> 98        h       156.009015     -79.3567920
#> 99        i       -92.378666     -85.0233059
#> 100       j       -78.134740      43.3523869

cc_outl(x, method = "quantile", value = "flagged")

#> Testing geographic outliers

#> Flagged 2 records.

#>   [1]  TRUE  TRUE  TRUE  TRUE FALSE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [13]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [25]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [37]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [49]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [61]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [73]  TRUE  TRUE FALSE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [85]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [97]  TRUE  TRUE  TRUE  TRUE

cc_outl(x, method = "distance", value = "flagged", tdi = 10000)

#> Testing geographic outliers

#> Flagged 0 records.

#>   [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [16] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [31] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [46] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [61] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [76] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [91] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE

cc_outl(x, method = "distance", value = "flagged", tdi = 1000)

#> Testing geographic outliers

#> Flagged 90 records.

#>   [1]  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE
#>  [13] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE
#>  [25]  TRUE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [37] FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [49] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [61] FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [73] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE
#>  [85] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [97] FALSE FALSE  TRUE  TRUE