Identify Geographic Outliers in Species Distributions

Removes out or flags records that are outliers in geographic space according to the method defined via the method argument. Geographic outliers often represent erroneous coordinates, for example due to data entry errors, imprecise geo-references, individuals in horticulture/captivity.

cc_outl(
  x,
  lon = "decimallongitude",
  lat = "decimallatitude",
  species = "species",
  method = "quantile",
  mltpl = 5,
  tdi = 1000,
  value = "clean",
  sampling_thresh = 0,
  verbose = TRUE,
  min_occs = 7,
  thinning = FALSE,
  thinning_res = 0.5
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimallongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimallatitude”.

species

character string. The column with the species name. Default = “species”.

method

character string. Defining the method for outlier selection. See details. One of “distance”, “quantile”, “mad”. Default = “quantile”.

mltpl

numeric. The multiplier of the interquartile range (method == 'quantile') or median absolute deviation (method == 'mad')to identify outliers. See details. Default = 5.

tdi

numeric. The minimum absolute distance (method == 'distance') of a record to all other records of a species to be identified as outlier, in km. See details. Default = 1000.

value

character string. Defining the output value. See value.

sampling_thresh

numeric. Cut off threshold for the sampling correction. Indicates the quantile of sampling in which outliers should be ignored. For instance, if sampling_thresh == 0.25, records in the 25 not be flagged as outliers. Default = 0 (no sampling correction).

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

min_occs

Minimum number of geographically unique datapoints needed for a species to be tested. This is necessary for reliable outlier estimation. Species with fewer than min_occs records will not be tested and the output value will be 'TRUE'. Default is to 7. If method == 'distance', consider a lower threshold.

thinning

forces a raster approximation for the distance calculation. This is routinely used for species with more than 10,000 records for computational reasons, but can be enforced for smaller datasets, which is recommended when sampling is very uneven.

thinning_res

The resolution for the spatial thinning in decimal degrees. Default = 0.5.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Details

The method for outlier identification depends on the method argument. If “outlier”: a boxplot method is used and records are flagged as outliers if their mean distance to all other records of the same species is larger than mltpl * the interquartile range of the mean distance of all records of this species. If “mad”: the median absolute deviation is used. In this case a record is flagged as outlier, if the mean distance to all other records of the same species is larger than the median of the mean distance of all points plus/minus the mad of the mean distances of all records of the species * mltpl. If “distance”: records are flagged as outliers, if the minimum distance to the next record of the species is > tdi. For species with records from > 10000 unique locations a random sample of 1000 records is used for the distance matrix calculation. The test skips species with fewer than min_occs, geographically unique records.

The likelihood of occurrence records being erroneous outliers is linked to the sampling effort in any given location. To account for this, the sampling_cor option fetches the number of occurrence records available from www.gbif.org, per country as a proxy of sampling effort. The outlier test (the mean distance) for each records is than weighted by the log transformed number of records per square kilometre in this country. See for https://besjournals.onlinelibrary.wiley.com/doi/full/10.1111/2041-210X.13152 an example and further explanation of the outlier test.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimallongitude = runif(100, -180, 180), 
                decimallatitude = runif(100, -90,90))
                
cc_outl(x)
#> Testing geographic outliers
#> Removed 0 records.
#>     species decimallongitude decimallatitude
#> 1         a       55.6708085     -24.1962277
#> 2         b      176.9795744     -22.0069766
#> 3         c      -20.8429972     -86.2727836
#> 4         d     -111.6536300       4.1832885
#> 5         e      -88.7059892      -4.5000270
#> 6         f      137.4144966      14.2426851
#> 7         g     -160.0712599     -10.2350200
#> 8         h       54.2106916     -65.8503977
#> 9         i     -122.2379586      11.8628017
#> 10        j     -159.5294142     -30.1113443
#> 11        a       43.5180937      28.6514499
#> 12        b      -98.5923805      14.0612991
#> 13        c     -139.0229885      61.5391188
#> 14        d     -152.4301972     -70.3369924
#> 15        e     -106.1678099     -38.0930796
#> 16        f     -132.0976120     -64.6384329
#> 17        g       -0.7046615      68.7513440
#> 18        h      -47.8812723     -31.3825617
#> 19        i      -25.1610252      72.8574108
#> 20        j       -5.2817515      38.8000224
#> 21        a     -162.8028901      80.4316243
#> 22        b      157.2022776       0.2027219
#> 23        c     -149.4789922       3.0735934
#> 24        d      -12.5126183     -10.6551352
#> 25        e      143.4320754      34.7013124
#> 26        f     -164.7663555      58.9281616
#> 27        g     -145.3151656     -22.9417162
#> 28        h     -151.5891054     -29.2073465
#> 29        i     -101.4372099     -87.1673417
#> 30        j     -130.7224004     -24.2655624
#> 31        a      -46.8410067     -20.4005129
#> 32        b      178.2909455     -24.4167532
#> 33        c       43.4565818      23.0258786
#> 34        d     -107.2686355      -5.7415353
#> 35        e       23.6699518      75.4248184
#> 36        f      -47.4349627      36.1973483
#> 37        g        7.8824494      16.2715906
#> 38        h       76.7733125      75.1428284
#> 39        i      -58.7289425      50.1875863
#> 40        j      121.5266725      -8.6138464
#> 41        a       -9.6852475       7.3084091
#> 42        b      -92.0010790       3.7368277
#> 43        c     -172.2577941     -79.1788049
#> 44        d      176.0284546     -84.5194205
#> 45        e     -128.4660723      21.2043667
#> 46        f       -7.9481974      41.0043872
#> 47        g      105.3946593     -75.6117715
#> 48        h       66.9219186      74.0367191
#> 49        i       89.3102210      55.4704136
#> 50        j       50.3449548      26.7666957
#> 51        a       57.7551597      82.0862767
#> 52        b     -112.6482773      28.9845177
#> 53        c     -164.6201793      -7.7263997
#> 54        d       93.5516212      -3.7650447
#> 55        e       58.9998386     -31.4340644
#> 56        f      -18.1419713     -43.3972132
#> 57        g      152.3256150      88.7679347
#> 58        h      168.8552412     -25.7539428
#> 59        i       27.7412316      24.2543685
#> 60        j      -52.3621741     -89.1230751
#> 61        a     -168.8649101      70.9559917
#> 62        b      -95.5788445      56.5180598
#> 63        c     -171.5691184      85.1711084
#> 64        d      154.6659352       7.9622833
#> 65        e       73.7365190     -39.6754868
#> 66        f      174.5667271     -14.5218831
#> 67        g      148.5884076      22.7000260
#> 68        h      -57.5005804      83.0090191
#> 69        i      112.2217498     -59.5046785
#> 70        j      120.6507385      73.1819670
#> 71        a       52.4592866     -54.6682948
#> 72        b     -150.8866119      77.0726774
#> 73        c     -172.4964179      48.0734499
#> 74        d       58.6130102     -39.6258591
#> 75        e       69.0483967      44.7322306
#> 76        f     -154.9193559      37.1797326
#> 77        g     -146.3994972     -13.9818751
#> 78        h     -174.7452543       3.7850509
#> 79        i       59.1627392      34.1276197
#> 80        j      -16.7915318      63.1356649
#> 81        a      -26.9831875       5.0346955
#> 82        b      144.5084273      23.5897112
#> 83        c      105.5253849      -0.1066766
#> 84        d       -5.8577725     -37.1210664
#> 85        e     -141.6565229      81.7695164
#> 86        f      141.9694718      52.9345661
#> 87        g       96.1237362      59.2229450
#> 88        h      108.6382021      45.2006184
#> 89        i      -44.4628532       4.3251699
#> 90        j       24.8519481      72.5679862
#> 91        a      -14.5745376       4.6609807
#> 92        b      -72.1031924     -34.8415520
#> 93        c       99.1173485     -39.1579259
#> 94        d     -165.2907619       7.2406812
#> 95        e       46.2037547     -20.4245186
#> 96        f      -92.9017315     -48.7729595
#> 97        g      -89.8135783     -21.0648497
#> 98        h      -55.3147522      -3.7649322
#> 99        i       77.1478182     -52.9597829
#> 100       j      131.4862031     -44.6475101
cc_outl(x, method = "quantile", value = "flagged")
#> Testing geographic outliers
#> Flagged 0 records.
#>   [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [16] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [31] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [46] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [61] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [76] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [91] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
cc_outl(x, method = "distance", value = "flagged", tdi = 10000)
#> Testing geographic outliers
#> Flagged 0 records.
#>   [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [16] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [31] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [46] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [61] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [76] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [91] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
cc_outl(x, method = "distance", value = "flagged", tdi = 1000)
#> Testing geographic outliers
#> Flagged 94 records.
#>   [1] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [13] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [25] FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE
#>  [37] FALSE  TRUE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE
#>  [49] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [61] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [73] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [85] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE
#>  [97] FALSE FALSE FALSE FALSE