Identify Invalid lat/lon Coordinates

Removes or flags non-numeric and not available coordinates as well as lat >90, lat <-90, lon > 180 and lon < -180 are flagged.

cc_val(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Details

This test is obligatory before running any further tests of CoordinateCleaner, as additional tests only run with valid coordinates.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = c(runif(106, -180, 180), NA, "13W33'", "67,09", 305), 
                decimalLatitude = runif(110, -90,90))
                
cc_val(x)
#> Testing coordinate validity
#> Removed 4 records.
#>     species decimalLongitude decimalLatitude
#> 1         a     -150.6378666      12.9831783
#> 2         b      125.8571717      45.6170932
#> 3         c      -77.1802055     -89.9733934
#> 4         d     -160.3688126     -13.7481884
#> 5         e      -64.9134149      42.8176967
#> 6         f     -108.3696242      12.7494509
#> 7         g      -68.8653999      -2.2445733
#> 8         h     -157.7196862     -45.3381658
#> 9         i     -175.3140258      68.6538444
#> 10        j      -10.3732739      44.6687170
#> 11        a      -21.0187280     -60.5378859
#> 12        b       62.8170699      63.9040029
#> 13        c      145.4468196     -16.4162861
#> 14        d      108.4606612      53.0273579
#> 15        e     -177.1561213      83.4046998
#> 16        f        0.7886546     -32.3468753
#> 17        g       90.1609995     -35.8019468
#> 18        h      175.8903135      66.6350640
#> 19        i     -166.1778869     -46.1962942
#> 20        j      -99.5766773      -9.4142063
#> 21        a     -136.1715484      23.4158389
#> 22        b      113.3219508       1.8519401
#> 23        c     -169.5972164      25.4734789
#> 24        d      -75.1510961      68.4592171
#> 25        e      -43.2580642       3.0385623
#> 26        f       97.4690525     -88.9405406
#> 27        g     -169.5851598     -68.8564727
#> 28        h     -150.6276083      36.9190763
#> 29        i       68.0520001     -88.1586545
#> 30        j       43.3290070      44.9945436
#> 31        a       76.8500909      -0.6940875
#> 32        b     -177.0020713      21.3681155
#> 33        c      -36.8548668     -14.9336973
#> 34        d      175.1118433      77.7573040
#> 35        e     -149.5944567      32.1996018
#> 36        f     -132.5261901     -74.2031030
#> 37        g       48.7943142       3.3272610
#> 38        h       36.2294180     -41.9335099
#> 39        i      127.8190517      40.1827742
#> 40        j     -117.7863825      62.6093883
#> 41        a       74.8101903      20.2602337
#> 42        b       23.8920529     -18.9931667
#> 43        c       78.3599674     -23.6251192
#> 44        d     -149.4144062     -32.4242319
#> 45        e       30.4399300      21.3136039
#> 46        f      138.8285416     -66.5968217
#> 47        g       57.6958501     -59.3113224
#> 48        h      -98.2835952     -66.8809166
#> 49        i      162.9885538       0.2566234
#> 50        j     -132.3249120     -47.0871055
#> 51        a       49.9888619      54.7015313
#> 52        b       66.1092496      37.2433086
#> 53        c       36.2298053     -29.5300072
#> 54        d      -99.5819537     -26.9518795
#> 55        e      -68.6395215      41.5777338
#> 56        f     -162.9856670      63.2288814
#> 57        g     -166.9083182      22.2344659
#> 58        h       60.6261592     -51.0964777
#> 59        i      -18.1638261     -46.2803681
#> 60        j      -67.9135883       9.8201741
#> 61        a     -177.7255395      73.1779986
#> 62        b      -14.0795788     -62.2451080
#> 63        c       49.1339555     -40.4339646
#> 64        d     -135.6350655     -35.7955892
#> 65        e       -3.8958455      62.7650256
#> 66        f       64.3142400      86.8273714
#> 67        g       44.9284547      10.7508222
#> 68        h      121.0846648     -39.5864307
#> 69        i       70.0025379      67.0263523
#> 70        j     -146.1485019      32.7401446
#> 71        a      109.9035814      89.2417523
#> 72        b      -44.6494736      70.1375654
#> 73        c       -4.1083488      13.6345240
#> 74        d      -12.7683912     -54.2842900
#> 75        e      -60.4686250      32.1921753
#> 76        f       55.1848197      44.0908888
#> 77        g      -44.4338879      58.0805731
#> 78        h       88.9499971     -50.4249453
#> 79        i     -141.5092298      21.4263665
#> 80        j      157.2278463     -73.8783867
#> 81        a        0.8491827      22.0986835
#> 82        b      -24.0427174     -15.5205952
#> 83        c      110.4501368     -37.1184963
#> 84        d       60.5285633      47.9840469
#> 85        e       69.0783614     -68.7373119
#> 86        f       70.4285698      -4.5104756
#> 87        g       89.8093253     -16.0965970
#> 88        h       75.4712795     -68.9888917
#> 89        i       76.4582248      27.7114289
#> 90        j        5.1869379     -66.2341100
#> 91        a     -133.6955410     -69.7605295
#> 92        b       74.4608156     -88.1092541
#> 93        c     -167.7854019      -4.3543642
#> 94        d       71.8303941      65.7086974
#> 95        e       30.1125380     -61.2819539
#> 96        f      -78.0583453      18.3201047
#> 97        g     -168.3314487      66.0324513
#> 98        h      -44.6428054     -21.5664276
#> 99        i        5.3986043     -18.7849042
#> 100       j       81.0747130       6.3353073
#> 101       a     -176.3284638      66.0153389
#> 102       b      -35.9631321     -54.1918713
#> 103       c     -174.7255074      88.5373613
#> 104       d      108.4697005     -32.7533940
#> 105       e     -105.4637679     -62.6434668
#> 106       f      109.4560695     -77.7932616
cc_val(x, value = "flagged")
#> Testing coordinate validity
#> Flagged 4 records.
#>   [1]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [13]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [25]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [37]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [49]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [61]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [73]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [85]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [97]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE
#> [109] FALSE FALSE