Identify Invalid lat/lon Coordinates

Removes or flags non-numeric and not available coordinates as well as lat >90, lat <-90, lon > 180 and lon < -180 are flagged.

cc_val(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Details

This test is obligatory before running any further tests of CoordinateCleaner, as additional tests only run with valid coordinates.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = c(runif(106, -180, 180), NA, "13W33'", "67,09", 305), 
                decimalLatitude = runif(110, -90,90))
                
cc_val(x)
#> Testing coordinate validity
#> Removed 4 records.
#>     species decimalLongitude decimalLatitude
#> 1         a       -56.101558      39.9143406
#> 2         b       -64.857598     -19.9052121
#> 3         c       129.033575      12.9142753
#> 4         d       171.071810      65.1458581
#> 5         e       -43.962335      46.2982236
#> 6         f        -3.896064      69.2688770
#> 7         g       -69.561484      65.7241250
#> 8         h       -94.713551      66.7613952
#> 9         i        96.521729      49.8547042
#> 10        j       156.801882      11.0029955
#> 11        a       153.226085     -55.0115331
#> 12        b       -22.344139      69.8257112
#> 13        c        90.825037      41.8466525
#> 14        d        23.910276      27.5344595
#> 15        e        49.035442      21.1851241
#> 16        f        30.319846     -27.2760746
#> 17        g       145.151547      82.6121462
#> 18        h      -125.057174       5.6414725
#> 19        i       -53.820584      45.6396006
#> 20        j       127.595650     -84.8835863
#> 21        a       135.770067      -5.4714180
#> 22        b       -68.059744      77.9577840
#> 23        c       121.540741     -56.6670468
#> 24        d       132.568743      26.7972200
#> 25        e       127.851428     -57.2814369
#> 26        f        -8.254665     -40.0643124
#> 27        g      -107.435992     -39.6424274
#> 28        h       166.609988      32.8067645
#> 29        i       -92.970035     -10.9059143
#> 30        j       -90.000591      30.6641054
#> 31        a        73.050185      37.2979489
#> 32        b       -95.701097     -85.1376493
#> 33        c       141.374969      13.3726166
#> 34        d       -72.409564      89.4434009
#> 35        e      -104.229780      46.9089187
#> 36        f       166.454257       6.8427684
#> 37        g       -65.536347      21.9779871
#> 38        h       -71.528231      42.8340386
#> 39        i      -136.764614     -14.1323238
#> 40        j       134.095199      80.6521370
#> 41        a       -84.311357     -26.3555113
#> 42        b       -24.912990      -0.6276272
#> 43        c        60.672475     -84.8875113
#> 44        d        96.370102       5.3688127
#> 45        e      -154.459990     -72.9527056
#> 46        f       -62.496977      53.7454642
#> 47        g       132.427615      83.1539241
#> 48        h        53.536760      54.5035304
#> 49        i        16.113735      24.1010105
#> 50        j        58.750607      -9.7965050
#> 51        a        64.434813      52.7236626
#> 52        b      -115.411181     -50.0591081
#> 53        c       -33.841569      50.4984717
#> 54        d       -45.422496      85.9368729
#> 55        e        13.099165      63.1728180
#> 56        f       118.073932     -39.4267166
#> 57        g       153.972444      76.6461134
#> 58        h       133.232943      45.6619607
#> 59        i       151.262263     -56.6302795
#> 60        j       -67.685073     -20.9474945
#> 61        a       -54.396791      11.7955342
#> 62        b        58.123205      87.6027622
#> 63        c      -142.984156      15.7313464
#> 64        d       134.930305     -22.5178689
#> 65        e       172.702483      89.4536052
#> 66        f       119.484975     -69.5993387
#> 67        g       -52.836274     -51.6634752
#> 68        h        89.456442      32.8568604
#> 69        i       -30.679775      62.6381488
#> 70        j      -120.106859      69.4460353
#> 71        a      -119.682210     -40.8658222
#> 72        b        -1.907854     -46.4436051
#> 73        c      -126.535523     -21.5228301
#> 74        d        82.965101     -68.0789318
#> 75        e        32.297507      30.9062357
#> 76        f        17.641199     -15.3929848
#> 77        g       177.838833      86.5334121
#> 78        h      -157.004644     -82.8040253
#> 79        i       150.617776      41.1932286
#> 80        j       168.556330      43.6412739
#> 81        a       170.121511     -26.0671346
#> 82        b      -165.578556     -52.4154401
#> 83        c       176.802248     -76.3232976
#> 84        d       -82.878321       9.0155816
#> 85        e       -89.752283     -54.6466524
#> 86        f        31.659393      54.5320886
#> 87        g        32.154074       0.5618503
#> 88        h      -125.974475      -4.9520287
#> 89        i      -167.300460      55.4973153
#> 90        j       -89.156656      44.1974205
#> 91        a      -125.049251     -12.0262882
#> 92        b       -82.636630      80.7738010
#> 93        c       -24.162910      44.1603377
#> 94        d       -23.438628      89.1815595
#> 95        e      -141.706529      66.9725834
#> 96        f        36.942330     -26.3111375
#> 97        g       170.767100      71.0738811
#> 98        h       -52.343640      62.0649450
#> 99        i       171.455505     -50.8941951
#> 100       j       139.294070      69.7837234
#> 101       a      -138.610373     -86.0957669
#> 102       b        74.419669      46.6081975
#> 103       c        59.810614      48.4366708
#> 104       d       -97.007898      -1.5246868
#> 105       e        45.478959      65.7722751
#> 106       f       129.393632     -83.1751156
cc_val(x, value = "flagged")
#> Testing coordinate validity
#> Flagged 4 records.
#>   [1]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [13]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [25]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [37]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [49]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [61]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [73]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [85]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [97]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE
#> [109] FALSE FALSE