Identify Invalid lat/lon Coordinates

Removes or flags non-numeric and not available coordinates as well as lat >90, lat <-90, lon > 180 and lon < -180 are flagged.

cc_val(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Details

This test is obligatory before running any further tests of CoordinateCleaner, as additional tests only run with valid coordinates.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = c(runif(106, -180, 180), NA, "13W33'", "67,09", 305), 
                decimalLatitude = runif(110, -90,90))
                
cc_val(x)
#> Testing coordinate validity
#> Removed 4 records.
#>     species decimalLongitude decimalLatitude
#> 1         a      -12.1612549    -27.19679228
#> 2         b       34.3391654     76.41557040
#> 3         c      -87.0693401      0.95650348
#> 4         d     -179.1851675     -2.10014229
#> 5         e     -108.5796183     50.49205660
#> 6         f       74.9766441    -76.88045095
#> 7         g        7.3201151     41.04649751
#> 8         h      -20.7071384    -10.53785435
#> 9         i     -107.0113669     19.63060406
#> 10        j       18.7977322     -3.77730014
#> 11        a       99.0552903     60.96239179
#> 12        b      -56.0395411    -65.58685851
#> 13        c      -48.6349458    -16.75592304
#> 14        d      167.1868983      9.53507812
#> 15        e       18.0307251     53.22441061
#> 16        f      -23.9400679     26.44615932
#> 17        g      147.3774076     -8.97173481
#> 18        h      -73.5662516    -36.19095475
#> 19        i      163.8834772    -32.55446970
#> 20        j     -153.1262064    -47.31538324
#> 21        a     -131.7706957     68.95147846
#> 22        b      -66.2775956     -8.58813215
#> 23        c      -72.6063297     68.83850293
#> 24        d      112.2305115    -16.59077440
#> 25        e       89.2004461     66.70432710
#> 26        f     -163.4789831    -25.46465816
#> 27        g      -35.4063249     70.17652642
#> 28        h      173.7794330    -31.63914350
#> 29        i     -121.0435399     -0.14826778
#> 30        j        6.4143902     10.41528163
#> 31        a       98.6457932     67.68762232
#> 32        b       30.8926258    -45.80848510
#> 33        c      -37.8177992    -46.85581463
#> 34        d       -0.4457759    -61.10909408
#> 35        e      -61.5116934    -16.96327079
#> 36        f      179.6024862    -50.42768216
#> 37        g      127.4450832     26.18525582
#> 38        h      122.7277997    -65.70170681
#> 39        i      -16.4881737    -63.46216445
#> 40        j      152.2177936    -56.46993896
#> 41        a        6.0527972    -87.01223663
#> 42        b       18.6148068     27.00218137
#> 43        c     -130.1468748     15.93441350
#> 44        d     -151.7546909     22.40039055
#> 45        e     -179.7139665     77.68873618
#> 46        f      149.0016124     35.96961203
#> 47        g       79.9490859     89.93339270
#> 48        h      166.2221680    -14.68888394
#> 49        i     -175.2017906    -55.68258421
#> 50        j     -157.0539657     79.18995197
#> 51        a     -160.5952894     84.80284281
#> 52        b      -86.9628240     13.66187286
#> 53        c       29.0892225     68.42193437
#> 54        d     -128.2644437     84.00133782
#> 55        e      -76.1792617    -76.69214575
#> 56        f      176.2212654    -29.76407383
#> 57        g       67.7081777    -80.86448100
#> 58        h      110.1732192     12.83807515
#> 59        i      161.1140304    -50.85218164
#> 60        j     -148.1726670    -89.75779334
#> 61        a     -138.7293408    -31.99084581
#> 62        b     -146.4730333    -45.52617393
#> 63        c      172.3953189     26.40850706
#> 64        d      111.6457256    -29.64123201
#> 65        e      -70.5411256    -78.03120658
#> 66        f      163.2142082    -40.62262374
#> 67        g     -177.5067227     36.33434845
#> 68        h      113.9193450    -13.10126582
#> 69        i      -89.3496672    -47.78299708
#> 70        j       72.1596492     64.46440425
#> 71        a        6.8522629    -14.32076391
#> 72        b      -92.9663152     66.69665162
#> 73        c       69.0528040    -54.26897302
#> 74        d      -86.9744986     71.92467379
#> 75        e      -76.9277458    -22.92809925
#> 76        f      -88.4583766     88.05576417
#> 77        g     -110.7744728     19.97035361
#> 78        h      -37.5856922    -87.24854184
#> 79        i     -102.1112796    -72.47905785
#> 80        j       89.6976937     39.18746116
#> 81        a       31.7845629     71.91590531
#> 82        b      146.9611559     26.32892223
#> 83        c     -135.7235105    -37.07953543
#> 84        d      -16.7739488    -30.26535900
#> 85        e       12.0334518    -73.21945651
#> 86        f      175.1112891     39.08694555
#> 87        g      109.4232753    -31.73581723
#> 88        h     -129.6309432     72.49052865
#> 89        i      140.9782368     64.73878375
#> 90        j     -130.5391997     52.88727533
#> 91        a        6.3129089     21.08306693
#> 92        b     -148.8743408     71.53705548
#> 93        c       98.3949410     66.69009651
#> 94        d      -35.8698143    -88.23587148
#> 95        e      177.2105382     -0.82558058
#> 96        f      141.9031557     -6.79661572
#> 97        g      106.7599238    -12.78892555
#> 98        h     -159.5930861     68.73559500
#> 99        i      160.1335340     44.04666610
#> 100       j       48.1106173     19.70755354
#> 101       a      144.1205649      0.01893644
#> 102       b     -109.9583174     41.70222058
#> 103       c     -145.2872497     -2.99897639
#> 104       d       93.3391728     87.38675842
#> 105       e     -163.2382830    -86.58974219
#> 106       f      -48.8614513    -18.86274789
cc_val(x, value = "flagged")
#> Testing coordinate validity
#> Flagged 4 records.
#>   [1]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [13]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [25]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [37]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [49]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [61]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [73]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [85]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [97]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE
#> [109] FALSE FALSE