Identify Invalid lat/lon Coordinates

Removes or flags non-numeric and not available coordinates as well as lat >90, lat <-90, lon > 180 and lon < -180 are flagged.

cc_val(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Details

This test is obligatory before running any further tests of CoordinateCleaner, as additional tests only run with valid coordinates.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = c(runif(106, -180, 180), NA, "13W33'", "67,09", 305), 
                decimalLatitude = runif(110, -90,90))
                
cc_val(x)
#> Testing coordinate validity
#> Removed 4 records.
#>     species decimalLongitude decimalLatitude
#> 1         a      -62.2073574     -56.6197254
#> 2         b     -166.3856437       4.9545839
#> 3         c       72.7206566      20.6572763
#> 4         d      -40.1850301     -51.8238960
#> 5         e        0.5232006     -23.5792859
#> 6         f      114.4345184     -55.7881366
#> 7         g      -40.7258126     -89.0872433
#> 8         h       75.8255843       1.4761547
#> 9         i      -50.8806801      35.2067610
#> 10        j      174.7635742     -75.4281604
#> 11        a      -70.6044126      23.3797299
#> 12        b     -140.9209563     -29.9244347
#> 13        c      107.5106906     -39.0420429
#> 14        d       68.6098237      72.2382240
#> 15        e     -171.0774032      26.6962813
#> 16        f       39.7738320     -14.5078627
#> 17        g     -157.7030388     -66.8523119
#> 18        h     -141.1536478     -47.4454966
#> 19        i       80.0017093      82.7420984
#> 20        j     -131.4793893     -85.6921646
#> 21        a        5.4906632     -42.0682840
#> 22        b      120.8339799     -71.1487695
#> 23        c      137.9620695      48.7013482
#> 24        d       94.4971825     -15.3711374
#> 25        e     -174.1417202     -27.9674087
#> 26        f     -175.2289138     -39.6831278
#> 27        g     -176.1743615     -45.2977406
#> 28        h      -48.5082451      46.8711801
#> 29        i       84.6823022     -68.2379222
#> 30        j      -37.0589807       2.8868969
#> 31        a     -107.9751000      -1.0558393
#> 32        b      -58.7593256      38.7479381
#> 33        c     -172.4400854      -0.4730234
#> 34        d     -130.9871377     -59.8975064
#> 35        e       24.0575357     -66.2821015
#> 36        f       75.3977571      21.8453566
#> 37        g      -16.9196813     -72.8425448
#> 38        h      138.6536074     -89.3694932
#> 39        i      116.6252990     -51.1563697
#> 40        j      178.7120690     -61.8694348
#> 41        a      -20.4655409      46.8981948
#> 42        b      -44.0598764      -7.5698261
#> 43        c       75.3576867     -46.3867051
#> 44        d      -82.7995226      52.5584042
#> 45        e     -112.4217414      62.2071151
#> 46        f       88.2522354      44.8115349
#> 47        g      -72.4812748     -23.3967631
#> 48        h       49.6893831      89.2919606
#> 49        i      -71.5921195      12.6991739
#> 50        j     -100.1850992      12.5290802
#> 51        a      -16.3892245       0.5985166
#> 52        b       31.9451609      22.9176467
#> 53        c     -100.7429679     -71.9176918
#> 54        d      167.4858467     -31.1990511
#> 55        e       75.1761311     -60.2227245
#> 56        f      176.4291257     -84.2163312
#> 57        g     -154.1850552      44.6826899
#> 58        h       25.1923388     -26.9009462
#> 59        i     -107.1203523     -27.3705529
#> 60        j      -74.9687035      -0.2138296
#> 61        a       82.5691987     -81.7502942
#> 62        b        6.0117611     -82.6929488
#> 63        c       -0.8279501      40.2178293
#> 64        d      109.4017882      68.5601730
#> 65        e       75.8121297     -58.8945537
#> 66        f      152.4390319     -19.6558011
#> 67        g      116.3478294     -66.0072354
#> 68        h     -158.1255711     -66.4104694
#> 69        i       78.6240279     -31.7559328
#> 70        j      -55.0938852     -64.8416999
#> 71        a      151.6609735     -35.2856126
#> 72        b      -91.6220987      87.5538436
#> 73        c       -4.7607182      50.4973306
#> 74        d      141.8739310       2.7159248
#> 75        e       79.0104785     -22.8349614
#> 76        f       67.9952831     -66.6481192
#> 77        g     -114.4403212      58.9573556
#> 78        h      -60.9748243       8.9636520
#> 79        i      129.5053485     -87.3944833
#> 80        j     -142.6675452     -71.8267842
#> 81        a      164.1792125      86.4776796
#> 82        b       30.0304409     -31.2848203
#> 83        c      -75.8060206      57.5600145
#> 84        d      126.3913798      12.9830023
#> 85        e      115.6479597      17.8596210
#> 86        f     -138.1256219      38.4970424
#> 87        g      151.2155746     -71.6213471
#> 88        h      -27.8375805     -41.5839596
#> 89        i      -26.7194320      54.0818191
#> 90        j      163.0024075     -59.0295977
#> 91        a      109.2385540     -75.1608782
#> 92        b      -31.7773493      66.1265986
#> 93        c      -85.4571776      25.2665230
#> 94        d      -97.5837664      85.3438489
#> 95        e      -36.7506037      84.8772707
#> 96        f       99.9220639      27.3528003
#> 97        g      -38.8473723      74.7787261
#> 98        h      -88.1208392     -87.0937666
#> 99        i      -48.0729937      37.9795332
#> 100       j      -77.0579223     -46.6410344
#> 101       a     -145.5683923      77.9826293
#> 102       b      -72.4143082      64.4354884
#> 103       c     -136.9151188      68.5578430
#> 104       d       98.5337248       7.3566162
#> 105       e      -92.8079842      36.5576678
#> 106       f      116.2873843      80.6140371
cc_val(x, value = "flagged")
#> Testing coordinate validity
#> Flagged 4 records.
#>   [1]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [13]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [25]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [37]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [49]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [61]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [73]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [85]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [97]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE
#> [109] FALSE FALSE