Identify Invalid lat/lon Coordinates

Removes or flags non-numeric and not available coordinates as well as lat >90, lat <-90, lon > 180 and lon < -180 are flagged.

cc_val(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Details

This test is obligatory before running any further tests of CoordinateCleaner, as additional tests only run with valid coordinates.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = c(runif(106, -180, 180), NA, "13W33'", "67,09", 305), 
                decimalLatitude = runif(110, -90,90))
                
cc_val(x)
#> Testing coordinate validity
#> Removed 4 records.
#>     species decimalLongitude decimalLatitude
#> 1         a       122.017051      85.5188759
#> 2         b        36.238892       9.4213849
#> 3         c        81.924764     -87.7429191
#> 4         d        58.858313     -53.1326842
#> 5         e       153.983314      -0.4997437
#> 6         f       -30.087039      23.1478201
#> 7         g       -18.361383      51.2154982
#> 8         h        80.716332       2.4371481
#> 9         i       -92.718030     -18.1571829
#> 10        j        32.675191      70.3801156
#> 11        a       -88.234664     -40.7584675
#> 12        b       135.002506     -26.3160780
#> 13        c       -72.724957     -19.8522283
#> 14        d      -153.658730     -50.7373585
#> 15        e       -96.263585     -61.8912965
#> 16        f      -120.911503     -46.6954108
#> 17        g      -114.857832      39.2265626
#> 18        h       115.551933      57.6487605
#> 19        i      -106.978082     -59.8959800
#> 20        j       -89.292714      13.9752234
#> 21        a        27.167804      11.6684492
#> 22        b       151.583742     -69.1683753
#> 23        c       -60.961246     -52.5998765
#> 24        d        -8.189985      26.5903345
#> 25        e        56.946209     -77.7412876
#> 26        f        29.936584      10.4945333
#> 27        g      -151.769504      59.3876175
#> 28        h        97.114547      61.5397453
#> 29        i        -9.554710       6.8472687
#> 30        j      -164.490381     -82.8997016
#> 31        a       -76.732785      52.4005416
#> 32        b       -38.768074      73.0902990
#> 33        c        98.800291      82.7749766
#> 34        d       -19.035950     -22.6327963
#> 35        e       124.088070      77.6764650
#> 36        f      -111.938406      64.5265224
#> 37        g      -142.779909      -3.3871217
#> 38        h       -52.869550     -52.3045080
#> 39        i        99.587297     -26.6939764
#> 40        j        24.461162      45.2927849
#> 41        a      -179.267182      18.3242989
#> 42        b       -50.997383     -42.5011341
#> 43        c      -129.208210      30.4542008
#> 44        d       -32.617626     -23.2023202
#> 45        e       -59.329568      64.9104864
#> 46        f       -55.082220      22.4433660
#> 47        g       -68.341799     -51.1382600
#> 48        h      -151.974097      18.5389170
#> 49        i      -117.218327      14.7258931
#> 50        j      -126.838804      62.5883271
#> 51        a        76.656805     -64.4778124
#> 52        b        22.594941      55.0484721
#> 53        c      -129.022148     -80.5601897
#> 54        d       126.924540      40.7669174
#> 55        e        23.420823     -43.3921794
#> 56        f      -152.899779     -10.0303949
#> 57        g       169.149012      84.4876426
#> 58        h      -133.355883      31.0674987
#> 59        i        -6.345036     -55.7938455
#> 60        j       -36.069467      62.2729041
#> 61        a       169.083149     -53.6892966
#> 62        b       -72.193506      86.9189377
#> 63        c       151.173107     -71.5015924
#> 64        d        55.115757     -81.9956240
#> 65        e      -176.577787      26.3223960
#> 66        f       115.322226      30.6539101
#> 67        g        -2.542706      13.6237381
#> 68        h       151.031162     -56.9432821
#> 69        i       -87.920691     -75.0443660
#> 70        j        -9.552838      39.7841800
#> 71        a      -119.318356     -46.9760539
#> 72        b        87.183351      54.1994686
#> 73        c       -33.500309      45.8317323
#> 74        d        49.422701      34.8471793
#> 75        e        37.518421     -17.1873049
#> 76        f       -57.316584      23.4706688
#> 77        g         7.312242      50.0783718
#> 78        h       172.738731      43.7513485
#> 79        i        37.972024      53.5787960
#> 80        j        16.675518     -47.4016057
#> 81        a       163.247794     -27.3869370
#> 82        b      -138.561733     -20.3692455
#> 83        c       -70.295981      69.7802645
#> 84        d        94.467923     -84.9266139
#> 85        e         3.467947      -1.9272463
#> 86        f      -172.829138      55.8661875
#> 87        g      -101.571292      63.3217941
#> 88        h       129.597794      50.4927636
#> 89        i      -135.721814      29.9631253
#> 90        j       -27.924158      34.3351640
#> 91        a       -97.737204       1.9611761
#> 92        b        30.458229      57.8809652
#> 93        c        11.096707       0.3782271
#> 94        d        45.805643     -25.4113113
#> 95        e      -162.775582     -88.9503492
#> 96        f       -76.629897     -70.9620103
#> 97        g        33.257121     -45.1383911
#> 98        h       -80.398985     -14.8676877
#> 99        i       104.017671      52.1460935
#> 100       j      -110.966836      76.8413865
#> 101       a       -56.007940       0.4885847
#> 102       b        61.290170     -20.8984428
#> 103       c       176.028699     -12.7680736
#> 104       d       165.476390     -81.2067904
#> 105       e      -124.938894      70.7446427
#> 106       f        -3.477351      -3.2244526
cc_val(x, value = "flagged")
#> Testing coordinate validity
#> Flagged 4 records.
#>   [1]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [13]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [25]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [37]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [49]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [61]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [73]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [85]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [97]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE
#> [109] FALSE FALSE