Identify Invalid lat/lon Coordinates

Removes or flags non-numeric and not available coordinates as well as lat >90, la <-90, lon > 180 and lon < -180 are flagged.

cc_val(
  x,
  lon = "decimallongitude",
  lat = "decimallatitude",
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimallongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimallatitude”.

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Details

This test is obligatory before running any further tests of CoordinateCleaner, as additional tests only run with valid coordinates.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimallongitude = c(runif(106, -180, 180), NA, "13W33'", "67,09", 305), 
                decimallatitude = runif(110, -90,90))
                
cc_val(x)
#> Testing coordinate validity
#> Removed 4 records.
#>     species decimallongitude decimallatitude
#> 1         a      124.5904825     13.28386432
#> 2         b     -155.3993993    -61.85215458
#> 3         c       42.9057986     55.23507562
#> 4         d     -178.3184339    -78.99907869
#> 5         e     -139.4928774    -28.00677891
#> 6         f      111.3353475    -27.24798957
#> 7         g      117.6706998      7.64038939
#> 8         h      152.5451989     40.15411349
#> 9         i      -12.3619742     58.29593405
#> 10        j      148.3779364    -44.01011450
#> 11        a      148.3448156     10.63970340
#> 12        b      -90.3069410    -81.06141420
#> 13        c       -6.6248277     16.83444072
#> 14        d      -73.7779207     34.57453486
#> 15        e      -12.9682000    -15.50798850
#> 16        f     -130.7492674    -46.02698883
#> 17        g      172.0307073     44.82242569
#> 18        h       60.8963568     22.51341524
#> 19        i     -132.1368397     20.28249560
#> 20        j      -39.1908042    -37.56204145
#> 21        a     -101.4737775     53.16371486
#> 22        b       32.7415982    -86.00873016
#> 23        c      -89.8246066     55.25016799
#> 24        d      152.8109801     -1.22919709
#> 25        e      114.3270978    -43.92156120
#> 26        f     -168.9023542    -34.23648897
#> 27        g     -156.4617597     -0.02790327
#> 28        h      -43.9020972    -26.09011271
#> 29        i     -115.3589184    -26.65749602
#> 30        j     -125.8305479     51.08246613
#> 31        a      -38.0900730    -27.78998250
#> 32        b      -50.5962241     73.80216851
#> 33        c       42.6252653    -39.97008214
#> 34        d     -100.5319887    -72.69042150
#> 35        e      119.5722240    -13.89449841
#> 36        f     -102.9724391     -4.27471865
#> 37        g      116.9492991    -21.58911092
#> 38        h      140.1956997     81.09200580
#> 39        i      166.2014124     78.86320125
#> 40        j     -120.8116801     34.24702129
#> 41        a       67.0994451    -69.67158129
#> 42        b      -77.7444162    -15.29360146
#> 43        c      124.9317937     89.52938380
#> 44        d     -113.7276589    -19.83788077
#> 45        e      178.0767341    -37.95727966
#> 46        f      -97.4458437     66.02358752
#> 47        g       16.9615243    -44.13974885
#> 48        h       53.2914896    -47.97123481
#> 49        i     -134.4073476    -26.82918920
#> 50        j      154.7602561     33.72588271
#> 51        a      101.3052134     47.95451706
#> 52        b      151.1543304     80.75829947
#> 53        c      121.7162811     80.23961913
#> 54        d      -25.8227365    -87.94194614
#> 55        e       40.8666913    -24.67150913
#> 56        f        0.3655162    -65.50253636
#> 57        g      -85.4785792     37.50683379
#> 58        h      -55.7995552     28.50392392
#> 59        i       82.8477727     31.66411692
#> 60        j      111.3221649     62.71003055
#> 61        a      -62.9386372     77.17414688
#> 62        b      -56.7769150     18.36720009
#> 63        c       11.7720343    -54.06565068
#> 64        d     -137.4118989    -29.26673625
#> 65        e      101.8050855      5.03979270
#> 66        f       81.5952992    -26.97049149
#> 67        g      -29.2912613     76.56654704
#> 68        h       68.8155982    -22.99011813
#> 69        i      138.6646433    -38.92609360
#> 70        j      -48.0514391    -14.59036339
#> 71        a      -14.8096827     11.07614495
#> 72        b     -167.9202742    -53.98012425
#> 73        c       39.0955587    -72.97291368
#> 74        d     -114.4438706     49.43659684
#> 75        e      118.3213007      7.71121677
#> 76        f        7.6303310     67.19859066
#> 77        g      -45.9543093     31.74159651
#> 78        h       63.8466889     58.58027947
#> 79        i      -82.0390357    -56.78494484
#> 80        j      -40.0160540    -79.51653804
#> 81        a      -40.1552253    -73.02303247
#> 82        b      -82.2180574    -26.61836996
#> 83        c      171.4449710     16.63946948
#> 84        d      131.3537592     38.94680503
#> 85        e       69.5261055    -15.27152650
#> 86        f       60.1869988     36.82571598
#> 87        g      -97.3056246     53.95524756
#> 88        h       89.7440349     39.95872072
#> 89        i      -53.2310970    -36.05837031
#> 90        j       85.8741256     74.83842907
#> 91        a     -163.1191013    -17.14632640
#> 92        b      133.2364823    -37.96471131
#> 93        c      148.0263834     25.68172371
#> 94        d     -109.6456230    -46.84910002
#> 95        e       15.4684707     57.08260621
#> 96        f     -157.5111662     35.95105064
#> 97        g      -66.4420396     57.44562681
#> 98        h       99.1817441      6.70193353
#> 99        i       24.0115886     44.76987925
#> 100       j      108.9019174    -69.65836111
#> 101       a      -54.5746374      9.04952796
#> 102       b       71.2220317    -23.83255681
#> 103       c      -69.6377671    -41.90141156
#> 104       d       43.7186146    -50.58442807
#> 105       e      -62.0050736     71.56523125
#> 106       f      -18.4729978    -57.72390503
cc_val(x, value = "flagged")
#> Testing coordinate validity
#> Flagged 4 records.
#>   [1]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [13]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [25]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [37]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [49]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [61]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [73]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [85]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [97]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE
#> [109] FALSE FALSE