Check Coordinate Validity in lat/lon

Checks if all coordinates in a data set are valid in a latitude/longitude coordinate reference system. That is non-numeric, not available coordinates and lat >90, la <-90, lon > 180 and lon < -180 are flagged.

cc_val(x, lon = "decimallongitude", lat = "decimallatitude",
  value = "clean", verbose = TRUE)

Arguments

x
a data.frame. Containing geographical coordinates and species names.

lon
a character string. The column with the longitude coordinates. Default = “decimallongitude”.

lat
a character string. The column with the longitude coordinates. Default = “decimallatitude”.

value
a character string. Defining the output value. See value.

verbose
logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic. Default = “clean”.

Details

This test is obligatory before running any further tests of CoordinateCleaner, as additional tests only run with valid coordinates.

Note

See https://github.com/azizka/CoordinateCleaner/wiki for more details and tutorials.

Examples

x <- data.frame(species = letters[1:10],
                decimallongitude = c(runif(106, -180, 180), NA, "13W33'", "67,09", 305),
                decimallatitude = runif(110, -90,90))

cc_val(x)

#> Testing coordinate validity

#> Flagged 4 records.

#>     species decimallongitude decimallatitude
#> 1         a      -140.123324      51.3673425
#> 2         b      -109.580890     -49.1476593
#> 3         c      -177.454521      -9.3278763
#> 4         d       158.713414     -60.9790253
#> 5         e      -155.444634     -58.2998958
#> 6         f      -130.705495     -54.3189378
#> 7         g       -60.185871     -25.6299635
#> 8         h        -3.712215     -57.3600233
#> 9         i      -117.736366      11.0060748
#> 10        j      -156.621873      28.6349417
#> 11        a       -17.399884      28.8456136
#> 12        b       -43.415153     -89.5678582
#> 13        c       147.347412      88.8202861
#> 14        d      -169.219275      22.9480615
#> 15        e      -109.873309     -87.3665395
#> 16        f        64.612574     -53.0679209
#> 17        g        89.087869      29.3537995
#> 18        h        53.833577      -6.5260745
#> 19        i      -152.070956     -25.1373316
#> 20        j        35.875316      34.0859505
#> 21        a       136.443892     -43.4050581
#> 22        b        83.533376      63.2169058
#> 23        c      -151.180889      -3.5991556
#> 24        d       119.789932       6.7079502
#> 25        e       131.294051      29.0951815
#> 26        f        76.667390     -35.8461882
#> 27        g        92.666560     -42.5697272
#> 28        h      -135.923790     -35.0916966
#> 29        i      -110.268763      67.4924878
#> 30        j       -66.895816      39.3186990
#> 31        a      -130.488126     -16.7544551
#> 32        b       125.209037     -86.5152948
#> 33        c        -5.320085     -80.9675354
#> 34        d       -10.712619      36.4051858
#> 35        e       104.776335     -78.0533515
#> 36        f       -67.326022     -83.4961737
#> 37        g       -68.844775     -86.3723348
#> 38        h       -55.744079     -22.8994025
#> 39        i       118.239593     -49.3802273
#> 40        j       -81.361621      -1.4547420
#> 41        a       -24.899009      -9.6038718
#> 42        b       -29.916778      34.2368818
#> 43        c      -131.647608      61.6931964
#> 44        d        67.082904     -22.3499399
#> 45        e      -118.051878      87.2801439
#> 46        f       152.457640      81.4367393
#> 47        g        40.947693      88.1817316
#> 48        h       -66.373399     -10.2368962
#> 49        i       111.503913     -62.3225067
#> 50        j       144.854676      80.0509181
#> 51        a      -134.892792       3.5039545
#> 52        b        23.600590      -7.9857712
#> 53        c      -134.049202     -50.3837640
#> 54        d        45.935142     -64.9471708
#> 55        e      -152.883377     -51.0915814
#> 56        f       -77.237523     -16.9778389
#> 57        g       -68.183126     -47.9277735
#> 58        h      -127.178078     -85.4880553
#> 59        i       107.917385      45.4215526
#> 60        j      -118.081411      41.7730578
#> 61        a       102.846172     -16.3607806
#> 62        b       -10.787125      23.5544050
#> 63        c       -24.135218     -78.2199030
#> 64        d        26.831830     -63.2354127
#> 65        e        82.789180     -39.6536823
#> 66        f       141.907373      63.1182257
#> 67        g      -100.689707     -86.1226595
#> 68        h       -75.025282     -44.4906216
#> 69        i       112.643550      28.2080702
#> 70        j      -164.303161      52.0468937
#> 71        a         1.897799      37.5922166
#> 72        b       -93.003883     -78.3732614
#> 73        c        32.091271       2.0360419
#> 74        d       169.610605     -53.8328970
#> 75        e       -96.220439      56.5408067
#> 76        f       132.270486      27.6804137
#> 77        g      -140.125773     -26.4146432
#> 78        h       108.537041      17.3870813
#> 79        i        56.852804      40.9867872
#> 80        j        54.739547      89.2023371
#> 81        a       -98.523848      -0.3500511
#> 82        b       127.432667     -48.5469415
#> 83        c       -23.181466     -35.7742237
#> 84        d       -75.858575      39.6840018
#> 85        e        49.530481     -42.9470247
#> 86        f       -84.765795     -70.6146573
#> 87        g       -40.943384     -42.7639118
#> 88        h        42.909829     -56.3928870
#> 89        i        15.004110       7.6395023
#> 90        j        -5.460840     -78.2066746
#> 91        a        95.146171      68.4693246
#> 92        b       165.969697     -39.7417433
#> 93        c       -82.589568      31.0866609
#> 94        d        83.016080     -79.1764427
#> 95        e      -131.700403      87.9273990
#> 96        f      -162.576638     -39.1904224
#> 97        g        73.298545       9.7402260
#> 98        h      -127.003034      70.2151208
#> 99        i        96.895920     -85.8956301
#> 100       j      -116.513960      31.9839217
#> 101       a      -140.126760     -29.3292366
#> 102       b       157.803267     -64.5847030
#> 103       c       124.729610     -26.5384822
#> 104       d        25.611689     -17.7399258
#> 105       e        64.473626      70.0045834
#> 106       f      -147.843992     -33.7294462

cc_val(x, value = "flagged")

#> Testing coordinate validity

#> Flagged 4 records.

#>   [1]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [13]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [25]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [37]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [49]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [61]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [73]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [85]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [97]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE
#> [109] FALSE FALSE