Identify Invalid lat/lon Coordinates

Removes or flags non-numeric and not available coordinates as well as lat >90, lat <-90, lon > 180 and lon < -180 are flagged.

cc_val(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Details

This test is obligatory before running any further tests of CoordinateCleaner, as additional tests only run with valid coordinates.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = c(runif(106, -180, 180), NA, "13W33'", "67,09", 305), 
                decimalLatitude = runif(110, -90,90))
                
cc_val(x)
#> Testing coordinate validity
#> Removed 4 records.
#>     species decimalLongitude decimalLatitude
#> 1         a       123.964939     -76.8338669
#> 2         b        42.184869     -36.3316696
#> 3         c         5.110184     -72.7721552
#> 4         d        35.181756     -74.6769530
#> 5         e        47.468970      10.3735591
#> 6         f      -105.284291       1.7506877
#> 7         g       -73.003434     -59.8755029
#> 8         h       -66.364549     -46.5654559
#> 9         i      -162.632046     -78.9187892
#> 10        j        86.387998      56.8862980
#> 11        a       -46.893937      -9.0925980
#> 12        b        32.327108     -69.3476683
#> 13        c      -154.199598      80.3045919
#> 14        d       151.819203      65.2563752
#> 15        e      -157.951141      19.1614431
#> 16        f       -48.926296     -11.0425591
#> 17        g       -69.122688      -0.8782851
#> 18        h        22.322965     -29.5163913
#> 19        i       -39.415591      60.3983053
#> 20        j        46.621178     -19.4177766
#> 21        a      -167.388022       8.7124824
#> 22        b      -103.318051      16.9773566
#> 23        c       -61.519262     -20.4232113
#> 24        d       -73.186875      28.4490862
#> 25        e        76.024539     -83.9743430
#> 26        f         6.237786     -35.8486144
#> 27        g       174.569315     -80.7245382
#> 28        h       -91.876458     -44.4482643
#> 29        i         3.701151     -62.2514384
#> 30        j       -44.041157      54.4035968
#> 31        a        51.778589     -28.7364313
#> 32        b       -22.461695     -63.5232255
#> 33        c        26.719389      19.2589594
#> 34        d        73.127454     -32.0301598
#> 35        e       -57.474767     -29.0883859
#> 36        f       114.691009     -32.4189878
#> 37        g       -30.543392     -77.9669111
#> 38        h        -2.959365      -5.6112514
#> 39        i      -143.995255      46.0018764
#> 40        j        31.506385     -11.2721120
#> 41        a       -36.457718      39.2017336
#> 42        b       -98.738537     -18.9594428
#> 43        c       116.918455      25.6762999
#> 44        d       -17.927462       3.0245772
#> 45        e        46.776277      20.4923439
#> 46        f       -49.817078     -29.2379344
#> 47        g       139.280854      -2.8017542
#> 48        h        -7.689989      50.8139587
#> 49        i       -55.812371     -80.9773116
#> 50        j       -38.411080     -77.9631702
#> 51        a       -89.079992      79.0149776
#> 52        b      -125.711203      56.4870779
#> 53        c       157.192062     -75.4515535
#> 54        d       160.184297      62.9988721
#> 55        e       164.412258     -73.6097593
#> 56        f       -31.809686     -37.6803909
#> 57        g      -113.909275     -25.5597788
#> 58        h         2.060909     -12.8849761
#> 59        i        63.142689      -2.1870474
#> 60        j       -28.556012     -65.2429828
#> 61        a      -109.833718      73.7070573
#> 62        b       175.364471      -7.3084294
#> 63        c        -5.576123     -88.2354242
#> 64        d       108.336739      18.8730129
#> 65        e       -38.607101     -40.9961787
#> 66        f        81.944521      10.2361729
#> 67        g       -70.543890      51.4300691
#> 68        h       139.942199      23.3715352
#> 69        i       -41.338880      58.2218652
#> 70        j      -157.517448     -43.1286393
#> 71        a        68.345996     -13.1014268
#> 72        b        68.725022     -40.8516384
#> 73        c       147.593037     -22.3888133
#> 74        d      -122.140933     -83.4242843
#> 75        e        -2.119884      73.3687750
#> 76        f        89.890277     -10.7998350
#> 77        g       108.342657     -27.9604616
#> 78        h        16.293653       2.9402097
#> 79        i       152.756495      48.6979566
#> 80        j       130.243038      60.1955956
#> 81        a       119.474081      70.0922966
#> 82        b        80.735677     -35.4389490
#> 83        c       -88.730400     -13.6289758
#> 84        d        29.874733     -39.3057292
#> 85        e        36.316768      -1.2231574
#> 86        f        79.050148     -74.8174894
#> 87        g        63.240100      18.9313588
#> 88        h      -120.043947      56.2098320
#> 89        i       -84.941263      68.8661030
#> 90        j      -153.523508      -7.7279113
#> 91        a       112.230625     -51.9087739
#> 92        b      -124.957321      34.9682014
#> 93        c      -151.644146     -16.4759650
#> 94        d        21.545916      29.1832038
#> 95        e         6.801420     -69.6864727
#> 96        f        44.566095     -58.1219101
#> 97        g       166.984138     -57.6808322
#> 98        h        -4.180791      20.2093352
#> 99        i        61.028243     -56.9556251
#> 100       j       -55.747882      20.1500244
#> 101       a       -87.665020     -34.0014402
#> 102       b       -64.227994       4.8147739
#> 103       c       -35.653160     -43.8925381
#> 104       d       108.257843      61.3766328
#> 105       e        25.732997     -13.0778171
#> 106       f       -83.443801     -35.3466966
cc_val(x, value = "flagged")
#> Testing coordinate validity
#> Flagged 4 records.
#>   [1]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [13]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [25]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [37]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [49]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [61]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [73]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [85]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [97]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE
#> [109] FALSE FALSE