Identify Invalid lat/lon Coordinates

Removes or flags non-numeric and not available coordinates as well as lat >90, lat <-90, lon > 180 and lon < -180 are flagged.

cc_val(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Details

This test is obligatory before running any further tests of CoordinateCleaner, as additional tests only run with valid coordinates.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = c(runif(106, -180, 180), NA, "13W33'", "67,09", 305), 
                decimalLatitude = runif(110, -90,90))
                
cc_val(x)
#> Testing coordinate validity
#> Removed 4 records.
#>     species decimalLongitude decimalLatitude
#> 1         a       77.8643783      -65.598636
#> 2         b       59.9876184      -41.542036
#> 3         c       38.0530881       21.505645
#> 4         d     -101.9709551      -12.998252
#> 5         e       97.5920256       34.410202
#> 6         f       72.8090447       32.537822
#> 7         g     -129.4832957      -33.304158
#> 8         h       71.0060488       27.622494
#> 9         i     -150.9845208      -86.045812
#> 10        j     -159.0144078       80.751542
#> 11        a     -155.7747184       64.894780
#> 12        b     -124.6315817      -14.532344
#> 13        c      120.7505334       11.022844
#> 14        d      115.9876967       54.031236
#> 15        e       38.4266499      -32.848586
#> 16        f     -126.4901557        1.809583
#> 17        g     -131.2209066       74.870459
#> 18        h      -57.7296566      -11.067058
#> 19        i      -68.9941032       -9.634455
#> 20        j       52.1864909       39.934101
#> 21        a      -85.1082344       32.407851
#> 22        b      173.6522434       18.816853
#> 23        c       29.1115783      -79.343317
#> 24        d      -29.3446907      -60.241115
#> 25        e     -121.7916862       51.881150
#> 26        f      -14.9267005      -58.363643
#> 27        g      177.0433789      -67.817844
#> 28        h      127.9064808      -48.476123
#> 29        i      141.6896967       83.381831
#> 30        j      130.0609981       58.465092
#> 31        a      138.1377137       50.991060
#> 32        b      -29.0417957      -44.875445
#> 33        c       25.3792212      -88.445301
#> 34        d      -63.4506385      -43.795282
#> 35        e       69.3100817      -30.348437
#> 36        f       99.4985191      -17.888706
#> 37        g      -34.9443338       65.509224
#> 38        h      140.4960267       15.692794
#> 39        i     -133.8847648      -10.768187
#> 40        j       -9.7755929       63.396403
#> 41        a     -134.2956491      -65.112541
#> 42        b      161.4208670      -24.251991
#> 43        c       61.3172371      -15.563108
#> 44        d      -18.8059140       79.380457
#> 45        e      129.2388700      -11.282781
#> 46        f       79.4435945      -26.133823
#> 47        g     -112.6955769       52.126651
#> 48        h     -152.1104142      -58.885515
#> 49        i       47.3145544       52.522195
#> 50        j       27.6388314       30.832601
#> 51        a       25.7506641      -37.185595
#> 52        b      175.4209765      -62.462707
#> 53        c       -6.6625535       52.532532
#> 54        d      -33.6967463      -62.299216
#> 55        e       65.8497282        6.475056
#> 56        f     -160.8096651       -7.120481
#> 57        g     -141.8580485       41.422650
#> 58        h     -120.5941077      -72.347144
#> 59        i     -165.1154721       77.577002
#> 60        j       45.7223855      -66.380872
#> 61        a     -164.1625372      -42.725042
#> 62        b      156.3388244       -8.805360
#> 63        c       99.9471854      -18.727383
#> 64        d      125.1856964       10.699213
#> 65        e       96.4956988       73.665195
#> 66        f      -19.8319881      -75.061412
#> 67        g      136.3512079      -71.630050
#> 68        h     -164.6493013      -52.756976
#> 69        i      -29.6895575      -13.786802
#> 70        j      -28.8660396        5.830387
#> 71        a     -160.7525679      -51.949967
#> 72        b     -104.2280098       29.954529
#> 73        c      127.9040729       19.678944
#> 74        d      -40.2543456      -30.816931
#> 75        e     -121.9453358       10.188081
#> 76        f       -0.1045246       -8.686029
#> 77        g      140.4594714       83.452506
#> 78        h     -145.6403412       65.977659
#> 79        i      -23.3721282       28.103713
#> 80        j       56.1848199       -2.781816
#> 81        a      144.2815789       74.430107
#> 82        b      120.3428947       62.203516
#> 83        c       17.8578427      -48.134777
#> 84        d       67.1489516        1.673185
#> 85        e       34.3142950      -21.040804
#> 86        f        8.7409681      -58.361783
#> 87        g       18.2331012      -41.427523
#> 88        h       56.8121190        5.465928
#> 89        i      116.0046758      -25.021422
#> 90        j      -77.6132707      -29.447844
#> 91        a      140.2285505       18.745811
#> 92        b      117.0158775      -29.537177
#> 93        c      132.7405701       -3.084652
#> 94        d     -112.1510350       89.839130
#> 95        e      -77.6634068       85.070533
#> 96        f     -127.4589736      -57.459591
#> 97        g      -21.0806069       49.480168
#> 98        h       24.3255970       82.652275
#> 99        i       75.6168288      -60.546682
#> 100       j       58.1766329       20.670191
#> 101       a      162.4541218       46.315583
#> 102       b      168.3334337       41.900249
#> 103       c      174.0627845      -25.377515
#> 104       d     -150.6916454       15.699016
#> 105       e      -68.9605368      -68.664447
#> 106       f       85.3172068       22.165667
cc_val(x, value = "flagged")
#> Testing coordinate validity
#> Flagged 4 records.
#>   [1]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [13]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [25]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [37]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [49]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [61]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [73]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [85]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [97]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE
#> [109] FALSE FALSE