Identify Invalid lat/lon Coordinates

Removes or flags non-numeric and not available coordinates as well as lat >90, lat <-90, lon > 180 and lon < -180 are flagged.

cc_val(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Details

This test is obligatory before running any further tests of CoordinateCleaner, as additional tests only run with valid coordinates.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = c(runif(106, -180, 180), NA, "13W33'", "67,09", 305), 
                decimalLatitude = runif(110, -90,90))
                
cc_val(x)
#> Testing coordinate validity
#> Removed 4 records.
#>     species decimalLongitude decimalLatitude
#> 1         a       165.865575     -21.8210588
#> 2         b       -87.509353      45.7533892
#> 3         c      -154.067598     -46.0651815
#> 4         d      -158.608407     -40.9008672
#> 5         e       162.107276     -63.8093906
#> 6         f        42.392435      60.8651868
#> 7         g        88.669306     -73.5170536
#> 8         h       -99.089787     -80.7130684
#> 9         i       -62.495875      12.6668714
#> 10        j        90.535382     -51.4331904
#> 11        a       172.304785     -48.5446504
#> 12        b        45.107557     -77.9554026
#> 13        c        -4.818358     -16.0049135
#> 14        d        37.412693      71.9308610
#> 15        e        36.539962     -13.7064938
#> 16        f      -104.840500       7.2062200
#> 17        g        67.676037      88.4151850
#> 18        h       108.946050     -55.9290766
#> 19        i       -52.058440     -14.9825788
#> 20        j      -165.097908     -60.2610474
#> 21        a       -60.487622     -12.2229372
#> 22        b        11.996818     -20.4841125
#> 23        c        61.630940       0.7711996
#> 24        d         9.180686       6.7392740
#> 25        e       130.110074     -49.5563363
#> 26        f      -167.063901      50.3522065
#> 27        g       125.947233     -33.0620015
#> 28        h      -178.460701      78.0633930
#> 29        i      -148.330954      76.9589115
#> 30        j       120.771880     -80.5240671
#> 31        a       128.907723      61.0327828
#> 32        b        51.544032     -42.5404763
#> 33        c         5.656565      70.9313379
#> 34        d       -37.330790     -57.4881509
#> 35        e       -31.951893     -73.8776249
#> 36        f       -33.778301      11.1469222
#> 37        g       104.287937     -17.7749431
#> 38        h        60.892165      16.5508862
#> 39        i        90.678880      14.6880089
#> 40        j       -51.325164      -5.2213889
#> 41        a      -147.183082     -40.5488870
#> 42        b      -169.085769     -73.7632093
#> 43        c      -109.928285      -3.8480830
#> 44        d       -48.356559     -21.6594673
#> 45        e      -175.004955      19.7167866
#> 46        f       172.200976      40.7202246
#> 47        g      -105.996207     -49.1226201
#> 48        h       108.663670      12.0115933
#> 49        i       178.219621       6.8733955
#> 50        j       144.509980     -15.6339317
#> 51        a        17.679508     -71.0985146
#> 52        b       107.954730      -3.7144195
#> 53        c       177.679854     -61.4218990
#> 54        d        34.815629      69.0703513
#> 55        e      -159.492822      20.6187243
#> 56        f       174.752045      58.7050735
#> 57        g      -128.367311     -62.6613261
#> 58        h       119.200737      61.5594744
#> 59        i      -151.440373     -33.9114780
#> 60        j         8.790373     -66.3664776
#> 61        a      -141.725995      50.1102862
#> 62        b       -58.868813       2.0144845
#> 63        c       -86.161483      35.8043867
#> 64        d      -147.192225      57.3933919
#> 65        e        63.605653     -51.2545979
#> 66        f       -47.954242       4.8269290
#> 67        g       -31.414735     -16.6525523
#> 68        h      -103.760528      57.9273153
#> 69        i       -26.666800       0.7114552
#> 70        j       -19.270106      51.6449323
#> 71        a       126.845021      36.6700464
#> 72        b      -137.395921     -62.7405884
#> 73        c      -157.767117      31.7888260
#> 74        d      -106.089019       8.8599800
#> 75        e        86.463230      16.3981630
#> 76        f       163.876567       1.1082383
#> 77        g        28.280288      76.3811760
#> 78        h        15.365476      35.0678088
#> 79        i       142.771909      50.3947020
#> 80        j        94.752756     -63.6778584
#> 81        a       -80.505510      86.9897526
#> 82        b        28.784562      -9.0844800
#> 83        c        74.582163      63.0297372
#> 84        d       119.512830      33.2022491
#> 85        e      -135.499805     -88.3096052
#> 86        f        28.363287     -81.1599845
#> 87        g       106.344266      18.5588731
#> 88        h      -175.994887       5.8734067
#> 89        i      -143.971646     -30.5878505
#> 90        j      -171.247259      50.5983085
#> 91        a       -11.895587      66.6077386
#> 92        b      -100.950804      85.6828424
#> 93        c      -104.095097     -21.0733635
#> 94        d       133.416252      10.1278875
#> 95        e         7.910702      78.7415629
#> 96        f        21.001793       4.3959178
#> 97        g       -18.564520     -41.6269674
#> 98        h       144.160949      52.2941372
#> 99        i       -52.591902      57.5857647
#> 100       j       120.186037     -74.4435083
#> 101       a       -30.873197     -32.7201025
#> 102       b        52.659927     -28.0352036
#> 103       c       158.628623     -46.7088545
#> 104       d        99.181420     -59.0325581
#> 105       e       150.122238      13.9260145
#> 106       f       171.796886      39.7650469
cc_val(x, value = "flagged")
#> Testing coordinate validity
#> Flagged 4 records.
#>   [1]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [13]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [25]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [37]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [49]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [61]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [73]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [85]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [97]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE
#> [109] FALSE FALSE