Identify Invalid lat/lon Coordinates

Removes or flags non-numeric and not available coordinates as well as lat >90, lat <-90, lon > 180 and lon < -180 are flagged.

cc_val(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Details

This test is obligatory before running any further tests of CoordinateCleaner, as additional tests only run with valid coordinates.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = c(runif(106, -180, 180), NA, "13W33'", "67,09", 305), 
                decimalLatitude = runif(110, -90,90))
                
cc_val(x)
#> Testing coordinate validity
#> Removed 4 records.
#>     species decimalLongitude decimalLatitude
#> 1         a        49.520170       37.744903
#> 2         b        85.283972       49.772999
#> 3         c       -60.459268       42.024313
#> 4         d       172.993227       25.612864
#> 5         e        35.108847      -18.773264
#> 6         f       -78.714473       13.959387
#> 7         g       -90.710701      -66.887321
#> 8         h      -143.890912        5.171869
#> 9         i      -144.926692      -72.143916
#> 10        j       107.843596        6.508936
#> 11        a        59.247434       84.482729
#> 12        b        40.327750      -73.305040
#> 13        c       169.723105      -30.978369
#> 14        d       119.119589       83.553831
#> 15        e        89.597754      -26.781518
#> 16        f       -53.392417       62.380296
#> 17        g         3.793928       42.207274
#> 18        h       -74.527264       61.965585
#> 19        i       -10.087943      -24.546551
#> 20        j        23.077186      -38.986942
#> 21        a       104.491153        6.206106
#> 22        b        73.448346       66.777698
#> 23        c      -103.374358       -6.995751
#> 24        d       167.801267       30.548481
#> 25        e      -172.587245       28.821114
#> 26        f       -36.458226        5.042557
#> 27        g       -48.925700       13.278493
#> 28        h       126.218782        4.308643
#> 29        i       -11.118693      -78.998407
#> 30        j        35.879064      -87.909243
#> 31        a        -4.220322       49.537949
#> 32        b      -128.269754       13.063962
#> 33        c       171.663907       89.298665
#> 34        d      -178.904902      -63.251607
#> 35        e       -87.040392       35.654737
#> 36        f       154.434055       79.620657
#> 37        g        89.003860      -32.162420
#> 38        h       134.082361      -40.458021
#> 39        i      -157.839532       60.164080
#> 40        j       145.267199       15.802032
#> 41        a      -132.911217      -34.915291
#> 42        b      -172.873726      -18.242824
#> 43        c       -88.848835       57.652943
#> 44        d        74.240936      -62.293592
#> 45        e       -62.777495      -71.012244
#> 46        f        81.440542      -81.156450
#> 47        g        81.235277       22.473287
#> 48        h        46.747377       28.056418
#> 49        i        26.885972       31.601970
#> 50        j        89.283665        5.108212
#> 51        a       108.994570      -59.351650
#> 52        b        -7.117697      -28.718081
#> 53        c      -137.438851       -3.267989
#> 54        d       129.539840      -79.178717
#> 55        e        42.925778        8.026555
#> 56        f       -73.504940       27.057117
#> 57        g      -150.362123      -44.930065
#> 58        h      -134.908881       -0.924853
#> 59        i      -159.450799       74.149595
#> 60        j      -130.668142      -27.410324
#> 61        a       -88.329736       47.165718
#> 62        b       -76.376104        7.208978
#> 63        c        -5.304641      -76.624418
#> 64        d       120.981906      -16.268488
#> 65        e       132.029447       32.924504
#> 66        f       133.806730       36.972906
#> 67        g         3.668670      -21.519284
#> 68        h       -87.928084      -17.719354
#> 69        i      -113.782650      -60.796128
#> 70        j        -8.984733      -26.826182
#> 71        a       124.696963      -76.059164
#> 72        b       150.831621       82.137685
#> 73        c       -65.451492       18.743638
#> 74        d      -150.680928       -4.876263
#> 75        e       -56.176236      -14.589099
#> 76        f       142.410832       63.570228
#> 77        g        13.850642      -35.385125
#> 78        h         1.640161       -1.672395
#> 79        i       -97.160774      -54.864448
#> 80        j       113.651345      -15.691788
#> 81        a      -113.076374       65.690679
#> 82        b      -152.780652      -44.019230
#> 83        c      -154.741433       -7.702331
#> 84        d       -80.576595      -32.748874
#> 85        e        65.872055      -65.701115
#> 86        f       -58.145280       21.681025
#> 87        g       -60.918505        1.700578
#> 88        h        58.933064      -23.816216
#> 89        i       136.366310       71.135298
#> 90        j        69.077113      -82.979731
#> 91        a       135.430429      -17.678663
#> 92        b      -127.918521        3.268549
#> 93        c        53.220060      -74.321262
#> 94        d       -48.980157       35.101172
#> 95        e        66.448216      -16.222641
#> 96        f        62.089803       62.618696
#> 97        g      -144.469323        9.095891
#> 98        h       139.797529       78.154214
#> 99        i       -66.764363       20.535641
#> 100       j       144.452049      -51.207822
#> 101       a       174.895781      -21.134400
#> 102       b        89.178444      -55.621819
#> 103       c       178.053877      -11.328700
#> 104       d        77.605517       27.353010
#> 105       e        56.696793       51.884122
#> 106       f        -7.515809      -26.500748
cc_val(x, value = "flagged")
#> Testing coordinate validity
#> Flagged 4 records.
#>   [1]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [13]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [25]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [37]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [49]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [61]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [73]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [85]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [97]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE
#> [109] FALSE FALSE