Identify Invalid lat/lon Coordinates

Removes or flags non-numeric and not available coordinates as well as lat >90, lat <-90, lon > 180 and lon < -180 are flagged.

cc_val(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Details

This test is obligatory before running any further tests of CoordinateCleaner, as additional tests only run with valid coordinates.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = c(runif(106, -180, 180), NA, "13W33'", "67,09", 305), 
                decimalLatitude = runif(110, -90,90))
                
cc_val(x)
#> Testing coordinate validity
#> Removed 4 records.
#>     species decimalLongitude decimalLatitude
#> 1         a     -171.2603225     -25.7103587
#> 2         b      -28.7428296     -76.8463872
#> 3         c      170.3068257     -61.4160946
#> 4         d       11.0074088      54.4608785
#> 5         e       33.5851268     -39.3743631
#> 6         f      -64.0536377       5.7617231
#> 7         g      134.8252902     -76.4845802
#> 8         h      151.9532405     -31.7602345
#> 9         i     -174.0661853      47.2571627
#> 10        j     -151.2383254      72.9738944
#> 11        a     -104.3261173      43.1756517
#> 12        b      132.9911313     -64.0670893
#> 13        c      177.5086153     -29.6251872
#> 14        d      -18.3065576      32.6739223
#> 15        e       -5.4098360      -5.3601748
#> 16        f       83.8658954     -24.2528786
#> 17        g      -88.0081504      -3.5960528
#> 18        h       -2.1802245     -27.2631790
#> 19        i      -51.6053391     -12.9480669
#> 20        j       -2.0105265     -80.2561637
#> 21        a     -160.9450728      36.4830114
#> 22        b      -17.2957632     -25.0452044
#> 23        c     -120.5325098       4.9286485
#> 24        d       63.6039872     -89.5669541
#> 25        e       96.6485057      84.7984313
#> 26        f       75.5826641     -29.0329281
#> 27        g        4.8798833       1.7314479
#> 28        h      -74.7312280     -55.9576817
#> 29        i      -15.3647891     -25.2734597
#> 30        j     -114.7326665     -31.2163700
#> 31        a      101.4862993      12.1395223
#> 32        b     -157.7898746      -4.1872935
#> 33        c       35.5330894       5.5081692
#> 34        d      -60.3787533      14.1792878
#> 35        e       78.0932101      77.5561872
#> 36        f        9.6615739      34.3640313
#> 37        g     -129.0904973      85.0789964
#> 38        h        1.6791762      26.5464853
#> 39        i     -150.9219334     -76.3895160
#> 40        j      -58.1349319     -27.6589302
#> 41        a     -145.5848210      75.8482666
#> 42        b      -30.0328933      10.7937097
#> 43        c      -39.3287648      17.2766976
#> 44        d      -92.7868417       0.5935033
#> 45        e     -124.9513471     -10.8888133
#> 46        f      157.9366326      86.7939401
#> 47        g        2.1694898     -18.7239727
#> 48        h       96.3060224     -50.6976829
#> 49        i     -159.8901614     -63.3847465
#> 50        j      173.7787736      63.6197711
#> 51        a       70.0986151      24.6294862
#> 52        b      123.5262090     -68.5188227
#> 53        c      108.5639863     -78.1118717
#> 54        d     -100.9318452     -59.1711021
#> 55        e     -102.6605451      60.6580786
#> 56        f     -109.0496027     -77.7115201
#> 57        g      -43.0431415     -67.1389962
#> 58        h      -55.4739185     -31.0072785
#> 59        i     -152.6337360      -1.5444825
#> 60        j      -73.1552603     -85.2106989
#> 61        a       32.5712230     -31.4247529
#> 62        b       91.1046252      42.6287475
#> 63        c      136.0632111      44.1543079
#> 64        d       98.3337871      37.4483046
#> 65        e       67.9120536      51.8484024
#> 66        f      118.7648327      19.9516416
#> 67        g       52.9511169      -0.6238997
#> 68        h      -59.2863390      42.6371184
#> 69        i      123.7291063     -67.7480056
#> 70        j      139.0664781     -29.6534417
#> 71        a       73.8063751      70.0182131
#> 72        b      -42.6641318      61.7448444
#> 73        c     -120.0435983      87.4027881
#> 74        d      -79.7962924      49.8213160
#> 75        e       29.9676329     -24.1636831
#> 76        f       -3.5064290      79.2491681
#> 77        g     -159.0953240     -42.3531444
#> 78        h        9.1549261      20.6134410
#> 79        i     -179.0715518      47.0657493
#> 80        j     -136.7526096      42.3575430
#> 81        a      153.6990299     -10.5888483
#> 82        b     -174.8776067     -69.5020773
#> 83        c      -53.8379208     -79.3524037
#> 84        d      -58.6844933       7.8579023
#> 85        e     -129.2420131      12.9951808
#> 86        f       -0.7617522     -60.3520915
#> 87        g      139.7202015       7.2932854
#> 88        h      -91.5425786      70.0608733
#> 89        i      -89.1806213     -78.8696387
#> 90        j      163.4096094     -26.7187817
#> 91        a        9.6110073      76.4796823
#> 92        b     -124.1583976     -74.6323529
#> 93        c       38.8885083      35.9510093
#> 94        d       31.9124696       4.3966743
#> 95        e       27.8850351     -78.2566685
#> 96        f      111.2747367     -68.1411168
#> 97        g     -133.3475489     -45.9781686
#> 98        h       42.0228219      79.0286184
#> 99        i     -113.4753173     -54.2982035
#> 100       j       90.5694863      84.0453229
#> 101       a      160.2733013      26.6581712
#> 102       b      178.2956997      63.0129764
#> 103       c      114.5838553     -44.8274212
#> 104       d        9.1675843     -46.0482924
#> 105       e       -3.0067956      36.5753703
#> 106       f       -5.5631077      40.6361059
cc_val(x, value = "flagged")
#> Testing coordinate validity
#> Flagged 4 records.
#>   [1]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [13]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [25]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [37]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [49]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [61]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [73]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [85]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [97]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE
#> [109] FALSE FALSE