Identify Invalid lat/lon Coordinates

Removes or flags non-numeric and not available coordinates as well as lat >90, lat <-90, lon > 180 and lon < -180 are flagged.

cc_val(
  x,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  value = "clean",
  verbose = TRUE
)

Arguments

x

data.frame. Containing geographical coordinates and species names.

lon

character string. The column with the longitude coordinates. Default = “decimalLongitude”.

lat

character string. The column with the latitude coordinates. Default = “decimalLatitude”.

value

character string. Defining the output value. See value.

verbose

logical. If TRUE reports the name of the test and the number of records flagged.

Value

Depending on the ‘value’ argument, either a data.frame

containing the records considered correct by the test (“clean”) or a logical vector (“flagged”), with TRUE = test passed and FALSE = test failed/potentially problematic . Default = “clean”.

Details

This test is obligatory before running any further tests of CoordinateCleaner, as additional tests only run with valid coordinates.

Note

See https://ropensci.github.io/CoordinateCleaner/ for more details and tutorials.

See also

Examples

x <- data.frame(species = letters[1:10], 
                decimalLongitude = c(runif(106, -180, 180), NA, "13W33'", "67,09", 305), 
                decimalLatitude = runif(110, -90,90))
                
cc_val(x)
#> Testing coordinate validity
#> Removed 4 records.
#>     species decimalLongitude decimalLatitude
#> 1         a        79.962352      82.8218451
#> 2         b        40.484273      26.3755131
#> 3         c      -121.068792      49.9904139
#> 4         d       126.462288      86.5792229
#> 5         e        53.643200     -84.4205003
#> 6         f        70.150907     -75.4527707
#> 7         g       124.999486      56.7154486
#> 8         h       -51.497580      -9.2979822
#> 9         i       -46.776486     -58.8887333
#> 10        j      -130.462885      41.9847199
#> 11        a        16.678831     -22.1390434
#> 12        b        84.933740     -73.4964096
#> 13        c      -156.778473     -46.2795208
#> 14        d       155.442831      22.5470729
#> 15        e       -29.284296     -55.2782581
#> 16        f        63.089563     -77.3907425
#> 17        g       -22.980362     -63.0460264
#> 18        h      -162.567566      65.4452904
#> 19        i       -66.323014      58.5181762
#> 20        j       147.050652     -25.4228431
#> 21        a       130.728724     -55.9351660
#> 22        b      -131.271705     -41.5491379
#> 23        c       108.272370      -0.3464640
#> 24        d       -87.116467      25.6535324
#> 25        e       -38.077294     -19.8405940
#> 26        f       -13.081897       3.6278944
#> 27        g        60.552345      21.0103750
#> 28        h         6.184018     -86.2792960
#> 29        i      -107.584238     -31.8961632
#> 30        j       -66.994956      75.3440307
#> 31        a       -59.344956      38.6452124
#> 32        b      -157.564284       3.1418315
#> 33        c      -158.690061     -36.6843308
#> 34        d      -128.487645       8.5821879
#> 35        e        49.091441      45.5006894
#> 36        f       127.119986      82.4757509
#> 37        g       -44.296829     -71.0860470
#> 38        h        13.478890     -46.9117570
#> 39        i       -38.950506      19.2468861
#> 40        j      -169.141866     -12.5511303
#> 41        a        76.687504      20.6535579
#> 42        b        48.133029      13.2513207
#> 43        c       -42.542319       4.1678098
#> 44        d        92.169467     -61.9319718
#> 45        e        17.985988      -3.2560716
#> 46        f       -45.578314     -69.7926178
#> 47        g       -51.808841     -71.1111197
#> 48        h        -2.419177     -45.4945060
#> 49        i       100.014364      50.9495558
#> 50        j      -125.941846      -0.2253606
#> 51        a       -79.394058      25.2787357
#> 52        b       106.275642     -25.3310033
#> 53        c       -69.246532      62.2713421
#> 54        d       155.842538     -20.4173162
#> 55        e        42.468583      46.1834976
#> 56        f        24.223037      59.4515967
#> 57        g       136.055054      28.4462157
#> 58        h      -116.062548      -7.9161283
#> 59        i      -166.599834      83.7753504
#> 60        j       169.173688      29.2390642
#> 61        a        56.666610      17.5197855
#> 62        b       -58.930918      10.8566367
#> 63        c       142.131412      12.5199375
#> 64        d      -123.922526      34.6843510
#> 65        e      -119.717841     -81.8139656
#> 66        f        64.950021     -34.9023407
#> 67        g       -15.925141      69.6855645
#> 68        h       -61.594393      83.8592237
#> 69        i       -49.462548     -16.0181485
#> 70        j       -99.975804     -46.7698384
#> 71        a        42.388387     -87.8239472
#> 72        b        76.460352      74.4117000
#> 73        c        21.855444      -9.1692027
#> 74        d       -71.484719      62.2015524
#> 75        e        41.128288     -79.6752959
#> 76        f      -137.477718       9.0724459
#> 77        g        91.844768      44.7850212
#> 78        h        66.969647     -56.5002245
#> 79        i       -10.759217      62.8998741
#> 80        j         6.597837      65.5068855
#> 81        a        47.501711      27.8941931
#> 82        b        37.641839      58.9941315
#> 83        c        -8.290675      70.2782772
#> 84        d       132.959418      -1.7289456
#> 85        e       -88.058216     -33.9622474
#> 86        f        85.605092      89.6299546
#> 87        g        -4.104688      83.5144800
#> 88        h        96.031939     -11.1747554
#> 89        i      -110.329395     -45.7914698
#> 90        j       -15.829038     -38.0829291
#> 91        a        16.930910       9.2566672
#> 92        b        48.638792       7.6159279
#> 93        c       119.947678      72.3838982
#> 94        d       -71.330604      28.4174816
#> 95        e        27.045701      49.9887498
#> 96        f       170.018114     -42.0888870
#> 97        g       -11.814813     -44.6405909
#> 98        h      -146.312395      -5.9793268
#> 99        i       164.264065      29.8428774
#> 100       j      -107.648208      31.1557583
#> 101       a       -71.136634     -10.2957632
#> 102       b       153.366200     -45.7118482
#> 103       c       -53.712291     -61.6997942
#> 104       d       -41.045011     -80.8153997
#> 105       e      -153.752583      86.5452162
#> 106       f      -136.400016     -69.3177727
cc_val(x, value = "flagged")
#> Testing coordinate validity
#> Flagged 4 records.
#>   [1]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [13]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [25]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [37]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [49]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [61]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [73]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [85]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [97]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE
#> [109] FALSE FALSE