Country Centroids and Country Capitals

A data.frame with coordinates of country and province centroids and country capitals as reference for the clean_coordinates, cc_cen and cc_cap functions. Coordinates are based on the Central Intelligence Agency World Factbook https://www.cia.gov/library/publications/the-world-factbook/ and http://thematicmapping.org/downloads/world_borders.php.

Format

A data frame with 5,142 observations on 10 variables. #'

iso3

ISO-3 code for each country, in case of provinces also referring to the country.

iso2

ISO-2 code for each country, in case of provinces also referring to the country.

name

a factor; name of the country or province.

adm1_code

adm code for countries and provinces

type

identifying if the entry refers to a country or province level.

centroid.lon

Longitude of the country centroid

centroid.lat

Latitude of the country centroid

capital

Name of the country capital, empty for provinces

capital.lon

Longitude of the country capital

capital.lat

Latitude of the country capital

Source

CENTRAL INTELLIGENCE AGENCY (2014) The World Factbook, Washington, DC.

http://opengeocode.org/download/cow.php http://thematicmapping.org/downloads/world_borders.php

Examples

data(countryref)
head(countryref)

#>   iso3 iso2 adm1_code        name    type centroid.lon centroid.lat    capital
#> 1  ABW   AW       ABW       Aruba country    -69.97700       12.517 Oranjestad
#> 2  ABW   AW       ABW       Aruba country    -69.96667       12.500 Oranjestad
#> 3  AFG   AF       AFG Afghanistan country     65.21600       33.677      Kabul
#> 4  AFG   AF       AFG Afghanistan country     65.00000       33.000      Kabul
#> 5  AGO   AO       AGO      Angola country     17.54400      -12.296     Luanda
#> 6  AGO   AO       AGO      Angola country     18.50000      -12.500     Luanda
#>   capital.lon capital.lat
#> 1      -70.03       12.52
#> 2      -70.03       12.52
#> 3       69.18       34.52
#> 4       69.18       34.52
#> 5       13.22       -8.83
#> 6       13.22       -8.83